#100451# All the pseudo sprm handlers currently are defined as if to be called at the start and end of the fields and other similiar events. But they don't actually do this so remove their redundant second argument. We do fields to create start and end events so change the fld plcf to the the IsSprm type of PLCF which supports start and ends and now set the end point of field ranges.  i.e implement GetSprms instead of GetNoSprms.
To keep track of what fields are currently being processed create a stack of
field codes to push each started field onto and pop them off when their end
point has been passed. Hyperlink fields are attributes in writer and this
difference requires that the attribute that the link corresponds to remain on
the atrtibute stack while its word equivalent remains on the field stack.  At
the end of the hyperlink field we can pop the field off and set the end point
of the attribute correctly, the case of a single inline graphic where we
traditionally set its hyperlink attribute directly and the cases of multiple
paragraphs inside a hyperlink and the cases of recursive fields can now be
handled naturally as part of the existing parsing as opposed to the
preexisting and faulty lookahead and use of "last inserted graphic" hacks can
be replaced.

+ replace the hardcoded 256-260 pseudo sprm values with enums to help
cleanup that mess.