INTEGRATION: CWS aig02 (1.57.32); FILE MERGED 2003/06/30 13:16:01 aw 1.57.32.1: #108784#
In operator<<, force to write a naked SdrObj instead of the virtual object. Writing virtual objects is not supported at all, thus i am sure that this was not done before. It is now necessary to write this as naked SdrObject for keeping the OrdNums when using the new feature for Obects in Header/Footer of SW.