INTEGRATION: CWS pbrwuno (1.14.206); FILE MERGED 2005/10/17 14:09:27 fs 1.14.206.3: #i53095# some cleanup of remaining TODOs 2005/10/05 06:56:21 fs 1.14.206.2: RESYNC: (1.14-1.15); FILE MERGED 2005/08/18 12:44:29 fs 1.14.206.1: #i53095#, phase 2
moved (nearly) all property handling to dedicated handlers, the controller is
now simply managing a set of handlers

open issues for making the property browser completely generic:
- target page for a property - at the moment, the pbrw uses form-specific
  knowledge
- relative position of properties. Again, the pbrw uses the OPropertyInfoService
  which is not generic
- isComposeable for a given property. Also OPropertyInfoService-dependent ATM
- help ids of pages and the pbrw as a whole. They're hard-coded at the moment

other open issues:
everything in the code which is tagged with TOD/UNOize. Those are items which
do not immediately hinder phase 3 (real UNOization, i.e. definition of new
UNO interfaces for the handlers, the controller, and so on), but need to be
addressed in phase 4 (knit lose ends)