INTEGRATION: CWS rpt23fix02 (1.2.4); FILE MERGED 2007/07/30 13:08:49 tmorgner 1.2.4.2: Issue number:  80157
Keep-Together is now applied to the paragraph-style instead of the cell-style.
However, as cell-styles dont (yet) get inherited if the paragraph has a style
defined, this now breaks the fonts of the first cell. Either we or the text/
style team should fix or work around this issue.

Also fixed:
A possible nullpointer if the image-data is invalid.
2007/07/25 15:05:46 tmorgner 1.2.4.1: Issue number:  78128

This commit fixes the messed up repeating-header system. The whole thing has
been redesigned to create a clean state-flow that can be debugged and traced
without any weird mind-bending activities. The report-targets now have a
new state (IN_GROUP_INSTANCE) to keep track of the group's processing and
to separate repeating headers and footers from ordinary headers and footers.

This patch also enforces a clean state-flow in the OfficeReportOutputTarget
and explicitly checks for all allowed band-types and throws exceptions on
every illegal state.