INTEGRATION: CWS rpt23fix02 (1.2.4); FILE MERGED 2007/07/31 09:34:16 tmorgner 1.2.4.2: Fixed the non-display of shapes that was caused by some unnecessary CDATA copied from the original XML file. We now ignore all CDATA from <draw:*> elements. 2007/07/30 13:08:50 tmorgner 1.2.4.1: Issue number:  80157
Keep-Together is now applied to the paragraph-style instead of the cell-style.
However, as cell-styles dont (yet) get inherited if the paragraph has a style
defined, this now breaks the fonts of the first cell. Either we or the text/
style team should fix or work around this issue.

Also fixed:
A possible nullpointer if the image-data is invalid.