autorecovery: save open/modified (Writer-based) reports and forms, when doing a session/emergency save
Still, the documents are stored only, and not loaded. For storing, there's at least one known issue: The controller
is unable to deliver the proper object name for a form which has been newly created, and then saved. In this
case, an empty name (from the time where the form was created) will be returned, and used in the save process.