Improve randomisation in 'dissolve' transition.
The loop for randomizing a list in the dissolve transition is poor.

It randomly chooses two element and swaps them, and does this N/2 times.
Thus a total of N elements are moved.
It is probable that several elements will be swapped multiple times,
leaving it probable that several other elements will not be swapped at all.

This fact is quite visible when transitioning between two very different
slides (e.g. 2 photos).  While there is a lot of randomness, there is a
clearly perceptible top-to-bottom 'wipe' happening at the same time.

So change the loop to provide true randomness.
We randomly choose an element for the last position.
Then randomly choose one of the remaining elements for the second
last position, etc.

Signed-off-by: NeilBrown <neilb@suse.de>