Ported dde-reconnect-on-load-*.diff from ooo-build.
Improve reliability of DDE connections between documents opened in LO.
With this change, LO tries to reconnect to a DDE server document upon
opening if that server document is being listened to by one of the
open documents.

Also, the old implementation would load a DDE server document
invisible, and would never close it when the user updates the link.
This had the consequence that when the user tries to open this
document while it's loaded hidden, it causes some weird focus issues,
and closing it and opening it again would disconnect the DDE
connection.  The new implementation closes the server document
immediately after the DDE link update is complete.

This change also fixes a bug in Calc where DDE link updates to cells
would fail when the formula syntax is set to something other than Calc
A1.