Attempt to fix some vos removal fallout on Windows
After of the vos removal, some classes in other parts of the code now
are subclasses of SimpleReferenceObject.

To avoid compilation and linking errors in those other parts, we seem
to need to have the operator new[] and operator delete[] in
SimpleReferenceObject as protected instead of private. We also need an
implementation of operator delete[]. As it can be unimplemented with
gcc, I hope it isn't actually called, though, so it for now crashes
intentionally if called.

Will see later if this actually works, this is just a commit to make
stuff build.