Temporarily disable the extra ACL check.
The fix for i#102464 no longer seems necessary.  In fact, leaving
that "fix" would interfere with documents saved in samba shere;
when a 2nd user tries to open a document in a samba share that's
already opened by someone else, it would silently open in read-only
mode, but the correct behavior is to launch the "Document in Use"
dialog telling the 2nd user which user has it open.

If disabling this checks causes no regressions, we can remove this.