work around tables with many columns for msoffice (fdo#30860)
Theoretically the .docx spec doesn't seem to have any limit
on the number of table columns, but in practice msoffice seems
to have a hardcoded limit of 63 columns and fails to open such
documents. Since the expected use case with .docx is that
somebody with msoffice will try to open it, keep the limit
and merge exceeding items into the last column.