SwCallLink: Removed unnecesarry Invalidate() call.
In commit 47472e8 the author intended to fix some repainting
issues with nested tables. This made the cursor flicker and
disappear everytime it moved inside the table.

The problem was tracked down to an Invalidate() call inside
SwCallLink::~SwCallLink(). The lines are removed. Neither of
both problems reproduce on tests so far.

See discussion "Annoying cursor behavior on sw tables"
http://www.mail-archive.com/libreoffice@lists.freedesktop.org/msg04197.html