deletion of StgWriter::Write and its dependancies
During this deletion part, all the deletion have been done within
the binfilter/bf_sw, not touching the generic part. Further, all deletion
have been done based only on the deletion of the above mentionned function
and of its dependancies.
This step left behind some orphean functions, that are probably not called anymore,
this will be deleted in further steps.