We're not going to be called "StarSuite" again
Once StarOffice came in two varieties, StarOffice and "StarSuite" in the Asian
market. This code effectively converted "StarOffice" to "StarSuite" if product
itself was called "StarSuite", which ain't going to the be the case again. So
we can remove all that and just hardcode the old file format names as
"StarOffice".