Fix overflow problem in the custom action for UI language de-selection
In the _all_ installer we have slightly over 100 LibreOffice UI
languages. The code in the custom action that de-selects UI languages
based on the Windows UI language(s) and the user's locale had a limit
of 100.

The consequence was that a few of the alphabetically last LibreOffice
UI languages: Ukrainan, Urdu, Uzbek, Venda and Xhosa, were not
de-selected even if none of those languages was used by the user or
Windows, and thus got installed in a "quick" install and also in a
"custom" install unless the person running the installer noticed.

Reported by aseques_ on IRC, no bug filed yet.

Signed-off-by: Fridrich Štrba <fridrich.strba@bluewin.ch>