Ensure that the selected field is displayed when the selection moves.
In the DataPilot main dialog, Set the selected field in the data
select area to the one at the lower right corner, and press 'r' to
move it to the Row field.  This causes the selection to move to the
next field in the select area, but the selected button was not shown.