Related: fdo#33510 right align symbol preview area and center text on line
Stick the symbol preview over on the right, do more careful shrinkage of fonts
to fit height if neceessary, chop bits off the preview text to get it to fit in
available space to make it looks pleasently aligned.

Be way more careful about centering the text in the available area.