Related: fdo#33549 remove bogus CorrectResIdForCharset
This looks totally bogus, and even if it isn't then in
SwWW8ImplReader::SetNewFontAttr the *original* nWhich
needs to be the one used for the push onto the encoding
stack, otherwise its all busted afterwards on the pop.

I suspect this is an attempt to fix some other problem
earlier, maybe idcthint related or maybe default font
related. Impossible to know as there was no comments
or bugids mentioned originally to see what the problem
being fixed was.