docx import: don't add all characters from the parser to paragraphs
When the content XML file in the docx archive is formatted with white
spaces and lien breaks, those are imported as characters in the
paragraph... which makes no sense.

The stream handler was rerouting all parsed characters to the dmapper,
making the characters action defined in model.xml useless. Removing the
overridden characters from OOXMLFastContextHandlerStream fixed the bug