rework confused and racy clipboard event polling (bnc#683941)
It does not work reliably to just poll on the X connection fd,
as there may be already incoming events in the Xlib queue (removed
from the fd connection). So XPending() call is necessary; which
additionally ensures the output queue is flushed before polling.
The second poll() seems to have been just introduced as a result
of failing to understand this, and the strange XPending() nesting
later is no longer needed either.