create a pseudo-module tail_build to take advantage of gbuild features
gbuild allow for multiple 'module' to be build as one Makefile.
This help with parallelism.
This module build all the 'tail' end module. these that are not
dependencies or other dmake-based module, and that pseudo module
is declared as a dependency of postprocess instead of the individual
modules it 'group'