fix off-by-one when exporting postscript fonts to pdf
The problem shows when exporting a text document to pdf and there
a .pfb font is used (e.g. Standard Symbols L or Bitstream Charter).
As far as I can say this is actually harmless, as this is just
a padding section, but Acroread shows an error dialog for this
(bnc#690005).