Fix build on Windows
No need to mention $(LB)$/$(TARGET)$(DLLPOST) in the ALLTAR
prerequisite list. It causes errors, and is unnecessary, as pyuno.pyd
gets built anyway thanks to being the TARGET in this makefile.

No need to do the symlink if DLLPRE is empty. In that case
$(LB)$/$(TARGET)$(DLLPOST) equals $(LB)$/$(DLLPRE)$(TARGET)$(DLLPOST)
so dmake will complain about a circular dependency.