Fix lotuswordpro unit test on Windows
file: URLs on Windows are of the form file:///c:/path/to/file , and
the SRC_ROOT environment variable is of the form C:/lo/git/master, so
we can't just concatenate "file://" and SRC_ROOT, but have to insert a
"/".

Windows programs don't understand Cygwin symlinks, so we have to
specify the path from SRC_ROOT to lotuswordpro through an explicit
clone/filters path.