Make checks for db work when cross-compiling
When looking for the db,h header, use Autoconf mechanisms instead of
manual checks in hardcoded directories. So yeah, this means that you
need to make sure the correct -I flag is passed if you have db
installed in a weird place where the compiler doesn't find it.

Use checks that require only compiling, not running code. Nice.

Don't AC_SUBST variables that are not used.