fdo#36690: Don't broadcast setting changes during painting of button.
Calling SetSettings() when the output device is Window causes it to
broadcast data change.  PushButton::Draw() in fact calls this method
when the button contains colors, which ends up broadcasting its change
via Window::DataChanged call.  But depending on the output device
this DataChanged call may end up painting it again, which basically
causes a recursive loop.

The solution is to make sure OutputDevice::SetSettings() is called ( which only
 changes the settings without doing anything else ) and *not* and derived
classes 'SetSettings()' method

Signed-off-by: Noel Power <noel.power@novell.com>
Signed-off-by: Michael Meeks <michael.meeks@novell.com>
Signed-off-by: Jan Holesovsky <kendy@suse.cz>