Stylistic follow-up cleanup
Rename the bHighlight parameter to ImplDrawItem to nHighlight. It is
no longer claimed to be a boolean (even if it was passed values 0, 1
or 2) but instead a sal_uInt16.

When used in a logical expression, use explicit nHighlight!=0.

Pass nHighlight integers 0 or 1 in calls instead of sal_False or
sal_True.