Fix raptor to not use xml2-config on Mac
Christian suggested modifying the configure script not to use
xml2-config/xslt-config at all. Here's a patch which does exactly
that. I'm not exactly thrilled by the size/complexity of this patch
for what it does, but at the same time it does seem a bit crazy for
LO not to build just because libxml is installed in MacPorts (which
is not uncommon: it's a dependency of ImageMagick for instance).