always get GNOME MIME type icons from the gnome subdir
sysui/desktop/share/create_tree.sh put the GNOME icons under
the gnome subdir and KDE icons under the hicolor and lolor dirs.
GNOME and KDE root dir is the same (/usr) on RedHat and FreeDesktop.
It is different (/opt vs. /usr) on SUSE. The directory
/opt/gnome/*/hicolor is available on SUSE but the LO icons
are installed only under the gnome subdir.