Fix build by updating to new api
Hi,

I was rebuilding master, and had to change a line in libs-core which used
the old api for SetWindowState. Could this get pushed to master (if it's not
already done by the time people read this)?

Kind regards,

Luke.

Hi,<div><br></div><div>I was rebuilding master, and had to change a line in libs-core which used the old api for SetWindowState. Could this get pushed to master (if it&#39;s not already done by the time people read this)?</div>
<div><br></div><div>Kind regards,</div><div><br></div><div>Luke.</div>

>From b759a7cf43544a3eb9614d506637652ba3b20465 Mon Sep 17 00:00:00 2001
From: Luke Symes <allsymes@gmail.com>
Date: Thu, 23 Jun 2011 20:50:45 +1200
Subject: [PATCH] Update for new api for SetWindowState