Replace List with std::vector< AInfo* >
I actually moved the List member from the class to a class method. I think
this is safe because the member was only used in one method and it's life-
span was limited to the one method (populate, process, dispose - all in one
"if" block.

I also removed all the member's export methods since no one used them.