Import of xrefs to numbered paragraphs from docx
The first part of this patch adds importing of cross-references to numbered
paragraphs from docx file.

The second part provides a fix to imports of bookmarks from docx, which in
the test files I have end up importing with an empty bookmark name. The
problem is that for each bookmark, DomainMapper_Impl::AddBookmark is called
twice - once with an empty string for rBookmarkName, and once for the
string that is in the file. The fix I have used is to ensure the passed in
name is used if the bookmark has presiously been added, and the old one has
an empty string as the name. This works, but I am not sure it is correct -
I have not investigated why it is called with an empty string in the first
place.

Signed-off-by: Luboš Luňák <l.lunak@suse.cz>