Bin unused symbolic constants
Is it really good Mac style programming practice to define your own
non-namespaced kThisAndThat constants that look like they might be
defined by some platform API?

And seriously, defining symbolic constants for small integers like
const kOneBit = 1 ?  These were not used, but if they had been would
it really be cleaner to use the identifier "kOneBit" instead of a
plain number 1?

Cases where small integers have magic meaning and don't stand for just
themselves are obviously different; there it makes sense to use
symbolic names and not the integer values. But in the case of things
like kOneBit that is not the case, as one might guesss from the name
it was supposed to be used to indicate a bit depth of one. Etc.