Fix fdo#36868 - doc import list numbering error
This patch appears to be fix fdo#36868. The problem is that the paragraph
style was never being assigned an outline list level. The saved file did
not have an sprmPOutLvl for that paragraph (although it did have one for
the level 0 style), just an sprmPIlvl and an sprmPIlfo. Therefore there
was no call to SwWW8ImplReader::Read_POutLvl to set nOutlineLevel.