remove the concept of modules_type, source_config and related things
all these bells and whistles seems to be hamburg specific.
modules_type in you build seems to always be 'mod', so no need
to track it.

the whole source_config things is not use in our build either.

note: this is a rough cut... there is probably left-over to trim as
a consequence of this.