Fix MSVC build with debugging C/C++ runtime, i.e. with --enable-dbgutil
Whether this stuff actually works with the debugging runtime is
another question... The reason why the .cxx and .hxx files in this
commit explicitly undefined _DEBUG is unknown. As usual the old commit
message doesn't give any clue. Hamburg apparently didn't use the debug
runtime in their "dbgutil" a.k.a. "non-pro" builds?