Delete class XGradientTable
It seems that we don't use it...

Class XFillGradientItem referenses it as a parameter which is never passed;
however, the code is: "if(!isIndex()) else Table->" which will always fail
if the Item is an Index. This makes me belive that the XFillGradientItem
class could be simplifed to not be based on the NameOrIndex class.