Polyline improvement for Windows
I tried to draw polyline.... But on Windows... the end point is not
drawn... the main problem is WinAPI LineTo() function. MSDN says:

The LineTo function draws a line from the current position up to, but
not including, the specified point.

I think WinAPI Polyline() function uses LineTo() to draw a polyline. I
see a hack in VCL module in WinSalgraphics::drawLine(): "we must paint
the endpoint" But there is no such hack for drawPolyLine() function. A
possible solution is same as for drawLine().