Adapt for MinGW cross-builds
Adapt some of the DLL names to match conventions we currently use in a
MinGW build. If those are changed to be exactly like when built with
MSVC (for SDK ABI stability reasons), will have to change here, too.

Bypass stuff that we can't build with MinGW when necessary. Should be
synchronized with the corresponding makefiles, obviously. We can't
currently build the Explorer extension or MSI installer custom actions
with MinGW due to lack of some required headers and/or import
libraries.