Tell RootAccess when Components instance is destroyed.
This is to avoid RootAccess instances, which are ref-counted, to
avoid accessing the Components instance, which is a singleton.

The problem may arise when shutting down the entire app where it's
entirely unclear whether the RootAccess instances get destroyed first
or the Components instance does.  In fact, on my Windows XP environment
this repeatedly caused crash-on-close issue which would lead to "DDE
server error" Windows system error dialog popping up afterwards.