fast_merge: fix mis-merge of first module's strings
The problem was that in write_lines() when $first_run was true, the
add_to_buffer() call already set the current module's name to the second
module.

A fix for this is to use let make_paths() take a parameter, and in case
of the first run, pass the really first module name, not the current
one.