Fix E_ACCESS failure under windows on attempt to remove temp qa data file
The reference counting of ScDocShell is very confused. It starts off with ref
count of 1, assigning it to a ScDocShellRef gives it a refcount of 2, so
clearing it or its dtor gives a refcount of 1 again, so it doesn't auto-delete,
and so doesn't close its stream, so the attempt to remove its stream gives
E_ACCESS under windows.