bnc#726152: Avoid adjusting cell-anchored objects on other sheets.
The old code would incorrectly move anchors from one sheet to another
when updating the anchors of an object that was not on current sheet.

That can happen e.g. when modifying a cell value which triggers a
(cell-anchored) chart object on another sheet to get updated.

Interestingly, this issue ended up causing a write error during file
save in some situations, and in others silenting removing the affected
chart objects on save.