For Android, we don't actually want to build programs, but shared libraries
That's what the NativeActivity stuff wants.

For now the name of the produced shared library does not follow the
normal libfoo.so convention, but is the name of the corresponding
program on normal Unixes. To be fixed later. Or then, when
constructing an app .apk, just rename "foo" to
"libnative-foo.so".