Revert "Just put the Android ARM compiler options in the mk files"
Nah. Sure, doing it like that is sane if you consider only our own
code. But it won't help in making sure these options are used when
compiling the 3rd-party external libraries. For that having the
options already in the CC and CXX environment variables works nicely,
I think. People just need to look at the samples in README.cross.

Perhaps we could add the options to CC and CXX in configure.in,
though? Is that elegant or not, don't know...

This reverts commit 3c4f7242a967d6a3cb9e5e3f88bfb2cc348cc3eb.