Use RTLD_DEFAULT on Android
dlopen()ing NULL (the main program) and dlsym()ing on that handle does
not work on Android as in GNU/Linux. So don't bother with that, just
call dlsym(RTLD_DEFAULT, symbol) instead.

But, it doesn't help, it won't find the type_infos anyway, as they are
weak symbols and the dlsym() at least in Android 4.0 does not like
weak symbols, "Symbol is not global."

So the fallback branch that dynamically generates a type_info, which
normally should get used only for inter-process UNO (and thus
presumably never on Android) is taken instead.

But, that didn't work either at least with NDK r7 thanks to a bug in
its libgnustl_shared.so. But we now patch that at run-time in
android_main() in sal/osl/android/jni/lo-bootstrap.c.

Are we having fun yet?