Android hacking
Start of an app to just load some document. Uses API from the
org.libreoffice.android.Bootstrap class.

Not sure what is the sanest way to build an app like this. It needs a
bunch of shared libraries of course to be copied into libs/armeabi-v7a
so that they get included in the .apk. Perhaps a Makefile similar to
the one for lo-bootstrap might be good?

But for debugging the Java code Eclipse is the way to go (?), and to
be able to do that Eclipse wants a "project". So should this then be
built only through Eclipse? Can one build Eclipse projects from the
command line?