process element in endFastElement(), not in dtor
Otherwise with malformed xml the dtor is called in a place where
the processing of the formula apparently corrupts the internal
representation somehow and it is impossible to write the document
back as docx. As a bonus, all UI elements related to saving get
disabled, so the document then cannot be saved at all. Although
this is only with malformed xml, which hopefully should never
happen, it's still better to avoid this. Moreover it seems to be
the proper way of coding the handlers anyway.