allow to install desktop integration in Linux dev build
The desktop integration searched for the package name by the dash "-"
delimiter on many locations. It was not easy to add support for
package names with the dash.

In addition, there was an inconsistence. Most package names used the
"libo-dev" prefix name but the ure packages used "lodev". It was
installed into the prefix /opt/lo-dev.

So, I remaned everything to "lodev". It removed the inconsistence
and solved solution with the dash.