add docx starmath testing documents
It's simply a bunch of documents with various math features in them.
It'd be nice to have an automated test built on top of them that
checks the formula is the same (roughly, in the meaning, it can't be
precise because import adds e.g. extra {}'s) after an export+import
roundtrip.