use PACKAGEVERSION as the base of MSI ProductVersion
Formerly PRODUCTVERSION was used but that did not contain
the microversion digit. It is better to have it in MSI
ProductVersion, because a higher build number of a lower
microversion looked newer from the MSI point of view, and
update failed.

MSI ProductVersion is now major.minor.micro.build, e.g.
3.6.1.1206 (build number is taken from minor.mk)