sw: build hidden debug features in all cases
To activate the layout and node dumping features in sw, set the SW_DEBUG
environement variable to anything before running the office. F12 will
then dump the layout as a layout.xml file in the working directory, and
Shift+F12 will dump the nodes structure as nodes.xml at the same place.