Register chart data ranges via tokens rather than string.
Doing it this way avoids having to re-generate the data ranges in
Calc A1 before passing it to the chart backend in Calc.  We need this
in order to remove the silly restriction that forces us to always pass
data range strings in Calc A1 format, which is error-prone.

This is also necessary in order to fix the bug that prevents editing
data ranges of an existing chart when the formula syntax is something
other than Calc A1.