Should use gb_Library_add_libs and not gb_Library_add_linked_libs
for system libs, says dtardon, and I trust him. Note that this crap
still doesn't work any more for Android, though. Some problems with
the gnustl_shared library. I have no idea where it needs to be
mentioned for the bloody gbuild mechanism to understand what kind of
library that is.