fdo#44147: Sheet 0 is selected by default.
ScViewData's default ctor sets current sheet to 0, but its selected
sheet data didn't have it selected which meant that the two data
were out-of-sync when created.  This affected entering data into cell
when the sheet was embedded (because no sheet was "selected").  In
reality at least one sheet is selected at all times.

The above bug was present before, but was masked by a hack in ScTabView
which always selected sheet 0 when activated.  I removed that because
it was interfering with syncing the sheet selection data before and
after print preview.  And removing that hack exposed this.

This commit is the right fix.