Add import of docvariable fields from word documents
Hi,

Attached is a patch that adds support for importing docvariable fields
from microsoft word documents. The variables themselves are already
converted correctly, so this patch only ties the field to the existing
custom document information.

Please consider this patch for inclusion.

I declare that this and my further patches are under LGPLv3+ / MPL license.

With Regards,
Kenneth Aafløy