remove pointless forward moving of a table (bnc#706138)
This appears to be broken. Why should a table where all rows want to be kept
together be moved one page forward just because there is something
before it on the page where it is? It either fits its curent page, in which
case it's wrong, or it does not fit, in which case it may not fit the following
page either and something else needs to take care of it anyway.