Fix UpdateCheck clog up server by keeping connection alive
The update check protocol uses the webdav ucp, which uses neon,
which in turn defaults to keeping the connection alive. this is nice
for webdav, but disastrous for millions of clients phoning home to
the same server.

Using neon directly in updatefeed.cxx is not an option, due to the
thread safety problems around that (see e.g. rhbz#544619) - so we
had to extend it to accept connection options, and reuse the
existing webdav provider.