gbuild: UnoApiTarget: initialize target local vars
Fixes the problem in offapi, where a rebuild after changing an IDL file
would produce an offapi.rdb that contained the stale content of the
old version of the IDL file.  This was because in offapi 2 rdb files are
built, offapi.rdb and types.rdb, and types.rdb is a merge of udkapi.rdb
and offapi.rdb, hence it depends on offapi.rdb.  Unfortunately this means
that the UNOAPI_MERGE variable for types.rdb is inherited to
offapi.rdb, with the result that after the workdir offapi.rdb is built
from .urd files, it is overwritten by a merge of udkapi.rdb and a stale
offapi.rdb from the solver.