fdo#40261: Fix crash in XML Form Document.
The data structure holding the UI elements in the browser listbox was a
terrible mess - it held the items in an unordered_map, but then accessed
them via a vector containing iterators to this unordered_map.

Fixed the problem (and cleaned all this up) by removing the vector of
iterators, and turning the unordered_map into a normal vector.  When we need
access by name, we just go through all the items; it is always just a handful
of them anyway.