add back close button to the about dialog
Some windowmanagers on small devides do not have the close window button,
so all dialogs should have its own way how to diappear.

It is now done like in other application. They use "Close" button instead
of "OK". They put it into the right side and delimit it by a line, so
it looks like another dialog in the application.