WaE: expression result unused
Ifdef out a couple of weird leftover lines that effectively do
nothing. The code was changed to be like this in 2004. Could some
Calc expert have a look please and clean up? Also, if the intent
really is that bEmbed is always false, a couple of lines below it is
pointless to check if it is true anyway.