Rewrite of the format paintbrush
With this patch the format paintbrush differentiate automatic character
attributes apply to character than those apply to paragraph.
Character attributes applied to paragraph are treat as paragraph format
when paragraph format are copied (and so apply to paragraph), and treat
as character format when paragraph format are not copied.

It also change the behavior of the format paintbrush on "restart
numbering" and "restart numbering at" attributes. With this patch those
attributes are simply copied… as expected.