tubes: quick hack to allow running two instances
To record a screencast, I wanted to get two instances of LibreOffice
running in the same session. But the Client object(s) claim D-Bus names,
so initializing the second instance failed.

This patch changes things so that if LIBO_TUBES=master, no Client
objects are created. As a result, only the slave can receive files.