resolved fdo#40021 don't let CSV import get confused by erroneous HTML detection
HTMLParser::IsHTMLFormat() is convinced that anything containing a valid HTML
tag would indeed be HTML, which is a rather idiotic assumption for us in the
case of "foo <br> bar" with a preselected CSV filter. So keep this detection to
the end.

The original order where preselected CSV had precedence over others was changed
with 9f1cc58c20ee365ff2a158ad69c1091e6ad11ac6