enable -fno-enforce-eh-specs only for non-dbgutil builds
The code uses exception specifications, so it doesn't make much
sense to have them, but disable them using a gcc option. MSVC
ignores them completely, so this would leave only clang as
the only compiler to enforce them. Either the majority compiler
needs to enforce it at least in dbgutil builds (in product builds
the option can be seen as a kind of NDEBUG), or alternatively
the exception specifications should be removed.