introduce ENABLE_DEBUG_ONLY
with this you can now do:

 ./autogen.sh --enable-debug="sw sc"

and your build will have debug version of the libraries sw and sc, but a
non-debug build elsewhere. This currently only works for libraries. It would be
also possible to extend that to excutables, but that would need explicit in the
enable-debug switch then:

 ./autogen.sh --enable-debug="Library/sw Library/sc Executable/mkunroll"

so I dont know if it is worth it.