n#757910 fix WW8 import of textboxes with thin border/inner margin
There were two problems here:

First, nLineThick was already in twips, so converting it again from EMU
rendered thin borders invisible.

Second, when the border width is set to zero, that means a minimal
border, but we interpreted it as default width (which resulted in less
space for the textbox contents than required).