Resolves: fdo#33599 cursor in notes is reset to start on focus out/focus in
SidebarTxtControl::LoseFocus calls mrSidebarWin.UpdateData which takes the text
from the postit if the postit has been modified and updates the field from it.
That updates all listeners of that field which includes the postit itself,
which in response updates its content from the field, which causes the cursor
to be disabled, i.e the original contents have been removed and replaced with
identical content, but the selection has been invalidated.

Change-Id: Idcde25a85a3a657a6057e70a59e2219e224e009e