split up top-level Makefile:
The top-level Makefile invokes autogen.sh (and thereby configure) in an
environment which is polluted by config_host.mk; this causes at least
the problem that following a "make clean", the bootstrap script will not
copy dmake to its destination because BUILD_DMAKE=NO is set from
config_host.mk, which is apparently due to the PATH being polluted from
config_host.mk, so configure finds the dmake in the build tree.

So split up top-level Makefile into Makefile, which invokes autogen.sh,
and Makefile.top, which does everything else.