Add stdout and stderr redirection to the Android log
On a (non-rooted) device it is not possible to set the
log.redirect-stdio property so that it would be effective (i.e. read
by the Zygote process when it starts). Such redirection has to be done
in-process.

Add a (JNI-callable) method to set it up: Point file descriptors 1 and
2 at pipes that are read by a thread that logs each line through the
Android logging API. Code based on Android's own logwrapper.c.

Change-Id: Id5308293595096a44a2ffed2dbc0c252be109de7