rework getting shape for the .docx import filter
This makes both sw/qa/extras/ooxmltok/data/n705956-{1|2}.docx testcases
work at the same time. As far as I understand it, the .docx filter
calls into oox and at the end if gets the shape, while the shape
is removed from the ShapeContainer. In the case of recursion caused by
<w:pict><v:shape><w:txbxContent><w:pict><v:shape>, clearing the whole
list is wrong. Also, the OSL_ENSURE seems to suggest that there should
not be more than one shape inserted by each read. So simply consider
maShapes to be a stack when importing .docx .

Change-Id: I7263aeaf74c8d31a05e64c56b880cbc6b00d7dd6