Delete SfxItemSet before the current shell gets destroyed.
SfxItemSet takes hold of the SfxItemPool instance from the current
shell, and accesses it when it gets destroyed.  The problem arises
when the current shell gets destroyed before the SfxItemSet instnace
does, in which case an illegal memory access ensues.

This fixes intermittent crashes when opening a document in Writer.

Change-Id: Ib5e74b43051f868f22f6efdb311e6c2a75326d9a