Don't test cairo if enable_headless is set
I've ended up using a feature switch, i think it is the less
risky way of handle this, assuming that enable_cairo_canvas is
enough to mark cairo as required, as in my previous try,  may be
not that safe.
The diff looks big because of reindenting a whole block.