fdo#43989 let unopkg.exe run with elevated privileges during install
... so it can do the pre-registration of bundled extensions in
INSTALLLOCATION/share/prereg/bundle.
The meaning of Type of custom action is the following:
msidbCustomActionTypeDll + msidbCustomActionTypeBinaryData +
msidbCustomActionTypeInScript + msidbCustomActionTypeNoImpersonate +
msidbCustomActionTypeCommit = 3585
So it is a binary in a dll from the MSI itself, it runs deferred
with system privileges and after commit, because msvcrt90.dll is
available only after commit. The same type was chosen for the
removing part, it works.

Change-Id: I14c347b5cbc47789eb0484769a676703361c2193