fdo#50141: Character count (with spaces) incorrect with numbering on.
There was a typo in the character counting code of text nodes. When a numbered list is detected, the code attempts to add the character count of the "item number" to the total count, but ends up overwriting it instead (uses '=' instead of '+='). This resulted in the item number strings to be the only ones counted in a numbered list.

Change-Id: I9d96311065f66026dd4f05d3cd5ae61e2c0dafb4