writer:crash(i118878) when adding columns and chang. properties in a frame
when adding columns, GetNumCols() reflect the new number of columns
before the actual columns available are updated.
Due to the way and order in which properties are updated
the code that adjust the columns can be run before the new columns
are avialable.. and was using GetNumCols().
This patch avoid using that api and rely on the column iterator
to determine the acutal number of column available.

Change-Id: I52e75ec82c1ace65b05b8df9e54f4689f2aec8d3