Resolves: fdo#45869 we want page preview text to layout for print output
We want text to be positioned as it would be for the the high dpi printed
output, not as would be ideal for the 96dpi preview window itself. It is
supposed to be a print preview surely.

This should avoid getting different text positions/widths shown in the "page
preview" of calc than gets used when printed/exported to .pdf

Change-Id: Iea3acce3c1f2a7051007c22768329c52a7c1d4ee