Compress Graphic functionallity added (currently only for Draw)
This is a initial commit of "Compress Graphics" functionallity.
Compress Graphics is accesible via pop-up menu option to graphics
objects. Currently this funtionallity is only available in Draw.
When the menu option is triggered, a Compress Graphics modal dialog
opens to define DPI, compression level and type of compression.
When OK is clicked, the graphics in Draw is compressed and exchanged.

The current implementation is not final as it needs handling in case
when the graphics is croped. Additionally cropped areas could be
removed when compressing, depending on a checkbox in the dialog.
Special cases of graphics need to be checked too, and some tests
would not hurt.

Change-Id: I9ec1a71f6ff4364a49b2521e42d88c3ba224dd87