Resolves: rhbz#836937 insanely slow with Zemberek installed
Zemberek is a java spellchecker extension. With it installed the collecting of
spellchecker information on first activation of spellchecking is insanely slow
as the cache of spellcheckers is thrown away on each iteration through each
language known to LibreOffice.

So...

move the config updating stuff from editeng down to the linguistic layer. Let
the linguistic layer keep its spellchecker cache and listen to the extension
manager to know if an extension which might be a spellchecker one has been
modified and only throw away the cache on that event or if (existing
implementation) the config data for linguistics changes.

The polling of changed linguistic data in SvxLinguConfigUpdate::IsNeedUpdateAll
can be removed and leave it up to LngSvcMgr to load everything in its ctor and
keep itself up to date with its config and extension listeners.

Change-Id: I9c93d998928e2e7f5128c36771b3e450a8057cd6