Rename FieldEntry to WW8FieldEntry
When compiled with Clang on OS X, the OS X dynamic linker appears to be
confused about which FieldEntry destructor to call at runtime. In some
cases, the FieldEntry destructor in register/source/reflwrit.cxx is
called instead (both destructors have the same symbol name). This patch
avoids this problem.

Change-Id: I7d69894318cb8fda0a7c5a9b1c2ff3ca0d47a37c