'make build' again generates .msi on windows
This is a partial revert of 0ec45dc41dd4872482b817387f253d6fab6bf20c ,
mainly for tinderbox nightlies. The .zip for dev-install is built
only in dev-install target. But IMO the build target is broken
this way too, I don't see why the .msi has to be built already there
and not in some install or make-msi or whatever target (it's pointless
in build target for a developer build).

Change-Id: Ifd63066499b67fa446127193b243d893d497b451