Import intended blank or error cached formula results instead of 0
There are cases where a formula cell is exported with an office:value of 0 or
no office:value at all, but the formula cell will have a text:p value which
contains the intended formula result.  If these cases are not taken into
consideration during import, a 0 will be displayed in the cell
instead of the intended special formula result (blanks or errors).
These cases include when a formula result:
 - is blank
 - has a constant error value beginning with "#" (such as "#VALUE!" or "#N/A")
 - has an "Err:[###]" (where "[###]" is an error number)

Change-Id: I8068cf1c9809c02513662f2b0a6fb16eb44920e0