fdo#53006 Force reinstall of all bundled extensions on upgrade
...as the per-user data about bundled extensions can otherwise contain stale
$BUNDLED_EXTENSIONS_PREREG references if the old installation used
share/prereg/bundled/, the new one does not (cf. the fixing for fdo#51252 "LO
cannot start (reports runtime error with Visual C++ Runtime Library)"), and a
bundled extension did not change version.

It is safe to tie this behavior to the existing "force" parameter of
dp_misc::syncRepositories, as the only place that calls it with force=true is
the call to Desktop::SynchronizeExtensionRepositories(newInst) in Desktop::Main,
where newInst=true is the relevant condition for this behavior, too.

As stated in XExtensionManager.idl, "this [...] can go again once no exisiting
UserInstallation's user/extensions/bundled/ data can contain any
$BUNDLED_EXTENSIONS_PREREG references any longer."

Change-Id: I630dec8f2d20282ee47c65ac61ed2c9b062448e1