support for deferred property processing in writerfilter
There currently does not seem to be any sane way to process
an attribute or sprm that depends on another one that may not
possibly be there yet (e.g. in e7ab4bb6b0e83f01148ffff41e8c5eaa0c5ba0a4,
or w:position which for Svx internal reasons depends on fontsize
and thus w:sz).
So make it possible to defer such properties and process them
only before they are actually used, instead of trying to get them
out of PropertyMap, possibly in more places and possibly having
to undo the changes that have been done to them already.

Change-Id: I1630057ecdf46443647ec1dd5253983ae15a083f