Related fdo#53006: Do not instantiate service manager too early
... in soffice_main calling GetCommandLineArgs, before Desktop::Init takes care
of synchronization of the per-user bundled/shared extension information (as the
latter potentially modifies data that the service manager consumes upon
instantiation; this e.g. lead to bundled extensions not working during first
start after an upgrade).

The only reason GetCommandLineArgs needed ensureProcessServiceFactory appears to
be the ExternalUriReferenceTranslator.  So defer its usage to when the relevant
cmd line args are actually processed (which, luckily, does not yet happen in
soffice_main's usage of GetCommandLineArgs).

Change-Id: I6ebbf0a4ad1c6f64c8fbbe2b0d7628fa42a1afb6