fix "uncover" slide transitions with presenter console
It turns out it is really not a good idea to render _anything_ on the
first run of performIn/performOut in classes derived from
SlideChangeBase. That is because the functions are called separetely for
every view, so, if there is more than one view, like when the presenter
console is active, it will be done only for the first of them. This
commit introduces another function, prepareForRun, that can be used
reliably for that purpose.

Change-Id: I2c690fb145ccc05a680d4c2991c73900336e6f42