fdo#53001 disable unlink graphics in image control dialog in base report builder
Embedding graphics in a report builder report is not implemented,
so do not propose it to the user.
It already was not proposed the dialog raised by
"Insert->Picture->From File"
This commit disables embedding the UI when selecting
an image for an already existing Graphic control
(as opposed to inserting a new graphic control).

Change-Id: I107bc59bef8bb29a75140542cc5c2779a37ed113
Signed-off-by: Lionel Elie Mamane <lionel@mamane.lu>