fdo#52451: Ensure that the extension for text is .txt from writer.
Added a preferred extension property to the filter which overwrites
one(s) given by the default file format type associated with it.

Normally this property is not used for the majority of the filters,
but for plain text filters, we need to use .txt when saving from
writer whereas we do .csv from calc.

Change-Id: I3453ad084e4379e5cc1101daeb9d2c3ddd919e92