Import bezier curves from .docx.
A shape element with a path attribute is interpreted as a bezier. We can
support both curve and line commands by using the correct co-ordinate
flags when using the UNO interface (omit control points for lines).
Alternatively, curves may be specified using the to, from, control1 and
control2 attributes of a curve element.

Note that when setting the PolyPolygonBezier property of the xshape, the
shape size and position are set too large somehow. The workaround for
now is manually forcing the pre-calculated shape size and position.

Change-Id: If3563474df357d161a2ed2e7a20856d5aed80a94
Reviewed-on: https://gerrit.libreoffice.org/649
Reviewed-by: Miklos Vajna <vmiklos@suse.cz>
Tested-by: Miklos Vajna <vmiklos@suse.cz>