workaround broken uno implementation with ScRangeList in ScCellRangeObj
ScCellRangesBase uses a ScRangeList to keep track of the range but the
ScCellRangeObj class which inherits from it can not deal with more than
one range. This is a temporary workaround for this problem. We need to
think about a better solution for ScCellRangeObj to prevent such
problems in the future.

Change-Id: I4403e55f605267ef99eb6a73302366cf024b3866