old Mac linker is picky about argument order - static libs last
and also after stuff that references external symbols, as it only looks
for symbols it has encountered so far in the object files. Stuff that
comes after the specified library then results in symbols not found
despite the lib being on the commandline earlier.

(another solution than reordering the arguments would be to use
-Wl,-all_load)

Change-Id: If4add0d3f81b13874c1620406fa40fcbd25b7cf0