Move GenPoEntry out of header
GenPoEntry is part of implementation rather than
an autonom class thus its place is in cxx file.
Because of pointer members we have to override
copy constructor and copy operator.
In PoHeader class these methods are unneeded
so make it non-copyable.
The m_bIsinitialized member's meaning widen with
that m_aGenPo points an object. So check it not just
in getter but in setter methods too.

Change-Id: I26ebb885c81d14820076e0d46625c60cc1cf7984