writerfilter: clean up PropertyMap::insert mess:
Rumor has it that MSVC 2010 has problems choosing the right overload in
PropertyMap::insert call, which is no surprise as that is a ridiculous
piece of garbage.  Special bonus points for having another class
TablePropsRef that must have a call compatible method because it's used
as parameter to the same templates, and for the call with a pointer
converted to boolean implicitly in StyleSheetTable.cxx:163.
Also remove the silly boolean parameter while at it, just have 2 methods
instead.

Change-Id: I022d5dd7008f3e6c510d94e19fbacbf04bcf1986