Use -Zi when debug info wanted with MSVC
It really sucks to have to propagate debugging options to the compiler
in an ad-hoc fashion like this. But then each dmake-based external
library module really is an individual, especially the ways they are
built with MSVC.

Liblangtag is especially exotic, as it uses autotools and libtool(!)
also for a MSVC build. Which works, but just barely, almost by
accident. For instance, it ends up using the Cygwin ar and ranlib on
MSVC-produced object files. Which luckily seems to work at least with
MSVC2008.

Change-Id: I1c5244b2a599775c04ae8a003e4a096a0a11648f