deduplicate OString std::basic_ostream operators
I wanted to add one, based on OUString one, and Stephan on IRC agreed
it's a good idea. Then it turned out we already had it in sal unit
tests, sdext and unodevtools as well. Get rid of all these, and have a
single implementation in string.hxx, so it can be used everywhere.

Change-Id: Ibb820f20a138d055fd87036253c627fb862a31fe