Revert "fdo#38878 upgrade xpdf to poppler"
Also revert my two minor follow-up popper commits.

Poppler doesn't build on Mac OS X. It uses fontconfig, and we don't
depend on that on the Mac.

There is a patch from 2009
http://lists.freedesktop.org/archives/poppler/2009-July/004932.html
(actual patch at
http://lists.freedesktop.org/archives/poppler/attachments/20090712/f3b154d0/attachment-0001.obj
) that makes poppler not use fontconfig on the Mac, but it was never
upstreamed.

And that patch uses the obsolete ATSUI API, which we want to stop
using in LO code, not introduce new uses of.

This reverts commit 39f9eb0d856293a974e2f024b29dca20068af073.
This reverts commit 0941c211858416f6117548e093b428b64785f98a.
This reverts commit a6e9bb4c7bad649cb44efaeef67fa18090348c53.