python3: add module for internal Python 3 build (not active yet)
The module builds here on Fedora 17 and with MSVC2008.
MacOS X is unfinished and probably breaks, which is why the module
is disabled now.

These patches from module python were dropped:

Integrated upstream:
- Python.mipsel-py4305.patch
- Python-2.6.1-py4768.patch
- Python-2.6.1-py2422.patch (modified, use --with-valgrind)
- Python-2.6.1-urllib.patch
- Python-2.6.1-py8067.patch

Obsolete:
- Python-2.6.1-svn-1.7.patch (migrated to non-toy HG now)
- Python-parallel-make.patch
- Python-2.6.1-nohardlink.patch (no idea why that would be needed,
                                 NFS should support hard links)
- Python-2.6.1-sysbase.patch (Solaris 11 setsolar specific patch)
- Python-2.6.1-cross.berkeleydb.patch (berekeleydb removal)
- Python-2.6.2-bdb48.patch
- Python-2.6.1-vc10.patch (upstream supports vc10)

An attempt to cross compile with mingw that proved unsucessful according
to dtardon; there is upstream work on this topic that is possibly
already in 3.3: http://bugs.python.org/issue8067
- Python-2.6.2-cross.patch
- Python-2.6.2-cross.fix-configure.patch

Change-Id: Iba9a3cab955983e173e12110f93a6f381d86f9ce