n#780645 DOCX import: add support for fixed column width
In general, DOCX tables provide a grid for the table, then use spans in
case different number of cells are used in different rows. In this case,
the cell width is ignored, as the wished width can be counted from the
grid and span values.

However, it's also valid to simply provide more cells then the amount
defined by the grid, and in this case the cell width should decide the
final width of the cells. This commit adds support for this later case.

Change-Id: I0dd6c1f0c06f81c2afa00489b7ad1f33ff300a7c