Header/Footer: fixed click on frames over the header/footer area.
The main problem here was due to the click to show the separator when
there is no header or footer under the frame. That code was just exiting
MouseButtonDown too early, which isn't good.

Moved the seperators showing code out of the SwCrsrShell as it has no
clue about the document-position and may lead to problems.

Change-Id: Ia62333b52c5d0659bbe7ec4fcfe6e664db132435