make presentation spanning all screens work (bnc#758138)
This reverts an API "improvement" from 6ee5dfa150d408179e20a5525ff7ec46649e0e58
that tried to use -1 as the current screen/display but failed. The check
for nDisplayScreen being in range broke all-screens (which is
represented as last_screen+1 value, which is broken in its own way, as e.g.
the screen count can change). The same way -1 as the default value is broken,
as everywhere else invalid invalid screen number means spanning all screens.
The extra overloads without a default value should take care of that.

Change-Id: Ie118038eacacebc007c25860732d5904ad0de2b9