Introduce a 'highlighted' state which differs from a 'selected' state.
And use that to only highlight child entries which should logically be
treated as not-selected but should have the same appearance as the
selected entries.

A 'selected' state is logically registered as 'selected' as well as
visually.  A 'highlighted' state should only appear as if it's selected
but logically it's treated the same way as non-selected entry.

Change-Id: Ic4bc6923c7678044cf552194ad9865371465c614