Related: fdo#44664 spadmin support for forcing "postscript of driver level"
pdf of 0 and ps of 0 == automatically pick pdf if pdf is default backend.
Which means that spadmin wasn't able to disable pdf if one selected
"postscript of driver leve", though picking an explicit postscript level
would work.

So we need to tweak things to use the pdf flag to be either automatic,
explicitly pdf or explicitly ps

Change-Id: Ic5382a3eb9cbcff39d363723825e75851d73520e