Disable MSVC warning C4986
This is a new warning in Visual Studio 2012.

Otherwise, when using /Wall (as we do), you get tons of warnings from
the compiler's own headers:

.../crtdbg.h(1041) : warning C4986: 'operator new[]': exception
  specification does not match previous declaration
.../new(79) : see declaration of 'operator new[]'

It seems to be generally recommended not to use /Wall with MSVC (that
option does not really have the same intended use as gcc's -Wall,
people say), but use /W4 instead:
http://stackoverflow.com/a/4001759/259398

So maybe we should change that -Wall to -W4?

Also, we should go over the long list of -wd options and check each
whether we really want it or not. Maybe, while at it, add a comment
for each giving its one-line meaning.