Avoid activating window pane when launching popup window.
Else, activating another window pane where the cursor is not would cause
the pivot table field popup to immediately get dismissed after launch because
ClickExtern() gets called from ActivatePart().

For example, imagine a field popup button is on C2, the cell cursor is
on E5, and the window is frozen between rows 2 and 3.  Clicking the popup
button in C2 would launch it for a brief second and dismiss it immediately
afterward.  Conceptually, launching a popup in another pane shouldn't
activate that pane anyway. So it makes sense not to activate it for that
scenario.

Change-Id: Ib970cb898fb9c79d254411e2519cfce74c60b72f