Make MSBuild actually run the 2012 compiler when that is what we use
The /p:PlatformToolset=v110 /p:VisualStudioVersion=11.0 options are
needed to indicate that we want that. Otherwise MSBuild, when
presented with a .vcxproj file created by VS2010, will assume it is
the 2010 compiler that should be used. This will then lead to link
errors pointing outthe incompatibility when linking the static
liborcus into the scfilt DLL.