Add a few missing files to be expanded (from *.in) to CONF-FOR-BUILD
But note that currently the config/config_*.h files are created and used so
that those created for the HOST platform are used also when compiling (the
cross-toolset) for the BUILD platform. Seems to work by luck so far, but this
is really broken. We need to have the config_*.h files separately for HOST and
BUILD.

Change-Id: I500694eee8bb2a8d2d7afbf78171c0e2ea1e66b3