The LineListBox drew lines wider than can fit in its text area
it takes the outputsize of the entire widget and take a little
(not enough) off that as the line width. The line is always
wider that space available, so with widget-layout enabled it
will request more space. On the next update of the lines it
repeats, so listbox grows ever wider.

Change-Id: Ib43239d297bb0ba0dc1096acf347ef17b5e41273