Check also for a (self-built) 64-bit NDK tool-chain
Also make the --with-android-ndk-toolchain-version option mandatory. Otherwise
it would be possible to leave it out and that would appear to work in the
configure phase but in fact mix stuff up when using a self-built 64-bit
tool-chain merged into a r8b,c,d NDK.

Sigh, maybe we shouldn't even try to have the configury generic enough to work
with any NDK version (which probably it doesn't anyway), but explicitly accept
only a small set of NDKs, like r8b,c,d (which probably are the only that we
work with anyway).

Change-Id: Iecdf3a2d841e7b29598228ae49f48eca9640a82c