support new hyperlink cell attribute
Allows a hyperlink attribute to be set for a cell, allows for import of say a
hyperlink associated with a formula cell. Since this patch does not add
any file format support it is not possible to save such an imported hyperlink
to ods. Note: such a hyperlink would not have been imported in the past so this
should not be an issue

Change-Id: Ieb234bb6e98e4a2630b596a90972a75be12a92d4