Use default visibility for exception types even when DISABLE_DYNLOADING
Otherwise the dlsym() lookup of the typeinfos in the C++/UNO bridge
won't work. And it has to, because at least for iOS on ARM the
dynamically created typeinfos don't match those the compiler-generated
code tries to catch, so otherwise exceptions thrown across the bridge
don't get caught. (Fascinating, I am able to talk about this stuff as
if I understood it.)

Change-Id: Ie0d4632743e18719782f55904fb76f45c944f20c