The "unx" and "win" OGLTrans_Shaders.cxx files were effectively identical
Just some OSL_TRACE/SAL_INFO changes in the "unx" one. And it isn't
like the "win" one would actually compile for Windows as is anyway, as
it unconditionally uses glX stuff. So certainly it is better to have
just one copy, and if we want to make the OpenGL transitions build on
Windows then add ifdefs for Windows to it.

Change-Id: If6887e24a0fc03c7e0e77f003d08feb51c264f30