Make the OpenGL transition code buildable for Windows
Whether it works at run-time is another question;)

Use the same source files as for X11, with relevant conditional
coding.

The Windows OpenGL import library corresponds to OpenGL 1.2, so we
can't directly use APIs from OpenGL 1.3 or newer. (Even if they as
such are present in the "Mesa" headers we use.) So look up also
glActivetexture() dynamically.

Change-Id: I2280e2ad7e190356e9409cc28142fa9fc33e078b