Resolves: fdo#59183 Copy 4 or more slides then crash
regression from 17afe4cea7e01aef1e5270cc09f438bc6fde3211 which is totally
forgivable as its riddled with asserts that suggest there should be no
out-of-bounds accesses and there probably shouldn't and those queries are
possibly bugs. But double-checking 3-6 it is the case that non-existant pages
were queried for in that version too, so return NULL on out-of-bounds
like the original pre-stl conversion code did.

Change-Id: Ic918419b6cb76b083de6b6911dde9d6e00258324