xmloff: refactor Generator version handling:
Since there are now 2 forks of OpenOffice.org, we cannot rely on a
simple total ordering of versions any more; add a new function
SvXMLImport::isGeneratorVersionOlderThan(), taking 2 reference versions.

Also extract the LibreOffice version number from the generator string,
and extend the BuildId property to store this as a third number.

This also allows removal of the "fake LibreOffice3 as OpenOffice.org
3.3 release" hack, which is not future-proof.

Change-Id: I44d8105eb537ac43fb9529a8b1b661ae0f2bba30