Use SAL_INFO instead of home-grown msgs_debug
If there is a need to optionally get a thread identifier or the
function name into logging output (as msgs_debug did), we should
figure out a way to do that in some elegant fashion in the sal logging
macros instead of using some local solution in just one place in the
code.

Yes, the iOS and OS X CoreText code should be de-duplicated. Will
happen soon.