default to --without-help and --without-myspell-dicts
Updated the distro-configs so that the default change
does not impact build based on distro profile

The intent is that with this change of defaults
the build does only rely on core and does not need
any submodules. Making a initial build simpler
for new dev, since they won't be exposed to submodules
unless they really need them.

Change-Id: I446e912d607ce7358fd11f54a3de025dc3f25144
Reviewed-on: https://gerrit.libreoffice.org/1791
Reviewed-by: Tor Lillqvist <tml@iki.fi>
Tested-by: Tor Lillqvist <tml@iki.fi>