make clear that this variable should not be used inside of ScDocument
Sadly it is not as easy as it should to remove this variable yet. There
are some users inside the import filters that prevent me from moving the
variable into the UI layer but we should not use it inside ScDocument.

We should inspect in the future if this variable is actually a good idea
at all or if it create problems with multiple views and we need to use
the current table of the current view.

Change-Id: Ic1a378bacfcba65a3df8e5782e64eb15ca057d7d