Start on Bluetooth code for the Impress Remote for OS X
Work in progress. It sets up the SDP service record and registers it,
and that seems to work. Does not do anything with it yet.

I wonder if the current approach to use the standard SerialPort
service class will work for Macs, as they by default have one
SerialPort service running, "Bluetooth-PDA-Sync", whatever that
is.

There is no code in the Android app to handle the case of there being
several Serial Port services on the computer it wants to connect to,
so what happens then when there are two will be interesting to see.

(I can't find in the Android API any way to enumerate services that
use the same service class. If there was, the app could choose the
correct one based on its name.)

So probably we should pretty soon switch to using a dedicated service
class UUID instead of using the standard serial port one. As we
already have the app out in the wild it will probably have to be
modified to look for both Serial Port and our own service then? Sigh.

Change-Id: I1bee6403642e9b8953b5a8fa8fb5fe6a9dc76a09