fdo#60738 SwPostItMgr: let the cursor know if we're in annotation mode
This solves the conflicting requirements:

If the annotation has focus ("in annotation mode"), then we want to
be read-only only in case the comment anchor is inside a protected
section, or so. Ignoring the fact that in the main document, the cursor
is right before the annotation anchor, which is a read-only position.

OTOH, if the annotation has no focus and the cursor position is before
the annotation anchor position, then we do want to be read-only.

With this commit, the first case passes around a boolean flag, so the
annotation frame won't be read-only without a reason.

Change-Id: I25781a4501f752f7c4d024795d57b884cb800547