Resolves: fdo#60627 wrong lib names for some custom widgets
Under Linux dlsym will search other locations and find them if they exist
elsewhere, but not under windows, so its easy to put the wrong lib name in if
developing under Linux because it'll generally work anyway.

So add a script: bin/verify-custom-widgets-libs which can be used under Linux
to verify that the required factory methods exist in the right lib.

Change-Id: Ic30f8da5acc4712684a7a25fbfb003e8b21cb867