fdo#60719 : fix crash on scaling bmp with target/source width/height = 1
The rendering is quite ugly compared to the same pdf opened
with pdf reader. but at least it does not crash anymore.

The function ImplCreateRotatedScaled seems quite sub-optimal
for one thing it is called even when there is no rotation
and still bother with doing a lot of work to 'rotate' things.

It was assuming Width and Height > 1 for both the source and the target
The patch tries to 'hide' the misery... but I'm not convinced
that this is 'right'.

Change-Id: I9aec5eb6655ea3678aa018ae5928dabb8058bec6
Reviewed-on: https://gerrit.libreoffice.org/2195
Reviewed-by: Arnaud Versini <arnaud.versini@gmail.com>
Tested-by: Arnaud Versini <arnaud.versini@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Tomaž Vajngerl <quikee@gmail.com>
Reviewed-by: Norbert Thiebaud <nthiebaud@gmail.com>