Use actual size of view instead of hardcoded 1000x600
The View size is available only after the view has been connected to the
activity, it seems, so move the Bitmap creation to onDraw().

Note that the code in SvpSalFrame::SvpSalFrame() in vcl/headless/svpframe.cxx
still hardcodes another (!) size, 800x600. This affcects the size of the
desktop-style "top-level window" displayed by the android/experimental/desktop
app. I didn't yet figure out the right way to pass the actual view size to the
SvpSalFrame. And there is also a hardcoded third (!) size, 1280x750, in
AndroidSalInstance::GetWorkArea(), although I don't know what that affects, if
anything.

Change-Id: I042bf764cd66efa7069c36601170b90d57fa174c