Make it easier to debug the app by sleeping for a while if a property is set
If the property log.tag.LODesktopSleepOnCreate is set to "VERBOSE" then sleep
after liblo-native-code.so has been loaded to give the developer a chance to
start ndk-gdb and set breakpoints. Yeah, a bit silly to overload a logging
property like this, but it was the first idea I came up with.

Change-Id: I665f87778d083d2d167a5d16f24e2d50b1fba042