bin PATCH_SO_NAME crack
In VersionInfo of Windows executables the FileDescription property originally contained
80 times 'x' character. These executables were marked as PATCH_SO_NAME and later the
installer maker perl program replaced the xxxxxxxxxxx to the actual file description
which was defined in scp2. WHy was it good, I have no idea. The problem was that it
happened after postprocess, where we signed the binaries, thus diigital signature became
corrupted. It is better to give file descriptions via makefiles.

Change-Id: Id5ad4470bb7a6313b33fbba09d72d9a009163a89