extensions: WinResTarget_activex: use absolute paths
This does not build from the top-level because it uses a relative path;
the command invoked from "make" runs in the top-level dir while the
command invoked from "make extensions" runs in the extensions/ dir, so
we have to use absolute path here.  Presumably the weird substitution
going on here is just a relic from the days when WORKDIR was variously
converted with cygpath.

Change-Id: I73a61699ec0394bd40e63864f3e9f566c7d28a32