fix for fdo#61801
... "crash on Tools - Options - LibreOffice - Personalization - Select Background Image"

The root cause of this is my conversion of the UNO code to use the
new-style FilePicker service constructor in commit
4b51374a7021d52f7f1be1861e2ee6a011b30ecd

Unfortunately, the new-style service constructor will always call
initialize(Sequence<Any>), which the old code did not do.
And initialize does not like being called without any arguments.

The cleanest fix for this is to simply remove the createDefault()
service constructor method and make all the call sites explicitly
choose the style of FilePicker dialog they want.

As a bonus, this simplifies some of the call sites.

Change-Id: I75f5e03cff4e39abe22cd9650a079ec78ab636c4
Reviewed-on: https://gerrit.libreoffice.org/2553
Reviewed-by: Kohei Yoshida <kohei.yoshida@gmail.com>
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>
Tested-by: Stephan Bergmann <sbergman@redhat.com>