Test all file format types regardless of document services.
The old type detection would only test file format types that are
relevant to the current document service.  But this would not work
if the user tries to open an Excel document whose name ends with a
non-default extension (such as .tmp) from Writer UI.

To amend this, we need to test all possible format types that are
supported by all our modules regardless of the current module, but
prioritize them per current module.

TODO: The above scenario doesn't work yet.  I still need to fix
the individual type detection services to get it to work.

Change-Id: I51dbdff28299aa64cedc6752b9ff82b96d5d79cf