fdo#60562: List *all* format types that we support and rank them all properly.
I did my best to check each and every one of them and rank them to the
best of my ability.  Any mistakes are to be fixed as they are discovered.

Note that while working on this, I've noticed that we don't actually have
any real type detection codes for many of these format types, especially
the graphic file formats.  If we ever have trouble loading any file type,
check if it's caused by not having any detection code for it, and if so,
we should write one for that file type.  That said, my casual tests on
loading graphic files didn't cause any trouble.

Change-Id: I1398cbb72b2d4ccf75434a9aa0d2a5c5d17bdaad