embed also view-only fonts, but do not use them
MSO embeds even fonts which allow only embedding for viewing the document
but not editing it. So embed such fonts too, but do not actually use
them from the document. What MSO does when such a font is not present
locally when opening the document is switching to read-only mode,
warning about this and providing a button for switching to editing mode
by dumping the font(s). That should be done for LO too, but right now
dropping view-only fonts is better than using them for editing.

Change-Id: I19c28fadb091e6b21beaf4cbf8b47e3078256d1c