Try to make the scrolling and zooming actions snappier
Now it does work nicely during the gesture when all the action is on the Java
side (translating and scaling the pre-rendered bitmap). Looks a bit sad, of
course, that nothing scrolls in to replace the parts of page(s) scrolled out
during the gesture, and correspondingly for zooming.

To then get the stuff down in the murky depths of the LO code to do what I
want still is beyond me.

Change-Id: I9ce33ed482013d18a877d1798de3bce5ac608e5e