Migrate also Java and Online Update setting (fdo#57061)
The format of the javasetting*.xml file is the same, so why not migrate
it. Note that it does help for 4.0 because we modified javavendors.xml.
This would cause release of the java setting even when they are not
migrated. Well, the migration might be useful in the future, for 5.0 ;-)

I was slightly scared to migrate the /org.openoffice.Office.Job
but registrymodifications.xcu should include only stuff that can
be modified. Minor updates handled this well. AFAIK, we did not
change anything in this area for 4.0, so it should be fine.

Change-Id: I7775c238a7c298bf877241429c197e76f2093f5d