Related: fdo#62154 correct font lang is a requirement for applying locl
If the language of the font changes we have to throw away the "magic" cache.
Otherwise we continue to use the parent paragraph style's font language, which
isn't going to fly if we want to propogate the language down to vcl to give to
icu to allow it to apply the expected locl tables.

This is the fruit of an 8 hour debugging session. I hate writer.

Change-Id: Ic32f7a223b527e702370d5d8b9f9ede59efd8a80