initial work on an odc export
We can now export a chart to odc when we are in chart edit mode from
Calc. I need to add support for it to Writer and Impress as well.

We can already open these files but copy&paste from the opened file
fails. The next step is then to add a new menu entry
Insert->Object->Chart from file

Change-Id: I14d1702e79517e7319a1929de2be5501d375885d