follow patch for fdo#38385 attempt to detect rtl
tbh this is somewhat of an optimistic attempt to swap start and end borders
if needed. I am not at all sure though that I am dectecting the RTL-ness
of a cell in the correct way. There are some comments in the code in any case
that reflect my uncertainty ( hopefully they will be of use if/when some tweak
is needed )

Change-Id: Ic3099fbab08e46899ca311c81edbcad9bf5ab2a6