Use the make binary that configure found in PATH
When auto-detecting the GNU make binary from PATH store its absolute
path in the GNUMAKE variable to ensure to keep using exactly the same
GNU make binary even when the PATH settings change. This is needed
because the gbuild system relies on the GNU make features detected and
reported by the configure script.

Change-Id: I05916f0ad7f2e98e4e908455ed72758ebf396863
Reviewed-on: https://gerrit.libreoffice.org/2924
Reviewed-by: Michael Meeks <michael.meeks@suse.com>
Tested-by: Michael Meeks <michael.meeks@suse.com>