Export current Writer page as jpg/png
This is first step to rework of graphic exporting. The idea is to
replace the exporter that works only for Draw/Impress and replace
it with a general exporter for any object. With this it will be
far easier to export objects as charts. Currently only Writer is
supported and only jpg/png.

Additionally, this commit introduces a new Export dialog which
supports setting the pixel width, height and DPI.

Change-Id: I7302b26bd432840d7ef0c3d2d2e13ff150cd2a07