Make autogen.sh prefer to read an autogen.input file
For backward compatibility, an autogen.lastrun file will still be used if it
exists and autogen.input does not exist. The recommended workflow is now to
keep the configuration parametets in autogen.input which is never written by
autogen.sh.

Most (?) developers used to treat autogen.lastrun as a valuable parameter file
to be edited manually anyway, and not as an ad-hoc backup copy of command-line
parameters last used. The name autogen.input better reflects this usage.

Change-Id: I7e3c747fa95e9f2f0bc44036419aaab8f4ad01e7
Reviewed-on: https://gerrit.libreoffice.org/3111
Reviewed-by: Michael Meeks <michael.meeks@suse.com>
Tested-by: Michael Meeks <michael.meeks@suse.com>