Resolves: fdo#61940 unable to change shape fill colors
The tab area page has a magic bit which is set when the page is activated that
enables its properties to be applied on ok. If another page is activated that
also affects area fill using an alternative set of properties, then the bit is
unset.

For layout we have to activate all pages that don't exist yet to get their
optimal size. So we need to re-activate the original/current page after forcing
the other pages to exist in order to make this trickery work.

Change-Id: I90cac4624a12e6c29f6b226ce831f71c7cc49117