Create a justified line if asked to
The end result is still not satisfactory but it is getting closer.

Oh how nice it would have been to have some clean useful documentation for how
the text layout machinery works. The number of entry points in the low-level
"backends" (like CoreTextLayout) is quite small, so I doubt it would have
required lots of text to document the "protocol" used between the upper layers
and the layout machinery. Now one just have to use educated guessing aided by
lots of debugging printout. But this is true for most of the codebase, of
course.

Change-Id: I791f190819a187418865279f6a5b4aa598e01d3c