Check for the C++11 "final" specifier and introduce SAL_FINAL
I think it is useful to use SAL_FINAL mainly as a documentation aid,
to make it clear to a code reader when a class is not expected to be
derived from, and when a virtual function is not expected to be
overridden in a derived class.

Possibly there is also some class of bugs that using SAL_FINAL will
help find?

Change-Id: I45002f020dcb52e8a9f2962ff98780f2b80627af