Resolves: fdo#61948 delayed delete deletes nothing under windows
The reason there appears to be two dialogs, is because there
actually is. The original dialog doesn't get deleted under
windows.

If we use the actual returned type of the dialog here rather than a lower
baseclass then the right thing happens. Original code works under Linux but not
Windows.

Change-Id: Iaf99fa4e16f743b1f5a3d29f42a140491f3d67b0