consume hectares of empty space on outline and numbering tab page
We now have hectares of empty space in the English ui on the right side because
the other tabs are wider than this one.

What to do with the extra space, leave it on the right, or consume it in some
other way ? I'm inclined to spread things around the available space. If we set
our elements to expand to consume it (but left align the spinbuttons within
their expanded space) then we get a page that looks very like the original.

I'm not entirely sure if I prefer it this way because it's genuinely more
pleasing to the eye, or because it's just familiar looking this way.

Change-Id: Iafe60713eb8c6f0bd1ecd70e3f288d52b453a247