Revert "Go back to using Visual Studio project files to build"
I could not get it to build using the project files and VS2012...

With one additional patch to drop the use of dllimport in env.hpp, the
"gcc-wrapper" way builds fine for me, using VS2012.

It is not necessary to specify the dllimport attribute for functions
imported from a DLL. Specifying dllimport if the library actually is a
static library (as it is here) leads to linking errors.

The way the ifdefs and -D options were set up here, the utility and/or
test programs that are built as part of the liborcus build were
compiled using dllimport, and thus failed to link. So just don't
bother with dllimport.

This reverts commit 58b4cd934fd28b796dc12590f33971dd32ec95ee.

Change-Id: I2ffc90d47ae018a9a14b1c6d07fb0a225ce10d6d