add spec. handling for xcu files to gb_Dictionary
The xcu files can either come directly from $(SRCDIR) or be generated,
depending on whether the build is with or without langs. This
distinction did not matter while Dictionary was using ExtensionTarget
for its impl., becuase there the file was just copied to the
ExtensionTarget's root dir. Now that Dictionary uses Package internally,
it is necessary to know where the file comes from to put it into the
right package, either the one for $(SRCDIR) or the one for $(WORKDIR).

(Yes, it would be possible to use just one Package with base dir /, but
I consider this ugly.)

Change-Id: I4549923287abc968f1c042b809db06d688b4c573