fix stupid over-optimization
There's so little to be gained by storing pointers to simple
or refcounted types instead of using an instance. At least
SwTxtNode::GetMinMaxSize() and SwTxtNode::GetScalingOfSelectedText()
pass string temporaries to SwDrawTextInfo, so keeping a pointer
is clearly broken, and MSVC manages to optimize the code enough
to crash because of this. As all the pointers point to const references
and are accesses using const references as well, there shouldn't
be any harm in making a copy.
This fixes smoketest crash on Windows.

Change-Id: I7d7cb42e175a2e64fe9c84c98c0d0204a0c25c13