fdo#59259: Suppress the expensive parts when we are only writing to metafile.
When we are launching the printing dialog, we first draw the page using
drawinglayer to a metafile, and then render the metafile.  Unfortunately, here
we did the real operation of allocating large bitmaps, and destroying them
again; all that just to throw all that away at the end of the operation.

The preview sets the mbOutput to false correctly, so we can just skip the
expensive parts.

Change-Id: Ice77d83100eba339602bbdf374fec8546d4d1e12