SwTxtSizeInfo ctor: take a pointer, not a reference
The ctor later will turn that reference to a pointer anyway, but the old
code made it easy to pass a String, get it implicitly converted to
OUString, and then we took the address of the temporary OUString, later
resulting in accessing already freed memory. Instead, take an OUString
pointer, and then the compiler will warn about a String* -> OUString*
conversion. Adapt one remaining caller accordingly.

Change-Id: I4084dea1d245f0c8919d6afe47c5f391729f6eaf