make parent style for 'frame contents' be 'standard' rather than 'text'
The actual reason for this change is a .docx document, which has a textframe
and a modified 'text' style. The textframe in MSO has 'normal' style
(which is like our 'standard') and so the style is not written to the docx.
LO however uses 'frame contents' for the textframe, and since it was
based on 'text', it used the modified style, unlike in MSO.

But even on its own, this change seems to make sense. It's similar
to 4014d40c96b9d70796a154f0ac48cf5b685cf187, and I see no particular
reason why contents of textframes should be the same style like the main
text of the document, while footnotes, footers, headers, tables, etc.
are different.

Change-Id: Ibe7bfd6a2820cff5cf29c669fd60b575c63500b0
Reviewed-on: https://gerrit.libreoffice.org/3697
Reviewed-by: Bosdonnat Cedric <cedric.bosdonnat@free.fr>
Tested-by: Bosdonnat Cedric <cedric.bosdonnat@free.fr>