gbuild: add Package equiv. for copying whole dirs
This is in preparation for using doxygen to generate IDL docs. Because
the file names generated by doxygen are essentially random, it is not
possible to use a Package to deliver them. And we really do not want to
go back to using Zip, as unpacking slows the installation down.

Change-Id: I5f4aa6bf10c6aab30a5b1870e04dca723c123317
Reviewed-on: https://gerrit.libreoffice.org/3774
Reviewed-by: David Tardon <dtardon@redhat.com>
Tested-by: David Tardon <dtardon@redhat.com>