make Excel happy
Why the f*cking hell do we need file format specs if they are ignored
and we just write into an implementers note what we want to see in the
spec. And at least if there is an implementers note file make sure to
mention all the cases and save me the reverse engineering time.

Change-Id: I08c2c9963fc6f2780e49a89f988a758665e007fa