[harfbuzz] Fix text width calculation, 3rd try
It turns out storing the width in the layout is not so good idea,
because in some mysterious cases when font fallback is involved we call
GetTextWidth() without calling LayoutText() first, and we return the
width of the previous text run.

It seems all I needed is to pass down the X offset with the glyph item,
and take it into account when calculating the width.

Change-Id: Idbdb6bba00573fb6ca773701757d667e21ac0912