Personas: Kill the non-finished "Background Image" feature.
It seems not to be a too good idea anyway - the image has to have some
specifics, so the value of setting any random image as the background of
Writer / Calc / Impress is questionable.

Either way, if anybody wants to finish that, easy to recover from the
history ;-) - better than keeping the unused code around.

Change-Id: I5079210bd91b5e70ce20681e5477b860bb28ff07