svtools FormattedField: don't reparse value from text after ReFormat
If our format is lossy / not a bijection
(as is e.g. our default date format, because of 2-digit year),
then reparsing the string that came from the format
leads to data loss (the value is changed to something else).

The existing code tried to do that, by calling Modify()
and then setting m_bValueDirty to false.
However, this fails because listeners are called
while m_bValueDirty is true. If any of them calls
e.g. GetValue(), the reparse happens.

Change-Id: I272f377927f83c71ede1eb80eafbc689f36fb17a