Keep track of current block position when pasting a range of cells.
This avoids re-starting the search for insertion position in the cell
text attribute array, on every single cell insertion. Instead of always
searching from the first position, it starts the seaarch from the previous
insertion position.

This reduces the paste time from the previous 3.5 seconds to 1.7 seconds
on my machine.

Here is the scenario.

1) Put =A1 in B1, and leave B2 empty.
2) Select B1:B2 and copy it via Ctrl-C.
3) Select B3:B50000, and paste via Ctrl-V.

Because of the empty cells interspersing the destination range, this makes
the data array highly partitioned, which makes the position lookup very
very expensive without the position hint.

Still, I was hoping the duration would become a fraction of a second. 1.7
seconds still seems a bit too slow...

Change-Id: I7742ce7e22935b6d0e082e4569d53dbd2072c4e5