Use cp -R instead of cp -r for OS X compatibility
In GNU cp, -r and -R are equivalent.

For the cp in OS X: "Historic versions of the cp utility had a -r
option. This implementation supports that option; however, its use is
strongly discouraged, as it does not correctly copy special files,
symbolic links, or fifo's."

Using cp -r meant that the symlinks in the LibreOfficePython.framework
were not properly re-created in instdir.

Change-Id: I8367269a77b876c063fd21ceb919936215fb7d37