fdo#60967: re-enable deletion of paragraph following table
This was actually a documented feature, so removing it just because it
only happened to work sometimes by accident was a bad idea :)

(regression from f4b6c94c68b5f67b931cde8d0acda6ec8b288bb)

So implement this in a less accidental way, so that Ctrl+Shift+Del
at the last position inside a table will delete the following paragraph.

Do not implement the weird "only delete first sentence and merge rest of
paragraph into cell" behavior, because i don't think it makes sense;
much better to just delete the whole paragraph, that is also the most
likely use-case for this (with table in frame/header/footer).

Change-Id: Ia88b3c8958798fd9c64ee75b56c61d787079133e