bnc#817956 fix VML import of rotation
In VML, positive rotation angles are clockwise, we have them as
counter-clockwise. This wasn't noticed earlier, as the n751117.docx
testcase also had flip:x. (For example, rotation with angle 90 + flip:x
is presented as 270 by the UI.)

Fix this, and also mirror the angle when flip:x is present.

Change-Id: I591ec3369a5bdca53f9684006a459d11e37fbc33