i#30979: sw: fix crash on Flys anchored in ToX
When a fly is anchored at the first paragraph of a ToX, as happens in
the WW8 bugdoc in i#30979, updating the ToX will remove the anchor
node, and bad things happen, including ~SwIndexReg assert and also
eventually a crash when editing the Frame and updating the ToX a few
times.  Probably best to really remove all anchored flys here.

Note that it's not enough to re-anchor to the empty node that is
inserted here because that one will also be removed again.

While we're at it the _DelBookmarks is superfluous because CorrAbs
does exactly that 2 lines above.

Change-Id: Id40ae1d88ad99813b93bea7470f5dc790edcc80c