Add sw/source/core/text/guess.cxx
A reminder: Listing some source files in this Xcode project does *not*
mean that Xcode would have any deeper knowledge about them. So Xcode
(and other IDE) fans, don't get overly excited, this is *not* a start
of making an Xcode project for all of LibreOffice or anything like
that. I just add arbitrary source files to this project when I happen
to want to set a breakpoint, and can't conveniently do it by going up
the call stack until something in that particular file.

Change-Id: Iac77dff17f56ec3f2826fd887cc065160b65936e