sc: XLSX export: remove mpAttributes crack:
When there are attributes, XclExpXmlStartElementRecord::SaveXml() would
not call startElement() but instead write the XML element all by itself,
which means there will be an endElement() for which no previous
startElement() was called, which makes checking that these match hard.

Since this silly mpAttributes over-engineering is only used in one place
anyway, give that a separate subclass and use startElement().

There is still an issue with various WriteAttribute methods; if those
are used it's probably not possible to check that for well-formedness.

Change-Id: Id842f000a6fecd77211836705545d58031ebd0f8