fdo#64626 clarify use of border line settings on toolbar
Using the toolbar button to set border lines in calc, lines are added,
but not removed. I thought this to be a bug, but it isn't. Using the
toolbar button is for addition, and using the same button with the SHIFT
key is for addition+deletion.

Unfortunately, this is not mentioned in help files or hints.
This patch remedies this by giving a better hint when using the toolbar
button.

Change-Id: I0cf5cc9bffbec363cde3c7493f024cdfa72c54f4
Reviewed-on: https://gerrit.libreoffice.org/4220
Reviewed-by: Miklos Vajna <vmiklos@suse.cz>
Tested-by: Miklos Vajna <vmiklos@suse.cz>