fdo#62536: sw: fix AutoCorrect bold/underline on existing AUTOFMT
With the native AUTOFMT in Writer the SETATTR_DONTEXPAND does no longer
work reliably: if there is an existing AUTOFMT at the position then it
will be modified and no new hint with DontExpand will be inserted.

Work around this deficiency by inserting a no-length hint with the
preivous formatting at the end of the range.

(similar fix to the i#75891 problem in SwTextShell::InsertSymbol)

(commit 062eaeffe7cb986255063bb9b0a5f3fb3fc8e34c did not
 introduce the problem but made it far more annoying)

Change-Id: I58ece7f5bd5a786b22a066e5902f1784dafa5dce