Turn redundant member direct-initializations into value-initializations
...the direct-initializations were introduced with
032d5126e41562917b91b8f4826b0be628169968 "cppcheck: fix some uninitMemberVar in
configmgr part," presumably to avoid false positives from static analysis tools.
But the initializations are redundant, as the invariants of the affected classes
imply that the members have meaningful values at their points of use.  And the
direct-initializations with arbitrary values make it harder for a maintainer to
understand the relevant invariants, namely to ensure the members must have been
assigned /meaningful/ values by the time they are used.  Give a (subtle) clue by
making those into value-initializations.

Change-Id: Iadb25fa08b6d6b12d5bf8a8f04271270b6a7f7cb