Page border bug: margin edits are enable when no broder
To reproduce
1. Format -> Page
2. Set left and ritght borders only (All margins is editable)
4. Set inner margins bigger (using syncronize check box), for example 2 cm
5. Click OK

Result
Margin appaer just on the right and the left side, so just on those
sides where there is a border.

Expected:
Only the relevant edits must be enable on border tab page
just as in case of paragraph borders.

Note to fix:
Inner margin without border is used only for table, but not
for other objects (paragraph, page, frame)

Change-Id: Ie611eb29432cb7e75131154d94b4a8db1c3f1bb8