Avoid having formula cell directly update text attributes.
This would cause column cell and text attribute arrays to go out of
sync, because the formula cells may be stored and interpreted in places
such as change track and conditional formatting.

We still need to find a good way to mark text attributes "obsolete" when
formula cells are re-calculated.

Change-Id: Ida612806d3afec23c5ae501f2fc8cc7d52e019a8