Prevent crash during on-line spell checking.
The new ScHorizontalCellIterator internally uses iterators for each
column to track positions. This means that, if a cell value in the iteration
range chnages while the iteration is on-going, those internal iterators
get invalidated. Allow the client code to rehash the iterators when modifying
a cell content during iteration.

Having said that, it's best not to modify cells during iteration.

Change-Id: Ida453d4f883e1fbcbab4eb0401e37fea8c0b901d