tweak ppt ole import for libreoffice codebase
a) make sure backcolor is skipped when libreoffice backcolor
is default. Libreoffice default is 0 ( but if we write that out mso
reads it as black )
b) don't take the label default backcolor in exchange for the
libreoffice default. Libreoffice default backcolor for label is white,
mso's default is rather different ( and never seems to be used )
c) re-enable compObj export ( seemed this was disabled unintentionally
except for buttons )

Change-Id: Ie68267809cee2e92bd06e88e698f17a10eeeb6ad