fdo#63254: correct re-mirroring formula
This patch fixes horizontal scrollbar in Calc when a sheet is RTL and UI is LTR.

When a particular window (horizontal scrollbar in this case) is RTL and a whole
UI is LTR this code used for mirroring and re-mirroring back coordinates. AFAICS
mirroring is correct, but re-mirroring is not.

The used formula comes from commit 69b684b12e42bf1396bdffca88ca880ac3c2e00b, and
the same formula appears in the other variant of overloaded function "mirror":
 mirror(sal_uInt32, const SalPoint*, SalPoint*, const OutputDevice*, bool),
but it does not affect the bug.

Change-Id: Iec9c53edce9d33be6b676d1094bed8bf92f2e435