fix for fdo#66718, access correct clipboard
Changes to use SystemClipboard::createUnix break some subtle behaviour
( or is it even misbehaviour ) that users of the service
"com.sun.star.datatransfer.clipboard.SystemClipboard" depend on.
When creating the SystemClipboard the arguments passed to
X11SalInstance::CreateClipboard(...) are used to access cached
instances. Previously the window clipboard was created with 'no'
params and then initialised ( via XInitialisation ) later
This ensured users creating the "css.datatransfer.clipboard.SystemClipboard"
with no paramaters immediately get that same clipboard. Note: creating
the SystemClipboard with no param seesm a sensible thing from an scripters
point of view as the params that would be required to match the correct
Clipboard seem rather difficult to access in an uno only world

Change-Id: Ib71d6744086d354025de4db62039b9a98f230eca