fdo#66762 - Other: com.sun.star.awt.DisplayAccess removed
API_CHANGE:

I removed the undocumented internal API
  "com.sun.star.awt.DisplayAccess"
in commit dde234b6955a421d51d2b37e4fc3972c660146f0
  "fdo#46808, remove awt::DisplayAccess service."

However, it appears someone was actually using this.

Now, at the same time, we have a method in XToolkit, getWorkArea(),
which looks like it was designed to return this exact information,
but which has been returning zero for as far back as our commit
history goes.

So, to kill two birds with one stone, this commit changes
getWorkArea() to return the information the customer needs, to wit,
the size and position of the primary display.

Change-Id: I76300bac604e9e4b7be95b0872a7f95cb6781903
Reviewed-on: https://gerrit.libreoffice.org/4919
Reviewed-by: Michael Meeks <michael.meeks@suse.com>
Tested-by: Michael Meeks <michael.meeks@suse.com>