convert move table dialog to .ui format
Help needed:
I did not find a way to get the height of the warning area
included in the automatic calculated dialog height. So it
stays behind the action buttons if an error occurs.
(I am new to glade)

Code adaption should be complete - build and move-sheet-test ok.
helpcontent2 is done and I will try to upload.

Change-Id: Id987ac799b17787c35302890a95af2bd99456589
Reviewed-on: https://gerrit.libreoffice.org/5006
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>