Fix remote control server side build problems
The non-Bluetooth (TCP) code should not be inside Bluetooth conditionals.

You can't put a flag like -Wno-error in a list of objects. (And besides, why
would we want to disable -Werror like that, instead we should not introduce
warnings in what is after all freshly written code;)

This should fix the Android tinderbox build.

Change-Id: Ia698031a9815bf938c5c2a29bf68dee68cae07fe