Disable case insensitivity in findColumn (firebird-sdbc).
In theory it is possible to have case insensitive identifiers
in firebird however these are only used with unquoted identifiers.
Base by default always uses quoted identifiers, there doesn't seem
to be any way of determining whether we are dealing with a case
insensitive identifier when retrieving it from the database either.

Change-Id: I4df0ab10c5a7c15ecde099e36978ac2e6c4ad0be