Cut identifier to correct length (firebird-sdbc).
Currently firebird returns table and column identifiers as being
93 chars long (rather than 31 as per spec), this leads to problems
with ALTER and DROP commands which are unable to deal with the incorrect
length. For now we simply discard the extra characters (which are
blank characters).

(SELECT and similar commands have been able to deal with the overlength
identifiers previously, which is why this hack wasn't required
beforehand.)

Change-Id: Ib6c984b5fc72dc4e326b3f8d36f6cbc4a3c9086a