fix build with debug STL
Failed with:
/usr/include/c++/4.8.1/debug/safe_iterator.h:510:error: attempt to compare
iterators from different sequences.

Objects involved in the operation:
...

The problem is that miPos in _copied_ object points ot maAttrs in the
original object, not in the copy (and std::for_each takes a copy of the
functor). This could be solved by defining copy constructor and
operator=, but given the limited usage of the class, it is simpler to
let copies share the state.

Change-Id: Icf3f02ecd2fe4ce6dd77f3cde226d32beb4d4b3f