Adapt UnoControlModel::read/write
...to 8ee69b0ba13f74d1515fac71df92947eb6328ab1 "fdo#67235 adapt form control
code to time nanosecond API change, step 3."  It is a bit unclear to me how
exactly this code is used, so to be safe, just read and write in the old format
(of using a single integer to represent a Date resp. Time) at least for now,
loosing nanosecond precision and the UTC flag.

Change-Id: Ib5148f750a420ad09366c79b68370ad0efd501f4