Add a new option to bin/find-german-comments
The new option "-L" allows to print out potential candidates for translation
combined with the number of lines found like "-l" - but only if there was a
positive result (i.e., a file that needs checking).
This leads to a much more compact representation helping people to find the
few files that still need translations.

Change-Id: Iad554eb6ce5b062e96774d9488c54105679b7c4b