WaE: array subscript is above array bounds
Avoid a (presumably bogus) warning produced by gcc 4.7 and 4.8, at least those
in the Android NDK, in an optimising compilation. As such the code probably
was OK, even if a bit ugly and suboptimal, as it doesn't seem to be sure
whether various buffer lengths should be hardcoded or can be variable, etc.

Change-Id: I29790cb7a9cb65735e4ebb9a1a198f8575282ecf