Split type-description out of createStandardColumnPart.
createStandardColumnPart returns a string that is usable in table creation,
the type descriptor (e.g. CHAR(N)) is also useful in column-type altering
statements, hence this is split into a separate function.

This is needed primarily in the firebird driver where column-type alter
statements are in a different form to most dbs.

Change-Id: If44554f979f6dd14bf679455978f3f0b44784bf5
Reviewed-on: https://gerrit.libreoffice.org/5432
Reviewed-by: Fridrich Strba <fridrich@documentfoundation.org>
Tested-by: Fridrich Strba <fridrich@documentfoundation.org>