Work on the custom shape preset generation
Don't generate code that uses the old RTL_CONSTASCII_USTRINGPARAM horror.

Update the README file to have actually working command lines, and at the same
time make it into a shell script that can actually be run.

Use SAL_INFO with tag oox.cscode to output the stuff that the
generatePresetsCXX.pl is looking for.

Change-Id: I3f8f756fbe8e35857c2bd07beece0f3ab4234e7f