fdo#37230 Add legacy cell selection behavior and config. option
When selecting cells with ctrl+shift, the legacy behaviour was to
take in account to which cell the cursor moved. This allowed the
user to make some convenient selections. This behaviour can now be
turned on with a switch in the configuration (Tools > Options >
Calc > General > Input Settings > "Use legacy cursor movement
behavior when selecting").

Change-Id: I00ae7b4ee001b416cc680c7a650631daec71e51b