Revert "Reduce text background painting rect with border"
This reverts commit cc7a94cfa8ec901627e8f9abf28c724923deb520.
MS Word has the concept not to paint background for
the border line.
But LibreOffice follow the general html box model which
paint background for border too. So in case of double line
the border's gap has the same color as the background.
So the reverted patch can't fit to the general concept.