GSOC work, arrow navigation+TextSelection problems fixed
I've added a new function called EditorWindow::GetLastHighlightPortionTextSelection, which gets the last edited word (from the highlight portion), and creates a TextSelection from it. Later, this is used to remove/replace text in the listbox when pressing the tab key. The proble was, that is cleared the whole line, but now, it just clears the newly edited word.

Change-Id: I61b6721696e89002705c9980579023b42ad1faaa