treelist for autofilter ( nested nodes for dates ) ( using SvxTreeListBox )
Attempt at trying to get more interopable behavior, imho the way dates are
presented in the autofilter of the "other" well known spreadsheet application
is very nice and intuitive. This attempt here is lacking in that

a) for some reason I am not getting the node icons ( and subsequently can't sing
le click anything to expand the nodes ) Only double clicking works to expand
b) the 'check' behaviour is not quite right, e.g. as simple example is if all
nodes of a particular date are unset then clicking the leaf node ( e.g. day )
wont set the ancestor nodes.

Change-Id: Iaa89cac21fb18074ff64a984f099a35851c2fe3c