More work on a sandboxed LibreOffice on OS X
In particular, surround also the ftruncate() operation that
osl_setFileSize() does with access through a security scope bookmark
for the file, if available. This fixes file saving in a sandboxed
LibreOffice. (But oh boy, does simply saving an ODT document go though
a weird dance of file operations.)

Luckily the C++ oslFileHandle abstraction keeps the pathname that the
file was opened with, so even if ftruncate() as such takes only the
file descriptor, we can get at the pathname to retrieve our security
scope bookmark.

Change-Id: I8acb1b2f3fb3ec0cea833697b7f1d4a1912ed551