fdo#64491: Added functionality to export SmartArt diagrams
When exporting a Frame that contains an eDrawing, check whether it's
a diagram and export it properly if so.

Exporting the diagram means adding the relations for the four
mandatory files (data, layout, quickStyle and colors), and the
optional one (drawing), and also writing those from the imported
XDocuments to the respective files.

If the drawing file exists, the data XDocument must be updated before
writing it to disk, as it contains a reference to the drawing file
that must be updated with the new relation id.

With this change, the files saved to docx that contain SmartArt will
be properly opened by MS Word without data loss.
LibreOffice won't be able to properly open saved files though, as
it requires the theme file to be present, and it's not being
exported yet. This is planned to be fixed in an upcoming patch.

Change-Id: If69ca9815dd076597a37937fe2af8ed4d1c7acc5
Reviewed-on: https://gerrit.libreoffice.org/5896
Reviewed-by: Thorsten Behrens <thb@documentfoundation.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Miklos Vajna <vmiklos@collabora.co.uk>