fdo#32059: Fix this once and for all
Fix a regression from 356b2f0edfd5a01ed05976c1bdb74948ff84b5fd.

This regression was about the setSelection function. We inserted the command with a space
in the right side of the string, but our selection was not counting the " " character.

So, reasign the variable with the space solve this issue.

Change-Id: Ic48f1afaec89c6eefe73d051d3b0eba7edbf557e
Reviewed-on: https://gerrit.libreoffice.org/5688
Reviewed-by: Ricardo Montania <ricardo@linuxafundo.com.br>
Reviewed-by: Marcos Souza <marcos.souza.org@gmail.com>
Reviewed-by: Norbert Thiebaud <nthiebaud@gmail.com>
Tested-by: Norbert Thiebaud <nthiebaud@gmail.com>