Resolves: fdo#68726 make navigator child dialog float on top
i.e. set its parent to the thing it wants to be on top of

As an aside I'm not entirely sure that this little dialog has the decorations
set on it that it would like, I suspect it wants to be more like the
tearable-off font color floatingwindow. But lets not muddy the waters in this
commit.

Change-Id: I9e0c35fbb66e50c04c7b28904d3937cf9ae417f4