Check for FT_Face_GetCharVariantIndex at build time
No more dlsym() hacks. The downside is that LibreOffice built on systems
with old FreeType will not be able to use FT_Face_GetCharVariantIndex()
on newer systems, but most Linux users use their distribution builds
anyway.

This only affects the use of Unicode Variation Selectors which is an
exotic feature that wasn't even supported on Linux before 4.1.

Change-Id: I674822ef5bc8d7940a821a01cc85ae7a6d39a80e
Reviewed-on: https://gerrit.libreoffice.org/5844
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>