Add simple embedded firebird .odb loading test.
This opens an "empty" firebird-based .odb and tests that it is possible
for the firebird-sdbc driver to open the embedded database.

"empty" denotes that the .odb is marked as using embedded firebird
but doesn't in fact contain any .fdb file within. This is usual state
of a .odb directly after creation using the "New Database" dialog when
the sdbc driver first opens the database.

Change-Id: I83941c05b6328d8419dca49121988640c6f887bc