WaE: variables 'nDoub' and 'nFloat' are uninitialized when used here
Initialise them in the presumably intended way to get rid of the
warning. Note that the code around there is somewhat confused and
downright wrong anyway, though.

Also "+0", "-0" or "000" for instance are correct ways to write zero.
Also SAL_MAX_INT32 and SAL_MIN_INT32 are legal values for a UNO
"long". (And it will always be true that a sal_Int32 is <=
SAL_MAX_INT32 and >= SAL_MIN_INT32.) And does the code accept any
single-character string as zero?

Then there is copy-paste galore for the "hyper", "double" and "float"
cases.

Change-Id: Ie1920cfb4d17dca7cd966d8b200d2510ea975290