Resolves: fdo#69444 cursor keys don't move between recent used docs
because each item is identified by an id, but that isn't set on the recent used
so it stays stuck on the first one.

change the ctor of the base element to require an id to be passed to it and
adjust all the good code and then use the index+1 of the recently used elements
as the id.

Change-Id: I1e5978ddedbd4718923ef72d67598158d5ba8ee7