Add support for sliding average
Added a new subclass DynamicKernelSlidingArugment to handle sliding arguments.
Currently only sliding average is supported, and the sliding average is computed
on the CPU for now.

Refactored other DynamicKernelArgument subclasses so now the Marshal() takes
an extra argument nVectorWidth which is needed for kernel arguments that
require allocating intermediate result arrays.

Change-Id: I1fd95efd4af148ae9b270f1e45d2d2ffe74b005e