Kill PropertyDefinition; use PropValue instead.
PropertyDefinition was introduced long ago, probably in order not to have to
touch larger amount of code when it turned out that it was necessary to store
more information in PropertyMap - otherwise even then it was more logical to
extend the value from uno::Any only to some structure.

PropertyMap's API itself is quite non-convenient, it is half-way between
direct access to the std::map it is derived from, and something built on top
of that.

In the long run I believe we want to make _PropertyMap member of PropertyMap,
not an ancestor, and change the API; if anybody cares enough, that is ;-)

Change-Id: Ib8e9bfc4561f003b78ab6ce6620722e251f661d6