We want config_host, not config_build
We use GNU cross-compilation terminology: "host" means the platform
the code being compiled will run on (this is often somewhat surprising
to those unaware), and "build" means the platform the build tools (=
the compiler, linker etc) is running on.

For an iOS app, "host" is iOS and "build" is OS X.

There is also "target" but it is in this case the same as
"host". (Only in a so-called Canadian Cross situation is "target"
different from "host" (and from "build").)

Change-Id: I6eefa982916f0660afd259770bf455cf28793670