mailmerge: fix salutation list
Previous state: there were two salutation lists, one for men,
one for women. But they contained not only the salutation, but
title, placeholder, and a punctuation merk. Therefore, when the user
composed a Custom Salutation, and selected a pre-defined salutation,
the result was something like for example "Dear Mr. 2, Mr. Lastname,"
instead of "Dear Mr. Lastname,".
Current state: salutation list contains only salutations. There is no
need for two lists, because the Custom Salutation is customizable
separately for men and women and it is not necessary to choose
from the list.

Change-Id: Ibb4e195401f1d01d5e63c9085b99d7acf3ef1229