C++11: new/delete replacement functions cannot be inline
...as clarified by a new sentence added to paragraph 3 of 17.6.4.6
[replacement.functions]: "The program’s definitions shall not be specified as
inline."  Clang trunk towards 3.4 now generates errors for this.

Having these replacement functions in the fbembed dynamic library is extremely
fishy anyway; at least on Linux those symbols are not exported, and hopefully on
no other platforms either.  (If it turns out that this change causes the symbols
to change from non-exported to exported on some other platform, the best fix
would probably be to remove those replacement functions completely.)

Change-Id: I590d55e44814a6707030c42e1087377e75819666