fdo#70144 Mark database document as modified when data inserted in firebird.
This only applies to embedded databases where any changes to the database
only affect the temporary extracted db which still has to be stored into
the .odb.

This only fixes the status for modification/insertion/deletion of data
in existing tables. Insertion of new tables still does not result
in the modified state being set.

Change-Id: I553f67127672f2cf49b98ddcdf7d25f22bef5074
Reviewed-on: https://gerrit.libreoffice.org/6621
Reviewed-by: Lionel Elie Mamane <lionel@mamane.lu>
Tested-by: Lionel Elie Mamane <lionel@mamane.lu>