made horizontal cell alignment depend on writing direction, fdo#63546
If not left or right aligned or centered, the horizontal cell alignment
depends on context, whether cell starts with an RTL character, is a
numeric value, writing direction set at cell or sheet writing direction.
Do this for drawing the cell text so switching writing direction for a
cell or on sheet level actually works.

Change-Id: Ic03f432acbff35008305c90787a649fcae7320f8