CppunitTest_sw_ooxmlexport: speed this up a bit
We don't have too much testcases where we assert directly the resulting
zipped XML files, but where we did, we actully invoked filter() 5 times,
even if 3 times is enough, fix this.

That causes the testcase to run in 1m24.754s (previously it was
1m30.278s).  Additionally, at one case the XML result is an expected
one, having matching testcase and manual results is always good.

Change-Id: Ie905002b113bc84f1e517c35f14739f423fd8ea2