DOCX drawingML shape filter: import wps:txbx inside the shape
The problem was that previously the shape text was imported normally, and the
shape itself was simply appended after the text.

In case of the VML import, the following mechanism made it possible to have the
shape earlier: OOXMLFastContextHandlerShape::sendShape() is called twice, early
in OOXMLFastContextHandlerWrapper::lcl_createFastChildContext(), in case the
shape had text, and later in
OOXMLFastContextHandlerShape::lcl_endFastElement(), in case the shape didn't
have text. This works because the parent element of v:textbox (v:shape) is
already something that isn't handled in writerfilter, so the parent of the
w:txbxContent's parent was a OOXMLFastContextHandlerWrapper.

In case of WPS, the problem was that the parent of w:txbxContent is wps:txbx,
and the parent of that is wps:wsp, which is something (unlike v:shape) we do
handle in writerfilter. Fix this by adding an early call to sendShape() in
OOXMLFastContextHandlerShape::lcl_createFastChildContext() as well.

Change-Id: Ia24678871d7bbad89d18b1d5f468c17f68feec10