export data streams; pretend to be an area link
Fix ScXMLExport::GetAreaLinks to not use UNO, so we can hack DataStreams
into ScMyAreaLinksContainer.
We need to connect them to cells for export / import - that does not make
sense for DataStream without any attached range. We use an arbitrary one.
But you might remove it, if you create another ScAreaLink / DataStream
with the same range, without knowing about it.

Maybe it wouldn't be that bad to let DataStream inherit from ScAreaLink.

Change-Id: I5b85a9329ba1ca46fb2893b54fe5161d2fb22f47