fdo#64232: Add font theme info to CharInteropGrabBag
The attributes w:*Theme in the rFonts tag are saved to the character
InteropGrabBag so they can be saved back to the document on export.

Font attributes are also pushed into the grab bag because they will
be needed to check if the user has modified the fonts while editing
the document.

Change-Id: Ib247845a2d5543823798e1bbd8ed1958a5d9e884