Fix Cygwin PATH composition
Adding Windows-style paths (C:\foo\bar) happened to appear to work, as Cygwin
translated that into two paths C;\foo\bar (where the former typically just does
not exist and the latter is relative to the current drive, which is typically
C).

In theory, wrapping in $(shell cygpath -u ...) should not be necessary, but in
practice S is set to a Windows-syte path (which is probably a bug?) after the
paths of PATH would be shortened to use $I/$S/$W, which is prevented with the
wrapper.

In theory, the shell variable assignment PATH="$$PATH:... would more
idiomatically read PATH=$$PATH:"... but e.g.
unoidl/CustomTarget_unoidl-write_test.mk passes this variable assignment as a
command line argument to exectest.pl in which case it would potentially be split
into multiple arguments.

Change-Id: If870e9eba8b650fe75b324ac54c49891d19fcc55