DOCX import: parse drawingML group shapes in oox only
This is similar to commit d5c934d150cb6cea5f96cbbee4fb5e8312bf027e
(n#792778 DOCX import: parse group shapes in oox only, 2012-12-14),
except that was for the VML importer.

The only difference is that in case of drawingML,
OOXMLFastContextHandlerShape::lcl_createFastChildContext() is only
called for the children of the group shape, not for the element itself,
so compare against the start token, not the current element.

Change-Id: Iddeabb20bbd5f0153e2fc4e6df463830126fdd37