oox: render progress bar in main thread for threaded import.
Experimental only.

This avoids ,us deadlocking as the main thread tries to join it's children,
and the child threads wait to 'Yield' in the progress bar update.
Also it's generally safer to move progress reporting out of the other
threads, and to have the mainloop spinning here.
Finally  this allows people to continue to use the LibreOffice suite while
large XLSX spreadsheets are loading.

Change-Id: Id41c18f3941d6fc5eea593f7cfcf6a8b7215b3f8