Try to fix colour issues
Use RGBA consistenly. Wonder why the code was changed to use BGRA at
some point?

I got the picture in the document to show up with correct colours but
unfortunately not the RED GREEN BLUE etc text. Weird. Even weirder, if
I add a temporary hack in CoreTextStyle::SetTextColor() to use some
other colours for non-black text (instead of the ones passed in the
parameter), those colours do show up. This is a mystery.

Change-Id: I591424a19fa02b3f095035e989cbc49fff94b8ca