sdbc file driver (Prepared)Statement: created ResultSet owned by *caller*
That is the only choice that makes sense, since the
(Prepared)Statement could die (go out of scope) before the ResultSet.
User code could do that, if it does not "need" the (Prepared)Statement
anymore. Also, it is only natural for user code to dispose a ResultSet
that it does not need anymore.

So we need to create a fresh ResultSet each time.

The "luck" here is that the sdbc file driver does not implement
the XMultipleResults interface; things get more hairy then.

Change-Id: Ibf2cb5e5b7ca90432a289c185a6b4fe32d1ba565