fix deadlock on pressing cancel in find files for gallery
progress thread blocked on solarmutex
press cancel in main thread which has solarmutex locked
call terminate
call join, which waits for thread to finish, which it can't
because its waiting on the locked solarmutex

so...
just call terminate on cancel. Let the thread get the solarmutex
and do its last round of work, at which point it will exit and post
an event to the main thread to call Cleanup, put the join there
instead and which point we know the progress thread is finished and
has no more interest in the solarmutex.

Change-Id: Ibe3976377288ac76b1b37c875a32e20b8d5daae1