reverted the CSV comma-checkbox override
While for .tsv|.tab files forcing the Tab box to be checked may be
helpful, doing the same with the Comma box for .csv files is not. Though
named Comma-Separated Values, the file format is often used with other
separators, especially with localized content where the decimal
separator is also comma to not interfer and having to enquote each
numeric field.

Users frequently importing such files would have to uncheck Comma on
each import, while if it really was a comma-separated file and such
files are imported frequently checking Comma would have to be done only
once, as the options are persistently remembered.

Change-Id: I464d453bfa4e2a793faabbabfccb5986c454496e