fdo#72204 RTF import: fix crash on rotated, but not imported groupshape
In case the groupshape contains textboxes and they contain table, we
don't import the groupshape itself, just the children, as drawinglayer
rectangles wouldn't handle tables.

If that's the case, don't try to apply properties on the groupshape, as
that would crash (the Writer drawpage implements XShapes, but not the
XShape interface).

Change-Id: Ie34039170314b772dd050eb354681cfaec61c1fa