Fix stepping to the next change
Bug description:
In the "Accept or Reject Changes" dialog, when select a change
(it should be after the first few ones) and clicking "Accept"
or "Reject" button, acception/rejection is made but after it
not the next change is selected for futher work,
but a random one. This bug doesn't appear in all case,
just if changes are complex enough.

Solution:
The nPos means absolute position so we have to get the next entry
with the corresponding GetEntryAtAbsPos() method. It seems simple
position can differ from absolute positions if changes are
complex enough.

Change-Id: I7996f81c2a09c492f9334f071591291d200d533f