Need to have native-code.cxx as a separate file for the LibreOffice app
The executable of the LibreOffice app (which as such at the moment
doesn't work, since the tiled rendering changes) is built using
gbuild, and thus we can't generate the native-code snippet in the
CustomTarget that builds the app bundle, but need it already when
building the executable. This is one wayt to handle that.

Change-Id: Ifdab40c970e93b1f2608cefc637df8a8e5396efe