fix help packaging for Mac OS X language packs
helpcontent was not packaged into the tarball that is extracted into the
user's installation directory (and thus never was installed)

simplify the packaging by getting rid of the special "collect" dir
(LangPackHostName) that was the cause for the problem to begin with:
LanguagePack files were collected in LangPackHostName based directory,
while HelpPack files were collected in regular directory.
And that value wasn't even used for the actual path that is delivered to
the user, so get rid of it and the associated code.

Change-Id: Iadd24382a45722bbe8d58bd97be79609fd03021b