Drop unused nMergeType from SwNewDBMgr
All code uses the nMergeType from the SwMergeDescriptor. Actually
there is just code setting the nMergeType in the SwNewDBMgr but
none reading or comparing it.

I don't know if the removed assignments actually should work on a
SwMergeDescriptor object, so some parts of mail merge may already
have been broken.

Change-Id: Ic8c52e6171f122d3837d9f2de81053e30ac8f5b7
Reviewed-on: https://gerrit.libreoffice.org/7432
Reviewed-by: Michael Stahl <mstahl@redhat.com>
Tested-by: Michael Stahl <mstahl@redhat.com>