filter: audit customshape type -> drawingML preset shape table
It turns out that this table contained entries which are not valid
drawingML preset shape names.

Invalid entries are now either have a valid name instead (based on the
oox/source/export/preset-definitions-to-shape-types.pl mapping, which
already tries to map VML shape ID's to drawingML preset strings), or in
case I found no mapping there, the entry is now simply commented out.

It's still better to fall back to export a shape as a rectangle than
corrupting the whole document.

Change-Id: Ic2d8926e5819f3312aaca55d50a1492332e52a9d