VML import: fix height of textboxes when border distance is non-zero
The VML concept is that the height / width of a textbox is absolute, and
border distances only affect the position of the shape text, not the
size of it. OTOH, when we set the Text*Distance UNO properties on a
textbox, the size may change. Make sure that during VML import setting
those properties doesn't change the size.

Change-Id: I53b328b66572fc05027be344869bc1a78d855558