fdo#72639: send proper ACTIVE_DESCENDANT_CHANGED events
At ThumbnailView::SelectItem the piece of code in charge of the
accessible events related to selection change was mistakenly using
the accessibility helper class ThumbnailViewAcc instead of
ThumbnailViewItemAcc which is the proper one for managing the events
related to ThumbnailViewItem objects.

The methods FireAccessibleEvent and static getImplementation had to
be added to ThumbnailViewItemAcc to provide the full functionality
required.

Change-Id: Ia5ffb24c920f3c48854021700819be683c979411
Reviewed-on: https://gerrit.libreoffice.org/7551
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>