Resolves: fdo#73814 wrong name for drawing::PointSequenceSequence property
examining svx/source/unodraw/unoprov.cxx "Geometry" is a
indeed a drawing::PolyPolygonBezierCoords for the BezierShapes
and a drawing::PointSequenceSequence for the PolyShapes

regression since e44335abe006d05f0f915279605a03ef084116d6 because after
223f6b631c1b087754c0f9051fb55f029f2503ce we started getting
drawing::PointSequenceSequences in maPath which is the wrong type for the
argument to property "PolyPolygonBezier" for those shapes.  Which led me to
incorrectly assume that the all PolyPolygonBezier properties were named
"PolyPolygonBezier" which isn't the case.

so reverting the non maPath hunks of e44335abe006d05f0f915279605a03ef084116d6

Change-Id: I013a66778d11a472fc4567e53a9e17e73e2b91ce