filter: add sprmids header
This was generated previously in writerfilter, but keeping around an XML
file of 8000+ lines just to generate these constants is an overkill.

Additionally, these constants are hardwired in all available WW8 files,
so they won't change. By copying it here, it's possible to incrementally
remove what's still remaining from writerfilter's not used doctok.

Change-Id: I1d30da1f635946b15a37de68080fd9eebc2ad88e