Add support for 'Table Cell Redlines' in SW core
This patch adds support for 'Table Cell Redlines' (such as 'table cell inserted'
or 'table cell deleted' in SW core).
It adds the 'SwTableCellRedline' object, and adds a function for adding
objects of that type to the 'SwExtraRedlineTbl', which is the object
that holds all the redlines which are not 'Ranged' redlines.

Change-Id: Ic2e410be58683f171ea07d430b7544600780711e
Reviewed-on: https://gerrit.libreoffice.org/7873
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Miklos Vajna <vmiklos@collabora.co.uk>