fdo#74782 search view doesn't hide local view when searching first time
See bug report for screenshot of current behavior.
When the search button is hit, OnTemplateSearch() is called.
If we switch from 'no search' to 'search' view, mpSearchView was not hidden.
This'll result in SearchUpdateHdl that mpCurView will not hide.

It doesn't matter if we hide it anyway (either when we would start searching,
or end searching). If we start searching, we don't have any search term yet
(so no result anyway). If we stop searching, it has to be hidden for sure.

Change-Id: I6bcf540b4866eeb6a8196c171027b385635cf084
Reviewed-on: https://gerrit.libreoffice.org/7969
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>