drawingML export\import: cropping of shape's fill texture
In Writer shapes had no cropping property so far. With this
commit this is introduced as a FillProperty and has the same
type as the cropping used for pictures
(Picture context menu > Picture > Crop).

Layout and UI will be an other step. On the UI it would be placed
on the Shape context menu -> Area, when Bitmap is selected as fill type.

Note: In case of picture/graphic, cropping property is imported from
and exported to a:srcRect instead of a:fillRect.

Change-Id: Idc1ed2d40cb20b6992e94f14e7e4d853e1f55d02