inserting a picture in a blank placeholder is setting its bg
which isn't probably what we want here, we want to replace it fully like we
always did. The old behaviour that if the frame is non-empty then we fill its
bg with the color is still the case after this change.

regression since 5c75cd95178e3d57e53fee64a9d64023c6d18acf

Change-Id: Ib670a7f0dae5809d6936910b6ba4886a99fb7e64