fdo#72390: Let's not skip auto styles from unmodified sheets.
We do re-use sheet XML stream for unmodified sheets on re-save, which
is okay.  But skipping auto style population from unmodified sheets
is not okay because autostyles are shared across all sheets and written
outside the sheet XML streams. Skipping styles from unmodified sheets
could potentially fail to export used styles, not to mention could
cause a crash as reported in the bug.

Change-Id: I73e439f7354e341c2c07e28ecca5452193985860