fdo#72691: Write unit test for this.
One tricky bit was that in order to reproduce the bug in the unit test,
formula cell's bChanged flag needed to be cleared after each update,
because that's what happens when you display updated formula on screen.  Each
time an updated formula cell gets drawn, the UI code clears this flag.  That
same flag was used to control the code path during interpretation...

Change-Id: I2eedea3c9294f4f545422b8611840c81f8c2304f