fdo#72691: Allow overwriting of string value with numeric one.
This can legitimately happen when you have a matrix with a reference to
another cell inside, and the referenced cell originally contained a
string value then later overwritten by a numeric value.

Example.  Put a "Text" in A1, and in B1 put a 1x1 matrix {=A1}.  It
displays "Text" in B1.  Then put 11 in A1.  Prior to this change, B1
would become blank.  With this change, B1 will display 11.

Change-Id: I3feba3a8658e1a5ebf6f9e5ac34de2d579464ddb