external: Use gb_LTOFLAGS only in LDFLAGS to fix building.
liborcus was not building for me with -flto in CFLAGS, I would have to
fix ar somehow.
-flto in LDFLAGS is just to fix the build if the external library does use
another library built by us with -flto: does happen for liborcus and python3.
It's not like we would use -flto for external libraries consistently anyway,
the only exception is icu: no idea why we build with -flto there.

Change-Id: Ia54d619674b8999ce5e4b920ba77b1587c9cf48d