my upc router automatically adds .net to the domain search path
i.e. I get handed an /etc/resolv.conf of

  domain net
  search net ...

and so "nohost" resolves to "nohost.net" which actually exists, so I get an
authentication request which is unexpected by this test which therefore falls
over. "nohost.invalid.net" also exists btw, as are other various foo.net forms
where foo is one of the reserved tld which don't themselves exist.

So lets shove a "*" into the name which is definitely invalid, sigh.

Change-Id: I8627ced9199f339ad6cca8f27752d883ebe74aad