KDE4: sleep in yield for native file picker
As it seems to be the only way to poll the clipboard, reintroduce

 m_pApplication->clipboard()->setProperty(
    "useEventLoopWhenWaiting", true );

To prevent crashes, disable event processing in the Qt thread while
the dialog is open.

Instead this applies the same workaround as the Windows backend to
sleep a ms, which keeps the FP dialogs more usable, but feels like
a horrible workaround.

This is still slower then running processEvent in Yield but still
much better then the current situation.

Change-Id: I10c422f1c0d7448d4a7ad28e57a32ed2cb42f48f