Use the AppleLanguages preference instead of AppleLocale on OS X and iOS
AppleLocale is for things like date formats, it seems, while the first
entry in AppleLanguages contains the user's preferred UI language. I
think that is closer to what we want here.

Note that in AppleLanguages, the country (where present) is separated
from the language with a hyphen, while in AppleLocale an underscore is
used.

Change-Id: I6d8a728d53ca6b46e8fcf0e333a6d0aa34abff1d