fdo#73115: Always run timeouts as events
Right-click popup menus run click events throught the LO main loop.
In case of KDE4 the LO main loop is run by a timer in the main thread,
with Qt::DirectConnection execution.

If the timeout actually starts a nested event loop for a new dialog,
the timer is blocked, the nested mainloop detects it was started
from the timeout and drops the blocked timout from polling, which
blocks any further LibreOffice event loop processing.

This changes the timers to Qt::QueuedConnection, so they always
generate an event and are processed by the Qt event loop.

Change-Id: Ie626b22be3d8f9b8934bcc5e9e0e67a365549cfc