Don't Valgrind into java children in unit tests
...similarly to desktop/scripts/soffice.sh preventing that for LO itself.

One problem is that, even if it does not find any errors, Valgrind writes onto
a traced child's stderr, and when jvmfwk searches for a JRE to use, it takes
output to stderr into account, so would start to behave differently when run
under Valgrind.

--trace-children-skip appears to be new since Valgrind 3.6, and older versions
would probably fail early when they see this unknown-to-them option.  If this
turns out to be a problem in practice (current version is 3.9), we probably need
to make this conditional on a configure.ac check.

Change-Id: Ia6a72bdcd666d68ed0539170f3fc476292e82b96