fdo#74702 DrawOutDev() and CopyArea() functions now noops in Printer
DrawOutDev() and CopyArea() functions now assert when called from a
Printer instance. This has allowed the removal of meOutDevType, which
I hope to eventually completely eradicate from LO.

I've also included DrawOutDev(...) which holds a parameter with a
reference to a Printer object. This shouldn't be possible, so it
asserts.

Change-Id: Iabee853ca7c8bfc99349da086d1944ee13210a2e
Reviewed-on: https://gerrit.libreoffice.org/8592
Reviewed-by: Norbert Thiebaud <nthiebaud@gmail.com>
Tested-by: Norbert Thiebaud <nthiebaud@gmail.com>