Hopefully fix Windows HiDPI toolbar layout bug
On Windows HiDPI, toolbar buttons are cut off. This may fix the problem.

Here is a screenshot:
http://i.imgur.com/NADAvYi.png

I can't prove this fixes anything on Windows because I can't see this
on Linux and don't really understand the surrounding code. On the other
hand, it is easy to prove this is reasonable code.

Change-Id: I69c19ad46844bead942ce63883d163cb9d0690c9
Reviewed-on: https://gerrit.libreoffice.org/8637
Tested-by: LibreOffice gerrit bot <gerrit@libreoffice.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>