ooxml: import combobox sdt control
This patch modifies the importer to reuse the code that imports the
dropdown control and creates a combo box with it.

The only drawback is that the combobox tag is transformed into a
dropdownlist when exported back, but it doesn't seem to be an issue
from the point of view of the user.

Change-Id: I340d0cf02d8c5ea575ad4784a1242d2a3e9bd214