ooxml: preserve dataBinding tag in SDT properties
SDT blocks can contain document properties, like the author. This is
done with the dataBinding tag inside sdtPr.

Added a resource definition to the ooxml model file, so the parser can
process this tag. The attributes are stored in the existing grab bag
used for other sdt properties and written back on export.

The code to write the SDT block in the exporter has been modified to
be able to extract the new tag from the grab bag and write it to the
document.

Added a unit test.

Change-Id: I18f35a61784eb0a03ab9ca2cc398ca0df289956c