Resolves: #i49980# Autosense fax4cups devices and configure automatically
If there is a "Dial" ppd key then it's a fax device. If the Dial is set to
"Manually" then ignore that it's a fax device and don't prompt for a fax
number.

If we have a fax number, then set the jobname to the number given that
"printing will result in sending the fax to the number contained in the job
name"

Remove the weird and wonderful non-macosx-unix-only, non-pdf-print-backend,
text "scraping" of phone numbers out of the print out in the @@#PHONE@@ format.

Allow the phone entry dialog to accept multiple fax numbers separated by ;

Change-Id: Ie1d84b68584e82df15ceda5f32c7c75086b035c8
Related: rhbz#159153, #i49981#, #i49920#