Avoid "AquaSalGraphics::copyArea() for non-layered graphics" message on iOS
It seems this comes from the code attempting to 'scroll' the 'frame'
window. With tiled rendering there is of course no real 'frame window'
that would be displayed and need to be scrolled, so just bypass that.

Unfortunately this had no effect on the Smart Art rendering artefacts.
I had hoped those messages were related to rendering of them, but no.

Change-Id: Id4624eb5c1d825a120ef48bd2ae424f147060d35