ooxml: Make SdtHelper grab bag more generic
Flexibilize the way the grab bag in SdtHelper works, enabling it to
store several children of SdtPr.

For every tag inside SdtPr, we enable the DomainMapper grab bag,
store the children properties there, and insert the full grab bag
inside the SdtHelper grab bag. In this way, the SdtHelper grab bag
becomes a Sequence of pairs "SdtPr token name" -> "Bag with children
tokens and values".

The first advantage of this implementation is that we can preserve
dataBinding sdt property plus another one of those supported; until
now the second property overwrote the first one in SdtHelper bag. In
further patches we will add support for other tags.

Some lines are incorrectly indented to make the purpose of this patch
more evident, and will be corrected in the next patch.

Change-Id: I0ec7be7b96dca455bbbeb03fc2fed230df04c52a