use $(MAKE) for recursive make invocation
That's apparently the only proper way (to get e.g. -j right), and
even the toplevel Makefile is so full of gmake-ism that it simply
has to be invoked using GNU make already, hence $GNUMAKE is pointless
except for usage in configure (which says to start the build by
invoking it).

Change-Id: I6060da6f1dad2afc5845ac29b8ac02348c057d3d