SwXStyle::getPropertyValue: don't try to read a float into a sal_Int32
The problem was that for the CharDiffHeight UNO property,
SvxFontHeightItem::QueryValue() put a float to the returned uno::Any.
When we noticed that it's not sal_Int16, we tried to read that into a
sal_Int32 (which failed), and then wrote back the result to the
uno::Any, that's how the result value was 0.

Fix this by checking if reading the uno::Any into a sal_Int32 succeeds.

Change-Id: Ie2269a24fe82ae241940811e7a2f5deaf9f1aeb0