oox: Preserve table style.
Table style is a property that defines a set of background and line
attributes for cells. These attributes depend on theme settings (e.g.
theme colors).

We added a grabbag to the table object to save the table style name.
We detect that name on export and write it back to the document.

Finally, modified an existing unit test to check this attribute too.

TODO: To get the table style working properly after a roundtrip, we
must preserve it and also check that its values are not being
overwritten by different cell or table properties.

Change-Id: Id0e022a389561960c21ab874db33649499735024