fdo#76979 DOCX export: a:srgbClr should be always an explicit color
The brush color contains both the color itself and transparency as well.
In case of DocxAttributeOutput::FormatBackground(), the sColor should
contain the color part, and oAlpha the transparency part.

In case of drawingML export of Writer TextFrames, an automatic color
(0xFFFFFFF) shouldn't be recognized as "auto", as that's invalid:
instead an explicit 0xFFFFFF with zero transparency should be written.

Fix the problem by calling msfilter::util::ConvertColor() for the RGB
color only.

Change-Id: I2132375a4dbcac1667aed49c74e581183178b0cc