fdo#69157 Apply object-in front selection on mouseUp
Before, all selection was recklessly replaced when you clicked something
else than a scaling handle (or the like).
It caused bug 69157.

But now, you can still drag the frame by gripping the interior one.
Btw, that the timer did not correctly start was because of the return
statement in the prior state.

Signed-off-by: Lennard Wasserthal <Wasserthal@nefkom.net>

Conflicts:
	sw/source/core/uibase/docvw/edtwin.cxx

Change-Id: I5e02cfb2d5fe9cdb9fd7f50d0c961dcc418fadd6