Unify DrawGradient functions in OutputDevice
Turns out that the two versions of DrawGradient in OutputDevice are
almost exactly the same in every way, except one deals with a rectangle
and the other with a PolyPolygon. So I just convert the Rectangle into
a PolyPolygon and use the PolyPolygon function.

Now that the functions are unified, the need for a seperate function
to clip and draw the gradient is no longer really required, so I've
merged this back into DrawGradient.

Change-Id: I94d4af1bb7dd900495672f0c0481dc9a1083ff67