fdo#77342: sw: fix copy/paste of footnote cross references
The fix for sequence fields broke the footnote references some more;
simplify it by handling the footnotes in the same way as the sequence
fields, and not remapping GetExp fields for which the corresponding
SetExp field / footnote is missing.

Also, don't do any remapping when the target is a clipboard document, to
prevent modifying the source document.

(regression from bb665affbd8870652ade3951d626d76e99143f67)

Change-Id: If1be1e1d9742182a4085bbbff53e26a8fa8665b8