No need for XShape dumping (for unit testing) for non-desktop
We don't do unit testing when cross-compiling anyway. (And even if we
did, in some magic fashion, I don't think we would want
debugging/testing functionality like this dumping thing to be dragged
in into a production build of an app. It's not a huge amount of code,
but still.)

Change-Id: I12bc552b10452263572074aae865138ff30fbfc8