fdo#71556: Remove unwanted a11y event on text insertion
The accessibility event text-attributes-changed was being emitted
every time we typed because typing implies changing the RSID internal
format attribute. This change was triggering the event but it
shouldn't, as it is an invisible attribute.

I had added a new member to SwUpdateAttr class to be able to indicate
which format attributes are being changed (until now, it just
indicated that some attribute changed with RES_TXTATR_AUTOFMT and no
more details). This member is populated when an SwUpdateAttr object
is being created for a RES_TXTATR_AUTOFMT. Finally, I check the
contents of this list to look for relevant a11y attributes before
issuing the events, if there aren't then no events are issued.

Change-Id: Ieaedf888ccc1bff9c7af64a1412ca1d3b5411351