VCL: move native widget rendering functions out of OutputDevice
The following functions should be in the Window class, not in
OutputDevice:

+ IsNativeControlSupported
+ HitTestNativeControl
+ DrawNativeControl
+ GetNativeControlRegion

Additionally, moved nativecontrols.cxx to vcl/source/window/ and whilst
we are about it, it turns out that VirtualDevice isn't used by these
functions. Therefore the 'orrible check to for the type of class can be
removed and in fact as VirtualDevice doesn't use it at all then we can
just remove the function and replace it with a call to
IsNativeWidgetEnabled().

Change-Id: Idd0bfb1cba1c2902f7a6d55d258efb38b67fb827