do not have a separate action for accepting changes in a selection
It doesn't really make sense to have two different UI elements, one
for a change that cursor points at and one for all changes in a selection,
as it's always either one or the other, so just keep the one UI element
and do the right thing internally.
So this reverts a part of 4a60f9387ddd832f7c164102b76351e5e520d920.

Change-Id: Ic94a7704cbbc6273e91b0ee478f6304019220c4d