Correct keycode mappings for KEY_QUOTE{LEFT,RIGHT}
/Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX10.9.sdk/System/Library/Frameworks/Carbon.framework/Versions/A/Frameworks/HIToolbox.framework/Versions/A/Headers/Events.h
says that 0x27 is kVK_ANSI_Quote, which we call KEY_QUOTERIGHT,
i.e. the plain ASCII apostrophe, while 0x32 is not the ASCII tilde but
kVK_ANSI_Grave, which we call KEY_QUOTELEFT.

Makes the Command-' binding (see Kohei's
56cce6fd8151a916e1be540ac18724b46d1b15fc) work at least for the US
English keyboard. But not for the Finnish keyboard, for instance,
where the key labeled with an apostrophe (or "ASCII-style" single
quote) generates a different keycode than the single quote key on the
US English keyboard. (Which, I guess, isn't surprising as it after all
is a different physical key.)

In other words, situation normal, using keycodes is a horror as
always...

Change-Id: Ibf8f994370fec75849391e1dba30c7b7d54d237f