fdo#57197: sw: fix crash on Undo of table row insert on 1-row table
... with accessibility enabled: the SwCellFrms are deleted, but the
entry in the SwAccessibleMap::mpFrmMap is not removed, because Dispose
returns early since the connections to the nodes have been severed already.

This then crashes in some questionable code that iterates over and uses
the frame pointers in mpFrmMap that was added in
76c549eb01dcb7b5bf28a271ce00e386f3d388ba (where previously only the
WeakReferences were used out of that map).

Change-Id: I13885c6ab9a92e34e95925f275e15c20fdc600b7