provide the OpenGL window
This approach looks much better. We get size and position correct
without much work and can easily plug the window into the sc window
hierarchy.

We still have a crash on exit as the ScGridWindow goes out of scope and
the SystemChildWindow is still alive. We need to fix it and bind the
lifecycle of the SystemChildWindow to ScGridWindow.

Another open item is the OpenGL context work. Right now it looks like
the best way forward is to create a subclass of SystemChildWindow that
contains the OpenGLContext.

Change-Id: Ie0a74531e1b818cea92912345464c8fa219bbae2