AOO: #i124823# fix the calculation of leading metrics for CoreText
tml says: As such, for us the sample document from the above AOO issue
already gets laid out corectly (according to how the bug reporter says
it should be laid out), both in master and 4.2. We already set the
external leading field based on what CTFontGetLeading() says.

Also, our code around here is a little bit different as we don't have
any fFPIY paramter or mfFontScale field, so the change did not apply
directly.

But I adapted it and applied anyway. Effectively, for us the only
change is that the internal leading is now returned as non-zero. No
visible change to the sample document's layout.

Change-Id: I4f8967723e371fa2b695f726ff2f28c0f6d6ae27