Fail early in OpenGLContext::ImplInit() on platforms where not yet implemented
So far this code apparently has not been invoked except in the very
special use cases for which it has been developed, and which
presumably aren't such that one would come across in "normal"
LibreOffice use.

But now it gets invoked unconditionally, in some unit tests even, and
letting it call glewInit(), which calls glGetString(), without any
OpenGL context, leads to a crash. So instead, explicitly fail early on
OS X (and iOS and Android).

Change-Id: Ia4547cc1ddff1aaa0190fcdb69506ad444214b7a