kmz3collada: Workaround for missing textures
The write-only-.dae-file-which-we-need.patch makes the converter
extract only the *.dae file but textures are stored in seperate
image files we need to extract too.
On the other hand this patch makes the converter buildable on all
platforms, since this only works on Windows originally. So for now
disable this patch on Windows so here we won't loose textures
but enable it on other platforms for buildability.
Later we need to solve kmz conversion properly.

Change-Id: Iede9e3bdd5a68a9002f5fa38da8a6459dda8d805