vcl: Refactor scale "super" out of bitmap and make it independent
Introduce BitmapFilter as a general bitmap filtering class, and
make scale "super" algorithem independent as BitmapScaleSuper
which uses BitmapFilter as superclass.

This is an ongoing work to make some bitmap algorithms structured
and more independent from the big bitmap class This will make them
easier to work with, test and optimize.

Change-Id: I37d29709b2af95cab2f6da21129302f5be79318b