sw: let layout not mark embedded object as modified
The problem was that right after the import, the document wasn't marked
as modified, but as layout settled, it got marked as modified.

Given that we already have code to avoid modification when updating the
replacement image, do something similar here: when setting the view
area, make sure the embedded object's component is not marked as
modified.

Change-Id: I88a42829ec48db54178553661863a571cd2a268b