fdo#71558: Notify mispelled word to accessibility
This patch modifies the spell checking code to trigger an invalidation
of the accessible paragraph attributes when a mistake is detected, so
the 'text-attributes-changed' event is released.

Notice that 'text-spelling' is not an actual attribute of the text for
LibreOffice, that's why this didn't work in first place.

Finally, notice too that when the user disables the automatic spell
checking, no 'text-attributes-changed' event is released, and when it
is enabled again it will only release 'text-attributes-changed' events
in case some of the newly typed text is misspelled.

Change-Id: Ie792f040d84010cadcd41486d948a5e29421228f