Force --std=gnu89 for clang.
clang builds with a more modern standard by default

On Linux gcc builds with gnu89 by default. On Windows MSVC only
supports C90. Hence it makes sense to force gnu89 for clang to
catch potential gnu89/C90 incompatibilities locally.

(C90 and gnu89 appear to be quite close, however using C90
with clang results in warnings and errors to do with comments
on lines with code (-Wcomment), and also the lack of inline
(instead you would use __inline -- this is #define'd in for
 Windows only).)

Change-Id: I569af7541b86aa0f83ebb66770d43737bac5d21f
Reviewed-on: https://gerrit.libreoffice.org/9904
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>