SwFlyFrm::Format: inform our custom shape about size changes if needed
This adds support for the following scenario:

- Shape has no editeng text, but a textbox
- Shape has automatic size
- User edits the fly frame (serving as the textbox of the draw shape)
  and expects that both the shape and the textbox gets resized
  accordingly

Note that this way we can signal our size request for the textbox, and
the custom shape takes care of resizing the shape in a way, so that
after the resize, the text area counted from the custom shape geometry
will be exactly what we requested. (I.e. just asking the shape to have
the same size as the fly would work for rectangle, but not e.g. for
triangles.)

Change-Id: I1e7efd00db52cf28481f6821ea6949f748b855e7