Resolves: fdo#73806 reduce the size-request of the expert config dialog
the optimal size was set to be 1024x800, reduce this down to
effectively the same size as the options dialog itself.

This dialog just isn't ever going to be right given the huge
length of the strings and the number of strings. We need
something more like a file picker dialog view here

Change-Id: I019e951601c3df769a609b8b439b535b6d5e254f