fdo#70346 MM: add mail merge data to condition dict
Currently section hide conditions are just evaluated based on
document field data.

This adds the current mail merge dataset to the condition
dictionary, so this can be used in the condition evaluation.

The dataset values are named from their columns and added initially,
therefore DB fields with identical names will overwrite the values
in the dictionary!

Change-Id: I82a5f9f6962628a76c836e8e2a7c9e264fdc16e0
Reviewed-on: https://gerrit.libreoffice.org/10978
Reviewed-by: Björn Michaelsen <bjoern.michaelsen@canonical.com>
Tested-by: Björn Michaelsen <bjoern.michaelsen@canonical.com>