ww8import create a pagedesc if continuous section changes margins (bnc#875383)
This is similar to what writerfilter does. MSWord can have one page with several
different margins, which are saved using continuous sections, which causes all
kinds of trouble, because either we treat them as Writer sections, which means
we lose some of the data, or we treat them as Writer page styles, which causes
spurious page breaks if in the wrong place. Either option has its problems, but
here it seems slightly better to go for keeping the data and hoping the page
break will be in a place where a break will be anyway.

Change-Id: I8f52aa820750da6788ea04180a15ac334f6bf87b