Our log area checking Clang plugin requires it to be a string literal
And even if that wasn't a problem, we don't use a macro for the log area
anywhere else in the codebase, so why suddenly introduce such a convention
here? What's the gain from that? On the contrary, haven't we been *reducing*
the amount of pointless #define FOO "foo" style stuff?

Change-Id: I1247dc499136cdf3ae1921e6ffe7c7f63d57cfc1