writerfilter: Kill AttributeToResourceMap.
IMPORTANT: From now on, the order of attributes becomes stable, based on the
order in model.xml (not on implementation details of unordered_map), and the
code that handles attributes may depend on a particular order.  If you want to
change the order how the attributes are handled, change model.xml, and check
you achieved what you wanted in the generated ::getAttributeInfoArray()'s.

[Writerfilter loses another 250k (stripped dbgutil).  And the usage of
unordered_map here was just completely bogus from the very beginning, as it
was only iterated as a normal array anyway ;-)]

Change-Id: Ic70c37793e313c4ccda1d6f374cc2d366307ba1b