fix image dragging inside a document
Testcase is a document with a table and a wrap-through anchored-to-character
image below it. Dragging the image up to be anchored to a place in a cell
and positioned partially on top of the table does not actually change
the image position if the image was below the table before the drag,
but changes when the table was above. In the faulty case,
SwToCntntAnchoredObjectPosition::CalcPosition() returns prematurely
because of paragraph portion information not existing at that point
(interestingly, the problem is also fixed simply by commenting out the return).

This basically reverts 9b87346bf4a1d32abfa781aa66d4d5b627133edf . I'm unable
to check the original problem as the bugreport references a web page that
no longer exists. So if that one still exists, it'll need to be handled
whenever somebody runs into a way of reproducing it.

Change-Id: Id354237dace3179b94768146ec1a599cc3d49146