Ugly hack to fix build when srcdir!=builddir
The problem is that the ure/source/unorc file is now expanded by the configure
script, and thus exists only in builddir. But a further complication is that
the uno.ini file is in srcdir. This is one way to handle it. Seems to work for
me, let's see what the srcdir!=builddir tinderbox slaves say.

Change-Id: I6fb456cf849ce5077e2c5bd25dc9149096aab241