correct references after sort, fdo#79441
5c6ee09126631342939ae8766fe36083d8c011e3 introduced a different
algorithm for reference handling during sort. Unfortunately that clashed
with the SC_CLONECELL_ADJUST3DREL introduced a little earlier resulting
in relative 3D references effectively being "adjusted" twice.

Furthermore, in-sort-range range references to one row (or column) were
not adapted to the move at all if the formula within the range listened
only to ranges and not a single cell. Added collecting and adjusting
area listeners for this.

Last but not least, external (relative) references need to be treated
the same as internal 3D references, making them point to the same
location after the sort.

Change-Id: I492768b525f95f1c43d1c6e7a63a36cce093fa5a