fdo#82580: Win32 GetObject() simplification
We call the Win32 GetObject() in just one source file and there it is
the A version we mean. (As such, as we call it to get BITMAP objects,
which do not have different A and W variants, we could as well call
GetObjectW().)

Note that we have several own member functions called GetObject(), so
until those are renamed we still need to #undef the GetObject
definition (from <windows.h>).

(No, I am not working on the EasyHack fdo#82580 as a whole. It is
intended for some less mature LO contributor. I just did a few small
things because I couldn't resist.)

Change-Id: Idab79503b3d899ef19f0608677752a62b20302ff