Customproperties: Correct size for value field
Before this patch types "DateTime" and "Yes or No"
were one pixel too big in height.

Fix this by using the original aSize and aPos values, don't
depend on m_aYesNoButton size and position.

However there are some glitches with DateTime: If you scroll
some times up and down the list of "Type" the Date and Time
fields will get positioned somewhere left.
This has to be a problem of DateField and TimeField, because

m_aDateField.SetPosSizePixel( aPos, aSize );

and m_aTimeField get the same values as all other fields...

But this positioning error existed before this patch, too.

Change-Id: I793aebf39f5b6cb6e4b290f21a5dbcc7ce6ce964