Base the convert-to detection on the actual service
There is no need to run a type detection on the input
file (for a second time!), just to get the document service.

This also simplifies the conversion of some generic formats
like HTML and RTF, using Calc export filters. Previously
setting both input and output filters was required. Now it's
enough to set the input filter, or simply use --calc.

Change-Id: I825e0fe3e7368d6649acb434b13de53a4523064a