Avoid lookup table: just one single special case
Caller uses indexes from 1 to GetFormMax(), that is AUTH_TYPE_END+1
for TOX_AUTHORITIES and is smaller (23) that the number of entries
in the lookup table (30). The index is compensated for this offset,
so that the following check for special case AUTH_TYPE_WWW is
correct, but no checks on maximum value are performed since it
should have never reached the last special item of the table anyway.

Change-Id: Idd85e509da1e8ab4cdda0070f513c4e07008d410