bnc#770711: Make changing table text color in impress always work
We set text color for individual characters (in ContentNode::aCharAttribList)
and for whole paragraph (in ContentNode::aContentAttribs) too.
Problem: when table was loaded in an impress document with font color set,
you could not change it by selecting a cell because only
ContentNode::aContentAttribs was set, which has lower priority.
So, in that case, remove font color character attributes.

Change-Id: Ieb0c078481568dcb1739fe35d448c9591f687f2c