bnc#648251: Avoid crash when attempting to open embedded OLE object as "text"
On non-Windows, when double-clicking an embedded OLE object, our glorious
content type detection logic detects it as "Text". As a side-effect, we start
to calculate text statistics on it. Which surely could produce interesting
numbers (you know what they say about statistics), but sadly causes a crash
involving the ICU RuleBasedBreakIterator, SwScanner,
sw::DocumentStatisticsManager and whatnot.

Avoid this by checking for a detected filter of type "Text" explicitly, and
avoiding the fun code paths in that case.

This leads to double-clicks being just ignored. Maybe it would be more useful
to produce a "General OLE Error" message box?

Change-Id: Iae0726b5e9c511a92bdff7229d2978cbf76cb07b