fdo#55903 Immediate effect for zoom in button
That's what apparently people expect from this button
to do, similar to the nearby zoom out button. This is
also consistent with a similar button in math, and in
print preview (both are using the same icon and same
tooltip).

The zoom mode that was assigned to this button until now,
moved to a dedicated button (thus could be further improved
to host other functionality as well - see the next commit).

Change-Id: Ie640c72cf1aab1e3fc8a14211702a33a86bb3672