default n-up printing of notes to sensible 2 x 1 not 1 x 2
i.e. throw away the maFirstPageSize cache based on the initial slide view which
is typically in landscape mode

So that if we change to notes which are usually in portrait mode, and then
visit n-up print, we get a default layout based on the notes orientation and
not the slides orientation.

I hate printing

Change-Id: I8b7b81ce1eec0f9c5ecd7509b311cf4026958c2c