SwCursor::SelectWordWT: include comment anchor character of a commented word
In case the user double-clicks on a word and exactly that word is
commented, then we used to select only the word, but not its comment
anchor character. So when the user deleted the selection, the comment
left there.

It is still possible to manually select only the word, but on
double-click let's just select the comment anchor as well, so a simple
double-click on the commented word will remove the comment as well.

Change-Id: I7322af61c31e8aba108eef5596849d17d55fc7eb