fdo#81237: OpenGL window is not resized by view change
Reproduce:
- Create a 2D OpenGL chart
- Step out from edit mode
- Zoom in / zoom out

Behavior
- Chart window is not resized, it has a constant size.

When we step out from chart edit mode the chart becomes
an inplace active object (instead of UI active) which means
the m_xDockingAreaAcceptor is removed. On the other hand
it seems we don't actually use this m_xDockingAreaAcceptor here,
so don't need to check whether it exists.

Change-Id: I056109b6e5cac28cdd2aeb8fabd9190413a16ed5