fdo#35862 De-/Increase font when multi-sized text
Added two new function to the SwEditShell.
The first returns all items of a specific WhichId wich belong to the current selection. The items from the attribute set and from the hints are used.
The second returns SwPaMs separated at borders of an itemtype  identified by its WhichId. The SwPaMs must be deleted after yous.

Change-Id: I00e6d1b0f75b9f184343711c6d2bdc7e5694c711
Reviewed-on: https://gerrit.libreoffice.org/11857
Tested-by: Samuel Mehrbrodt <s.mehrbrodt@gmail.com>
Reviewed-by: Samuel Mehrbrodt <s.mehrbrodt@gmail.com>