android: rerender tiles when the zoom changes
Initial implementation which needs a lots of polish. Sometimes
the app crashes (for yet unknown reasons). Sometimes a gap
between tiles is visible probably because of rounding errors.
Tile handling and releasing after zoom is quite naive for now
(just clearing all the tiles) which needs to be improved to be
more seamless.

Other changes: The responsibility of LOKitThread to handle tiles
was moved to MultiTileLayer. TileProvider now takes into account
the current zoom level when requesting tile rerendering.

Change-Id: Ic188c03bfadf3a3dd2f79b04f07668eb63b705fb