SwTxtNode::Update: don't move anchors at insert position
Use cases:

1) User of the XTextAppend API's uses appendTextPortion(), then
appendTextContent(), then appendTextPortion() again. It's expected that
the text content will be in the middle of the paragraph, but before, it
was at the end of it.

It's possible to work this around by inserting the text contents only at
the end of the paragraph (like it's done for ODF import in
XMLParaContext::~XMLParaContext()), but probably it's not a good idea to
copy&paste this workaround in all the import filters.

2) User types some characters in a new document, then inserts a picture,
and sets anchor type to at-char. Now the anchor point is set to the end
of the paragraph, but it's impossible to type anything after that
character.

To be fair, now it's not possible to type something *before* the anchor
point, if it's at the first character, but hopefully that's less of an
issue, since setting the anchor type to at-char sets it to the end of
the paragraph, not at the start of it.

Change-Id: Ia8b33e8880aae86f62fae09cd5ac1cf82c83320f