windows: try to locate a good MSBuild.exe based on the version of VS
lcms2 for instance call MSBuild.exe, naked. it relies on what is found
in the PATH, but apparently MSBuild.exe can be found in many places
including for instance the .NET Framework.
This was causing failure with a VS2013

This patch inspect the registry to see if there is a MSBuild of the
same level than VS2013 and use that information to adjust the PATH
so that we invoke the right version.

Note: starting VS2013, Microsoft is numbering the version of MSBuild
with the same version number than VS itself.. so with VS2013 the version
of MSBuild is 12.0 (vs 4.0 before that)

Change-Id: I04c614116d1d365a10c9b9e0e8ce9571d770c065