Add more detailled configuration for OpenCL use, and make OpenCL default to on
Add a boolean whether to use OpenCL for just a subset of functions, and a
numberical minimum length of a cell group for OpenCL to be used. No
functionality or GUI for this yet implemented, just the setting fields.

Make whether to use OpenCL default to "on". Sure, this breaks stuff right and
left for now. Fixes and workarounds later.

Change-Id: I09b7aff9c7d1845df2bf01f09616e646373f030c