fdo#81264 Accessiblitiy focus not tracked for cells in Calc
When the accessibility roles for DOCUMENT_SPREADSHEET etc. was introduced
the following places in Windows specific files where not updated to take
these new roles into account. This patch should hopefully fix that. As a
side effect the initial code for tracking cell movement in Calc works again.
Well at least as well as it did before the change, there are still some
quirks left to fix.

Change-Id: If3cacdc27a73c4cee85b1450c406d490bd741456
Reviewed-on: https://gerrit.libreoffice.org/12261
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>