SwFrmDlg ctor: turn sRestype into an OUString
In practice, currently possible values are ASCII, but then this string
becomes part of URLs, which are in general no longer ASCII, and are
OUStrings. Instead of trying to guess the encoding of the string we get,
require callers to create an OUString right away.

A follow-up commit could adapt SfxTabDialog ctor,
SwFrmPage::SetFrmType() and SwFrmAddPage::SetFrmType() accordingly to
avoid the new introduced OUStringToOString() calls.

Change-Id: I087ed2bb341f5aca59e15e2ef4102556ca803363