Related fdo#82953: Forget package URL of image after it is loaded
It causes problems if we handle those imported images differently which
are identified by a package URL, so after the first load remove
this URL and handle images on the same way as inserted images.

Some related bugs:
* #i44367#
* #i124946#
* #i114361#
* fdo#73270

The image in the test document has a special ID which is different
from that one which is generated by LO internally so after ODF export
the new generated image URL is different from the imported one.

Change-Id: I4e7d3490674c5f86bec5c7c6e1c975dcafd7c265