Be honest about the maximum cache size
Implementation uses a multiplicator of ten. I guess the intention
here was that to calculate cache size based on the size of the inserted
images and not based on the real memory usage (potentionally more
instances of one image), which seems a bad idea because it covers the
real memory usage from the users (who will be suprised when LO's memory
usage increases tenfold of the set value).

It seems better to use the exact value in the implementation and increase
the default value on the user interface instead.

Change-Id: Ie3c0568d3518922d3a9931f894eebb9e7e5515ed