Add a spreadsheet to test OpenCL implementation on user request
Loading this spreadsheet and verifying that the end result in A1 is "PASS"
will be bound to the 'Test' button on the calculation details dialog.

Actually, I wonder whether it might be saner to just populate a spreadsheet
programmatically instead of having a (not huge, but still large) test file
bundled in the installation? We could make it so that the generated document
stays open and is usable normally, so that if necessary (for instance for
comparison with other spreadsheet applications) it can be saved.

Change-Id: Idea972cd0f598643e35314576f593047d49378c6