[Accessibility] Make the sidebar a bit accessible
The accessibility of the sidebar is heavily built on PANEL's.
Since A11y on mac ignored panels it did not pan out well.
Also the title was removed from AXGroupRole but is needed
to announce the different panels of the sidebar.

This might have some unwanted effects on other parts of
LibreOffice since a lot of things are built up from
AccessibleRole::PANEL.

Change-Id: I095d26b3c89e2c0403d3267baa53f3bec5e07662
Reviewed-on: https://gerrit.libreoffice.org/12985
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>