OpenGLContext stopped to support multisampling under linux.
glTF rendering needs to have a context with multisampling
so it also became broken.

Regresssion from:
07c0396ab9c94be5af392827f56f0d91551f3e25

This breaker change was made assuming getVisual{Info}() is always
called after the SystemChildWindow is already created, but
generateWinData() is run as part of the new child window creation.
It's aim to choose the best visual / FBConfig which potentially support
multisampling.

Change-Id: Ic881c6769260b5eaa9fed2552eca882191f6c56d