make FastSaxParser provide the whole content in one characters() call
SAX interface is not required to provide the whole node content in one
characters() call (e.g. if there's an entity that needs decoding). However
it's easier to consumers to assume this (e.g. writerfilter's
DomainMapper::lcl_utext() handles datecontrol that way), and expat apparently
never used this. However this can happen with libxml2, so ensure such consumers
still work.

Change-Id: Ib564f372fbea8451f84553a6d49a07d091a950e9