DOCX export: fix page border spacing in case of headers/footers
It is clear that Word/Writer has different page margin / border spacing
models: in Word, the page margin includes everything and border spacing
is a subset of that, in Writer the real distance between the body frame
and the page frame is the sum of the two.

What was less clear is what headers / footers add to this mix. It turns
out that in case we have a header as well, then for large borders, we
need to count the difference of the page frame and the header frame, not
the difference of the page frame and the body frame.

Change-Id: I3c828eeec2df461985110064ab0bf21b36594095