Add comment about a bug but how this stuff still works by luck
The 'suffix' GNU Make function returns the file name suffix including
the period. The test comparing to the string 'zip' will thus never
match, and gb_UnpackedTarget_STRIP_COMPONENTS_ZIP_DEFAULT will not be
used, but gb_UnpackedTarget_STRIP_COMPONENTS_TAR_DEFAULT. But as most
of the Zip archives we unpack do have a top-level with a single
directory anyway, that we want to "strip", that works out
fine. Apparently those that don't have a such directory level pass a 0
as second argument to gb_UnpackedTarget_STRIP_COMPONENTS_ZIP_DEFAULT
which has the effect of avoiding the "stripping".

Not sure what to actually do here, so I just commented the
situation... Should the code be fixed to do what it thinks it is
doing, but then to keep things working as before, should
gb_UnpackedTarget_STRIP_COMPONENTS_ZIP_DEFAULT be changed to 1?

Change-Id: I6436865dafe47e21e1365a602889cedab3c09784