don't copy stylesheet's HelpID if this's a rename
This causes problems for renamed copies of std. styles, as they have
special handling in SdStyleSheet::SetHelpId, which would set the API
name to std. value too. That means that there would be two (or more)
stylesheets with the same API name. On export, only one of them would be
exported.

Change-Id: I19d243bd3f73995e8de2d251e3d8b7f93219c3a9