in ScFormulaCell dtor remove cell also from FormulaTrack
It could happen that during a SetDirty/Notify cycle a formula cell is
appended to the formula track but not tracked yet so doesn't end up in
the formula tree. If it was deleted then without removing it from the
track the cell pointer shortly after was moved into the tree, possibly
setting pFormulaTree (and/or pEOFormulaTree) to that cell if it was the
last cell, and if immediately after that a new ScFormulaCell was
allocated at exactly the same memory location it virtually ended up as a
successor of itself in the formula tree ... leading to a crash if pCode
was accessed in a subsequent RemoveFromFormulaTree because the cell was
assumed to be already in the tree.

Change-Id: I9d1885a26b85c2331a084b5f89a2d7373cf2df0f