Don't use of the broken "Software" group interpreter
FormulaGroupInterpreterSoftware is known to be broken, says moggi. We should
not use it as a fallback to OpenCL.

Not sure whether it makes sense, but let's keep it in the code for now.  Make
using it conditional on setting the environment variable
SC_ALLOW_BROKEN_SOFTWARE_INTERPRETER (to any value). Only a developer who
wants to work on it should set that.

sc::FormulaGroupInterpreter::getStatic() can now return NULL, adapt callers
accordingly.

This might fix fdo#87119, but I am not able to check, as the document from
that bug causes an unrelated assertion for me.

Change-Id: I24454f46332014cbcbf696252059b9743f3c84d6