Draw a rect in the uncovered region of the native scrollbar
When the requested region is larger than the native scrollbar the
uncovered regions need to be drawn with a background color to
avoid artifacts.

This situation can happen when the TabBar is inline with scrollbar
when the width of the TabBar is greater than the width of the
scrollbar. Inline TabBar support is added in a later commit.

Change-Id: I70cb2e5bfd8d79401f77bbe55abe596adc160194