fdo#86674 paragraphs have ATK_STATE_INVALID
A few states have been added in the IA2 integration. These need to be
handled somehow on linux as well. At the moment they all get the state
INVALID. Let's add the state DEFAULT that already have a equivalent.
The state MOVEABLE is causing the issues in paragraphs of LibreOffice
Writer. The state is not invalid so let's avoid reporting it at all.
There must be a cleaner way of doing this however.

Change-Id: I859439f92fe19f5050195383f43fa4af787b249e
Reviewed-on: https://gerrit.libreoffice.org/13727
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>