Bin obsolete conditional compilation
USE_FMIN_FMAX has been defined a long time. There is no information in any
comment or in the commit that introduced it why one would not want to use the
OpenCL fmin() and fmax() built-in functions (intrinsics). Keeping such stuff
in the source code (and in the generated OpenCL code) is pointless and just
makes it harder to read.

Change-Id: Ibc0b950c0f16872867eb5eff089b33b00cb95028