gbuild: try to fix windows non-debug builds
The ENABLE_CRASHREP is no longer AC_SUBSTed by configure, so it is
always non-empty in config_host.mk, resulting in -Zi without -FS,
which breaks MSVC 2013 builds.

Leave a placeholder in configure for the new crash reporter vaporware,
and force the CFLAGS to be the same as in the --eanble-symbols case.

(regression from 8ab4d2f29fdd2c058db23f6ee7b2200655d4c6f6)

Change-Id: I476fb2e4f50de5304632628b25e3997ea6c31c27