SwRootFrm::CheckViewLayout: never center pages when doing tiled rendering
The size of the root frame is usually the max of the page frame widths,
except when the visual area is larger, in which case the root frame is
the visual area width, and page frames are centered (when there is a
single column).

When doing tiled rendering, we are not interested in the gray area
around the pages (not more than just having a gap between pages), so
never do this centering. This has two benefits:

1) When painting the tiles, we change the visual area already to always
have the given tile visible, that would undo this centering anyway, but
changing the size of the root frame causes a pointless
mass-invalidation.

2) When getting the size of the document, this way the width of the
document (root frame) will independent from the visual area, i.e. it'll
be always the same.

Change-Id: Ib0d86bbe6c7b5a83646264a2f308624179ad6982