fdo#86844 - CONTEXT MENU: Remove hyperlink not accessible
At the moment Remove hyperlink is only shown when text is selected
together with a hyperlink but not when you right click directly on
a link. Let's reuse the logic from when edit hyperlink should be shown
and append it to the currently working case where text and link is
selected.

Change-Id: I2524f1c3037f3101d24e592eadf2e7d5d66ab4f9
Reviewed-on: https://gerrit.libreoffice.org/13926
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>