Initial SwEditWin::LogicMouseButtonDown/Up
This makes single left mouse button clicks work while editing. The idea
is that when we do tiled rendering, we get the click position in
document coordinates (twips) and SwEditWin::MouseButtonDown() can check
if we're doing tiled rendering. For now, make sure that the position is
not passed to any other methods in the tiled rendering case, those may
interpret them as pixels, which is not true.

Change-Id: Icbd16411596393ed983a87fd3a59a08e4fda3696