Revert "fdo#88475 BrowseBox/grid: reposition data cursor to current row after paint"
(and its later fixup)

This reverts commits
 e8a68c1f50f32a0f9d8bcdf16c1270c319910baa
 e60b589952985edff12b1a28392ce6fa0ca8d9be

It was a work-around for the real underlying issue, which was that the
result of dbaccess::ResultSet::isFirst() et al were clobbered by moves
made by its clones.

The BrowseBox has two different cursors:
1) One for data to edit (which is kept on the current/active row)
2) One for data to *paint*
The second is a clone of the first.

The real underlying issue is fixed by:

  commit d7c9a1d9d65fe8b1a56c5c280d2ca6640a549d2f
  Author: Lionel Elie Mamane <lionel@mamane.lu>
  Date:   Thu Jan 22 10:49:42 2015 +0100

      fdo#88475 RowSetBase: reposition cache before interrogating it

      Change-Id: I28d62673fdf10ee6507d38bb7c79c08e4b40902f