Use LibreOfficeKit in the TiledLibreOffice iOS demo app, too
For iOS, the pBuffer argument to LibreOfficeKitDocument::paintTile()
is not a pixel buffer, but a CGContextRef. Thus we can keep the
existing CGContext-oriented vcl code for iOS as is.

I commented out the <touch/touch.h> -using code in sw, and it probably
would be OK to just kill <touch/touch.h> and remove use and
implementations of its API. I don't think they are used by our current
Android or iOS code.

This concludes the commits of the work I did at the post-FOSDEM
hackfest and on the flight home. TiledLibreOffice now works as before,
but uses LibreOfficeKit.

Change-Id: Ibbde6e11d660561840eab6335f7fc0db6758173e