Unmap the host buffer only after done accessing it
Most likely 64c479e9da02f724e1870649c99fac92f5f27cd3 accidentally made the
code unmap the host buffer before it is accessed, but the code continued to
work by accident in many (most?) cases. Either because in the case of OpenCL
devices that share memory with the CPU, the host buffer *is* the OpenCL
buffer, so even if the host buffer is "unmapped", it still exists. In the
case of GPU device with separate memory, using the host buffer after unmapping
corresponds simply to a case of use after free of a heap-allocated buffer,
which often happens to work.

Found by code reading.

Change-Id: I9e2b4574077a267938702c0f81c4b1cba9c9a183