Add ScCalcConfig field to DynamicKernelArgument
It will be needed to be able to adapt OpenCL code generation to the
text-to-number conversion settings.

To get the document-specific ScCalcConfig into the DynamicKernelArgument, we
have to pass it along in lots of places to constructors of various
DynamicKernelArgument subclasses. Originally it comes from the maCalcConfig
field of FormulaGroupInterpreterOpenCL, inherited from
FormulaGroupInterpreter.

Change-Id: Iac6a83e8081ab973e8b7e8161de79e411d2ed2dd