Handle string arguments better when attempting to use OpenCL
If a function/operator that takes numeric arguments but not string arguments
is passed string arguments anyway, handle that with OpenCL only if the
(document-specific!) setting for text-to-number conversion is ZERO (treat
strings as zero, which is how our OpenCL code for such functions/operators
works). Otherwise let the code fall back to the traditional interpreter.

Change-Id: I351da75b225baf4b721cd1ab67ce5094dbc5e458