android: discard image buffer as soon as we upload to a texture
Before TileLayer (which SubTile was a subclass of) held the
image buffer for the whole life cycle of the tile, which wastes
memory in case of SubTile as the image doesn't change (or changes
only when invalidated) for the whole tile lifecycle. This change
changes the SubTile inheritance to a general Layer and adds modified
logic from SingleTileLayer and TileLayer. In addition it now
discards the buffer as soon as it has been uploaded to a texture
so that it doesn't takes up double the memory anymore.

Change-Id: Ia104687d2df529182af412102459952b53e87950