tdf#80715: fix infinite loop in find-and-replace if SwTxtNode full
In a regex searching for "$" to remove the paragraph break, sw_JoinText()
may fail to join the nodes because the maximum size is reached; in this
case the cursor has to be moved to the next node, otherwise the search
will be stuck on that node forever.

This would fix the problem in LO 4.2 but it's actually rather harder to
trigger in 4.3+ because the 16-bit size limit of SwTxtNodes is gone.

(regression from b60ce8465c8f01242354abccebe00742d164af60)

Change-Id: Ie047cad37835adf95afe0d12b94a16ff4aecb17a