android: use locking to make tile redraw more predictable
CopyOnWriteList is a good thread safe container to store tiles,
however any change to the list makes a internal copy of the
underlaying array which contains the changes. The effect of this
is that this changes aren't immediately shown or only partially
in the other (UI) thread. So they are sometimes partially drawn or
drawn with a delay. This replaces the CopyOnWriteList with a
simple thread unsafe ArrayList and introduces Read/Write locking
to all ArrayList operations. Read operations don't lock, only
a write operation locks access.

Change-Id: I5783c6cde96360a6fd47faa801eec35e4debb792