android: do tile reevaluation in LOKitThread
Currently the tile reevaluation was done in UI thread which could
cause UI stutters as reevaluation is not a simple task. The
result was also a lot of tile rendering requests to LOKitThread.
This changes turns this around and the tile reevaluation is done
in LOKitThread instead. Now the UI thread just sends a LOEvent
when the reevaluation should be done. This should also reduce
the amount of messages that are queued in LOKitThread, however
an execution time should increase.

Change-Id: I01ce911226a71607c06da6100de323ca555db474