android: support ACTION_MULTIPLE KeyEvent
With ACTION_MULTIPLE more characters have been entered at the same
time. In this case we send more key press events to LO for now.
This is generally useful for IME input but Android also sends this
if non-ASCII key has been hit - even when the key is only one
character.

Change-Id: Iaf48ed4e40b8e4413ed5735add34870ec263b61d