Add OutputDevice::isTiledRendering()
This is similar to SwViewShell::isTiledRendering, but while SwViewShell
describes a state of the application, this one describes a state of the
output device. The output device may be created by LOK client code,
while the view shell may not be accessible from lower layers of the
stack, so we need both.

Change-Id: I82f7755eb73603bd79283272fba4970dccae8784