tdf#89694 Password dialog for opening OOXML shows url-encoded path
We need to convert a path string to decoded it for OOXML with a password as ODF
is so. This patch adds the converting process for OOXML.

In addition, this patch moves the decoded-point of URL into
css.task.Document[MS]PasswordRequest ctor from the outside. This decoded-point of
URL into ctor is common for OOXML and ODF.

(notes: Strange to say, css.task.Document[MS]PasswordRequest ctor needs "aDocURL",
    but methods in interaction-handler use "aDocumentName" as the parameter.
    Probably this will be fixed in near future.)

Change-Id: I0ecfddf0f70ea5810df52d9ca8637f483a26df2f
Reviewed-on: https://gerrit.libreoffice.org/14677
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>
Tested-by: Stephan Bergmann <sbergman@redhat.com>