SwEditWin: when tiled rendering, let double click select an empty cell
When not tiled rendering and the user double-clicks on a word, then
that's selected. If the double-click happens in an empty cell, nothing
happens.

When tiled rendering, still try to select a word, but if that doesn't
work, then see if we are in a table, and then select just that cell.
Then the user can use the usual selection handles to adjust the
selection further.

Change-Id: I6beabf5c73a02f1eb2ede6f8cf9062c1e6610e9e