Fix crash when timestamping PDF signature
Using the NSS API for CMS and ASN.1-based stuff in general correctly is
extremely hard. It is very easy to do things slightly wrong. Of course no
compiler warnings are produced. You just get code that happens to work by
accident when compiled with one compiler, but not another, or depending on
contents of uninitialised memory, or the phase of the moon.

The problem was that the "values" field of a NSSCMSAttribute struct apparently
is supposed to point to *two* SECItem pointers, one pointing to the actual
value, and a NULL one.

Anyway, now valgrind finally does not complain about any use of uninitialised
memory.

Most likely my earlier recent commits to this file were not necessary after
all. They just seemed to help by accident, at least at one stage. But
whatever...

Change-Id: Ic98401b5d151bbb2398f809f47699f670e9720fa