editeng tiled rendering: allow resetting selection with a single click
The problem was that when not tiled rendering, the selection is painted
with the inverse color, and so killing the selection can be achieved
with just painting it again -- using the inverse color again.

For the tiled rendering case, use IsInSelectionMode() to see if we're
killing the selection, and if so, just emit the empty selection.

Change-Id: I045f5b100ae8d4f5802b16f68376aa21de08b994