svx: if the handle bitmap is empty - create a replacement bitmap
When the bitmap with handles is not found (usually happens when
using LO as a library (for example on Android)), the returned
bitmap was empty. This caused that the hit test for the handle
always returned false. This commit replaces the empty bitmap with
a black 13x13 bitmap which should respond to hit test. This should
not influence normal behaviour when the bitmap with handles is
present.

Change-Id: I39330b1867abfb540c215f03b77e5eac2447f0dc