android: when in selection, avoid showing the middle handle
editeng sends LOK_CALLBACK_CURSOR_VISIBLE after the text selection is
already created. The effect of this is that after a long push on a word
you get not only the selection overlay + the start/end handles, but also
the middle one, which is not expected. Just check if we're in a
selection already, and if so, don't react to
LOK_CALLBACK_CURSOR_VISIBLE=true.

Change-Id: I939c957feb7889dfb1997fac1051cb4eaa77004c