add GtkSwipeGesture support and implement swipe left/right to change slides
keep it simple for now.

deliver to the same target window that gets the MouseWheel events, maybe worth combining
MouseWheel and Gestures into the same thing

and use it in slideshows so swipe toward the left to advance to the next slide,
to the right to return to the previous slide.

swipes are followed by mouse up events, impress already has a similar hack
to hide an mouse-up from the (incredibly complicated) interaction with
the slideshow so simply use that

Change-Id: Ib34f6fa0f15f3aa34eef887eb9d5642de9e5cdd1