SdrPaintView tiled rendering: ignore visual area
With this, a number of so far missing images properly show up in Draw
and Impress.

Images are loaded asynchronously, and the AsynchGraphicLoadingEvent is
registered in the SdrPageWindow. This means that in case the event is
registered in the temporary page window that's created by
SdrPageView::CompleteRedraw(), invoked by paintTile() call, then
chances are high that the page window will be destructed earlier than
the scheduler would execute the async event. This leads to missing
images.

Fix the problem by not ignororing Invalidate() calls depending on what
core thinks is a visible area: then the view that is an sd::Window will
also try to do the async image loading, and that will be actually
executed properly.

Change-Id: Icf5fd772b66c31d3876bf3673ba5e283c7e08083