tdf#90245: Support for table merge cell export.
Previously, the merged cells were not getting saved. Also it
was leading to corruption if saved as pptx.
While on opening the round-trip file of odf extension which has
merged cells, merged cell properties not used to persist.

In the current implementation XMergeableCell class was not having
any property to identify the parent cell of a merged cell.

This CL includes:
- Fixed the above scenario for export of file
- Now, the odf file with table merged cells persists its property
- MS doesn't complain for any corruption after export an odf file
as pptx
TODO: Writing a UT seems to be tricky for this change.
      Need to analyze and will raise the UT in separate CL.

Change-Id: I32f9daf77312a0ef3f291f36aef372671554c56d
Reviewed-on: https://gerrit.libreoffice.org/15282
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>