tdf#39547 Image title/description is lost after changing anchor
This issue has been present for a long time, at least since 2010.
When the old frames are deleted the SdrObjPlusData that holds the
title and description is deleted as well never to be transferred
to the re-anchored object.

Some types of frames only have the msTitle and msDesc so let's
keep them updated and also use them to transfer the strings in
question to the new SdrObjPlusData object.

Change-Id: If9b00b70d2556bd32fa2b4f312db4ad09102d9ee
Reviewed-on: https://gerrit.libreoffice.org/15194
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>