tdf#88812 sw: speed up SwDoc::MakeFlyAndMove()
In this function the anchor is not set by the time CopyRange() is
called, so the check always returns false, but in worst case it's O(n^3)
wrt. the number of fly frames in the document, so it's a significant
cost to perform that check.

Before:

$ time OOO_EXIT_POST_STARTUP=1 instdir/program/soffice.bin --headless /home/vmiklos/orig.rtf
real    0m46.833s

After:

$ time OOO_EXIT_POST_STARTUP=1 instdir/program/soffice.bin --headless /home/vmiklos/orig.rtf
real    0m26.832s

Change-Id: I7ab9565634e4df7e0adc701908815f778dc59d32
Reviewed-on: https://gerrit.libreoffice.org/15197
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>