editeng tiled editing: Don't overwrite bInSelection.
Setting bInSelection in MouseButtonUp() feels wrong, and against of its other
handling elsewhere.  Also it breaks the selection handling in the tiled
rendering case - on further redraw, selections may disappear.

But it has been so since the initial check-in of the OOo code, so I have no
idea what corner case I might break by deleting this, so let's guard the
behavior by isTiledReneding() for now.

Change-Id: I7f7b237f9f0f427e317266e05481bac439c3c8ad