tdf#89478: svx: ODF import: SvxBrushItem Graphics should override Color
The bugdoc has:

<style:page-layout-properties fo:background-color="#007f7f" >
    <style:background-image
     xlink:href="Pictures/2000003900002FA3000055F455EC9F63.svm"/>
</style:page-layout-properties>

In this case it's clearly better to prefer the Graphics.

The old code could actually store both in the model but the UI only
allows setting either Color or Graphic, so this should be good enough.

(regression from 7d9bb549d498d6beed2c4050c402d09643febdfa)

Change-Id: Id33231ad90815d9791f90289fb0b55455d334c14