implement a sane and working dxf font export, tdf#81918
Dxf font export needs a lot of special handling to detect which parts
are set in the itemset and only export these. The old code only checked
for font name, font family and text encoding and exported the whole font
as dxf font in that case. The new approach just exports the necessary
bits.

Change-Id: Ic9a9c547723a65f26ebb6c88040a51f04f3cc42b