tdf#71278 Revert both the previous cursor commit and the commit it fixed
The original commit was based on a misunderstanding. In fact, the call to
GetSalData()->getCursor() at the end of AquaSalFrame::getCurrentCursor()
already handles POINTER_NULL correctly, or at least in the same way as on
other platforms. The cursor in presentation mode disappears after 10
seconds - if this is to be changed, it should be changed for all platforms.

Reverts:

  commit 58a3a99efec9d4bcb4ab61c41bf2feb707c8e1ef
  Author: Matthew Francis <mjay.francis@gmail.com>
  Date:   Sun Apr 12 23:10:58 2015 +0800

      tdf#71278 Set an invisible cursor rather than hiding the cursor

  commit 8d351c59d4393aa37ab38139b3a741f15a49dbfe
  Author: jorendc <joren.libreoffice@telenet.be>
  Date:   Sun Apr 14 00:33:19 2013 +0200

      fdo#31082 cursor visible during impress presentation

Change-Id: Ie40882661dc1a2f4857027ff04393874042e04ec
Reviewed-on: https://gerrit.libreoffice.org/15316
Tested-by: Norbert Thiebaud <nthiebaud@gmail.com>
Reviewed-by: Norbert Thiebaud <nthiebaud@gmail.com>