StyleManager & StylePreviewRenderer + common implementation
StyleManager is a new class responsible to handle document styles.
The current implementation has only the responsibility to provide
a implementation of StylePreviewReneder, but the idea is to move
all style related functionallity to this class (and relieve the
SfxObjectShell a bit). CommonStyleMAnager is the common impl.

StylePreviewRenderer is responsible to render a preview of a style.
As styles can be handled differently by each component it is
possible to provide component specific StylePreviewRendered, but
currently only the common one is implemented and used by sw.

Change-Id: Id271485f571a777a3e94f855d3c614a2efc14628