fix re-export of novell590442-1.ppt to ppt
we always close this container, even if we didn't open it

original commit always opened the SpgrContainer, and always
closed the SpgrContainer. (but seems to have a loophole, presumably
never hit where the SpContainer could be opened and not closed)

make a container guard which opens in ctor and closes in dtor.

should make ppt export crash stats hit 0

Change-Id: I2aead7397448b674e433a4097c97285067a6dc6e