tdf#38693: Document property TotalTime misinterpreted as seconds
Problem Description :
The xlsx file has the extended document property <TotalTime>#n</TotalTime> in /docProps/app.xml
LibreOffice appears to read this as a a total editing time of in second_ but according to ECMA-376 2nd edition:
22.2.2.27 TotalTime (Total Edit Time Metadata Element)
Total time that a document has been edited.

The default time unit is minutes.
(ECMA-376 1st edition has the same statement in  7.2.2.27)

XML difference: None
After RT : None

Solution :
As LO uses the TotalTime in seconds, while importing the TotalTime,
the TotalTime mentioned in the app.xml [which is in minutes] converted it to seconds.

TODO: Writing a UT seems to be tricky for this change. 
Need to analyze and will raise the UT in separate CL.

Change-Id: Icbe998b9881ac241a9147628573bab6ef63582c4
Reviewed-on: https://gerrit.libreoffice.org/15479
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>