ImpEditView::SetEditSelection: fix missing tiled rendering selection callbacks
Commit 2f31ac76be64953d19a72e3d8eae42b1ac5c5473 (editeng tiled
rendering: allow resetting selection with a single click, 2015-03-16)
changed ImpEditView::DrawSelection() so that the selection is only
emitted when the view is in selection mode. However, this way selections
set by ImpEditEngine::Search() were lost, as it only set the selection, but not
the selection mode. So when in tiled rendering, always adjust selection
mode depending on if the selection is empty or not, that makes the
search result selections show up.

Change-Id: I00b0739c5263a98a437e38d01b73c5004e50957b