sw: fix SwAccessibleParagraph crash on delete
It happens that SwAccessibleParagraph::getCharacterBounds() is called
from ATK with a clearly invalid pPortionData member; the SwTxtNode has
no text but there are portions.

This is because after a deletion both a POS_CHANGED and a
INVALID_CONTENT event are created for the same SwAccessibleParagraph,
and they are merged into one POS_CHANGED event by
SwAccessibleMap::AppendEvent(), but
SwAccessibleContext::InvalidatePosOrSize() returns early if the
paragraph happens to not be visible, skipping the invalidation.

(regression from 76c549eb01dcb7b5bf28a271ce00e386f3d388ba, which
 removed a Dispose() call; not sure if the Dispose() is needed here)

Change-Id: I985e9a439ee6c7024963eace876186f2247b9e03