Get rid of the initial :: for the vcl namespace
We used it all over the place without leading :: already anyway, even
in many files in include. So let's be consistent. In the majority of
cases, prefer the easier-on-the-eyes choice, not the "safe" one.

In those files in include where *all* existing uses of ::vcl:: indeed
used the :: prefix, there let's keep it for consistency.

Change-Id: If99cb41d3bf290d38c601d91125c3c8d935e61d0