tdf#34555 add crop features to svx
Adds crop feature to SdrObject. In EndSdrDrag related to Crop, there is
a new branch. If object is a regular SdrGrafObj (known inside svx), it
is used. Else, a virtual method on the object is used. This enables to
forward End of crop action to SwVirtFlyDrawObj objects (when you crop
with handles in writer).
Regarding writer, coordinates based on Twip/MM100 are used, not the
matrix based one.

This is part of a serie of 4 patches that adds Save graphic, Change Picture,
Edit with external tool, Crop (by handles) in all products (scalc,
sdraw, simpress, swriter).
Main menus, toolbars and contextual menus are updated accordingly.


Change-Id: Ie1a133d18487f51bb9c44ae2f000e63d911bf6b3
Reviewed-on: https://gerrit.libreoffice.org/15588
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Tested-by: Jan Holesovsky <kendy@collabora.com>