Use a dummy icu-config when cross-compiling harfbuzz (to Android)
Seems that when cross-compiling to Android from Linux, we apparently
have used either the build platform's pkg-config files for ICU, or the
build platform's icu-config. Both of which are obviously the wrong
thing to do, but apparently it has worked by accident anyway.

This makes building for Android on OS X proceed past harfbuzz, at least.

Change-Id: I27351f6177438697a1cded642c8c669ba7221009