sw: fix crash due to redlines on tables on ooo121112-2.docx
Problem is that after import there are SwRangeRedline that start in the
first cell of a table and end in the paragraph following the table.

There are <w:del> elements covering every individual paragraph in the
table; all of these are merged into one SwRangeRedline.

This could possibly be fixed in writerfilter by buffering the
m_pParaMarkerRedline until after convertToTable() to prevent the
merging, but perhaps it's better to fix it in SwXText::convertToTable().

Change-Id: I853ae624fffedb59a48bd90decb0973bf33beb68