VclPtr: clicking on sidebar impress navigator crashes
not sure how this ever worked, the callback is called
at the end of the ctor and the callback is in the parent
obj which dereferences the pointer to this obj before
that pointer has been set.

Just split the ctor into a ctor and a follow up
post ctor init method.

Change-Id: I04855c8ac95592801f2ebf0b974a9a7b597d0327