sw: fix assert on loading fdo48559-1.rtf
SwXText::convertToTable() is called with 8 cursors that all point to the
same empty paragraph; that causes the paragraph to be split but the
SwNodeIndex always point to the 2nd node after the split, so they all
end up pointing to the last node.  So adjust the last cell's SwNodeIndex
after split.

Change-Id: I0fdcb19baf2fae840177fd621666c10f4f6f2a62