SwEditWin::SetCursorTwipPosition: support creating a selection
The Android LOK client always creates a text selection by double
clicking on a word, and then the start/end of the selection can be
adjusted using handles.  In the GTK LOK client, it makes sense to allow
the desktop-style selection, where you click somewhere, move the mouse
and finally release the mouse to create a selection. That can be mapped
to settextselect-reset on mouse-down, and settextselect-end on
mouse-move easily.

The only problem was that SetCursorTwipPosition() assumed that there is
a selection already -- fix that by  adding the missing Stt/EndSelect()
calls and limiting the lifetime of the SwMvContext instance.

Change-Id: Iaeeadd8e4d9030614ee069b9fcfa269ce74ed58a