tdf#68868: better handle xlsx with samecell anchor image
Problem Description :
In the bugzill attached file, a image is anchored to the cell,
it is shown as stretched image after opening in the LO.
The image anchor attribute <xdr:from> and <xdr:to> are having the
same value. It means <xdr:col>, <xdr:colOff>, <xdr:row>, <xdr:rowOff>
tag values of from and to are the same.

So this image will have width and height as 1.

Although, this is very rare case, where from and to has the same
attribute values. Eventhough, we tried to create such kind of EXCEL
file but failed to do so.

Solution :
When twocellAnchor attributes of from and to has the same value, then
hide the shape.
If we avoid to add the shape, then the original image file will be
shown as stretched. If we don't add it into the shape, the original
image will not be preserved.

Change-Id: I48ea1cffcdfb013cec115ef8df5401e532fd2efd