tdf#79639 DOCX import: don't delay text frame conversion of in-header tables
Floating tables may or may not be converted to table-in-textframes
during import, depending on if we guess that it'll be a multi-page table
with minimal wrapping or a real wrapped table.

If the floating table is in a header or footer, then it won't be a
multi-page one, so can do the conversion right away.

Change-Id: I8d5ff8c5fe00037d5cef92dea6b54de6806214bc