sw: LOK_CALLBACK_TEXT_SELECTION should be the union of all selections
E.g. if searching for a keyword and it's inside a text frame, then we
have two cursors: one is an empty selection at the anchor point, and the
other is the real selection. What happened is that we emitted two events
for the two cursors, instead of merging them together.

Fix the problem by not emitting the events in SwSelPaintRects::Show(),
instead do it at once in SwShellCrsr::Show().

Change-Id: Ie2c7691aaaea7ba8a32b5cfa718a45cba571f791