SwDBManager: fix relative reference of embedded data source
When the MM wizard creates an embedded data source definition, then we
used to always write absolute references to e.g. the data source Calc
file. Pass BaseURI to the filter, so in case the document and the data
source is in the same directory, relative references can be written.

Change-Id: Iaa316adaf435f82bb574c7e5085ac04bf13a71f8