Resolves: tdf#89104 Missing actions from toolbar
Toolboxes derive from Docking Windows, they are
not layout aware, which is ok, but they need
to tell the thing they are inserted in that their
state has changed if the parent is layout aware.

i.e. docking windows are not toplevel things like
dialogs and need to pass on queue_resize to things
above them

Change-Id: If0eeff90314bef4e828355661a75a0d34a6fc1bc