sw: fix assert on exporting ooo24533-1.sxw to WW8
There's a fly anchored at a paragraph inside of a table inside of a
redline, and for some reason MoveRange() calls SwTableNode::MakeFrms(),
which asserts in frmtool.cxx AppendObjs().

The problem is that _SaveFlyInRange() removes the frame format from the
SpzFrmFmts but it is still reachable from the SwNode::m_pAnchoredFlys.

Change-Id: I3853c0072662bfb4f71f78451d3604ed567cd584