tdf#91001 Fix inc/dec font for complex selections
By "complex selection" I mean:
1. Selection with an empty paragraph that has a bullet.
2. Selection that has different font sizes.

This is a regression of
5b32d9c4406a68507ac9737ee0ada8bd1d424815 that tried to
fix tdf#61492. In reality in order to fix that, it seems
to be enough to pass an empty SfxItemSet, because what we
really need is to invalidate the bullet width (by setting
it to -1), thus force recalculation, and this occurs
anyway.

Change-Id: I92545192f8c78ef7f1d66d72f0d8e8f256f87604