Removed singleton PropertyNameSupplier and replaced it with single function.
Measurements showed that the "optimisation" of caching PropertyIds and their
string equivalent leads to an increase of approx. 6 times in the total overall
time spent in PropertyNameSupplier::getName(eId), when running the unit tests.

PropertyNameSupplier was the only PropertyNameSupplier (no derived classes).
This means that getPropertyName can easily provide the same functionality.

Change-Id: I933b67c11d4cc35395a0c70e15f1c24ac9842ab0
Reviewed-on: https://gerrit.libreoffice.org/16665
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noelgrandin@gmail.com>