tdf#91378: sax, oox: avoid sending empty strings to character callbacks
This reverts the changes in FastSaxParserImpl from commit
16e8ffbd5ec1fe7b81835ea6584547669d55d751 and instead fixes the problem
of inserting string properties with empty value locally in
OOXMLDocPropHandler.

This change was not wrong in any obvious way, but it turns out there is
one doc rhbz583386-4.docx that, when imported with this change and
exported to DOCX again, results in a non-well-formed document because of
some weird SDT stuff.

That problem is rather baffling, but unfortunately the
DocxAttributeOutput usage of FastSaxSerializer::mark() makes the DOCX
export rather un-debuggable, so avoid that problem by reverting the import
change for now.

Change-Id: I0d874cbfe82d4f15d58b50116dda152341bdf7b0