Related: tdf#92645 svx rendercontext: avoid direct paint on ending text edit
If the Paint() is ~immediate after Invalidate() (which is the case
since the main loop has priorities), it makes no sense to
perform any kind of direct paint right after an Invalidate().

With this, we no longer hit the is_double_buffered_window() assert when
returning to editing Writer text after editing draw shape text.

Change-Id: I3b80ca08aa71d247fc0561133f907aef34de4001