Spell-check wrong-dirty text upon showing context menu.
This makes for a better user experience when the idle jobs
haven't yet ran on some text to check for spelling.
This can happen when the user is on a device with insufficient
compute power and/or other idle jobs with higher-priority
take precedence.

This change leap-frogs the spell-checking idle job when the
user might already know they mistyped a word and look for a
quick fix via the context menu.

Change-Id: Id1f7b4555050ded329ebeb56502f893ee7b2bc35
Reviewed-on: https://gerrit.libreoffice.org/17252
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Norbert Thiebaud <nthiebaud@gmail.com>