form document view activation: prioritise activation of already active form
This avoids arbitrarily switching to the first form in the document,
which would do a (premature!) save to the database of the
modifications pending in the active form. This may lead to a database
error, when the data is not in a shape to be written to the database,
e.g. when on an insertion row and not all mandatory fields have been
filled in. This then pops up an error message to the user.

Change-Id: I30bb533598ca707b892bb7155e54ce05d4ddf275