tdf#92911 SwAnnotationWin rendercontext: fix missing comment contents
The problem was that when inserting a post-it comment, the Paint() is
triggered for SwEditWin, which then shares a RenderContext with its
children, and at the end paints the buffer to the vcl::Window of
SwEditWin. When painting this buffer, the contents of SwAnnotationWin
was missing, as the parent SwEditWin was not allowed to paint on that
area. No explicit clipping is set at the OutputDevice level, but
vcl::Window::ImplClipChildren() still calculates a clip region based on
the child windows for the underlying SalGraphics, and that prevented
SwEditWin to paint its buffer fully.

Fix the problem by explicitly disabling this clipping for
SwAnnotationWin in its ctor in the rendercontext case.

Change-Id: I5eb47c20070057a79297d91242624fc2057e2dfb