tdf#93052 Fix abundance of calc annotation styles
This is a regression introduced by fixing tdf#47018, which
changed the default NumRule and pushed a lot of SVX_MAX_NUM
to the SvxNumRule constructors.

The changed NumRule default seems not needed to fix the crash;
at least it still doesn't crash as described in the old bug
report after reverting this part of the change.

Change-Id: I908eeb4a827799419d3f28584fdcbc9272cbc76d
Reviewed-on: https://gerrit.libreoffice.org/17527
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Norbert Thiebaud <nthiebaud@gmail.com>