tdf#90222: Removed ScaFuncDataList type
This type was a special case of a custom implemented list, and was used
like a map. I have replaced it with a std::map typedef'd as
ScaFuncDataMap.

Since the map key also exists in the value object, there are more
memory-efficient ways to implement this with a vector and a predicate
functor for searching at the cost of some code complexity. I opted for
the simpler code afforded by using a std::map implementation.

Change-Id: Idd450f4a908ff503cb02e8e363ddbe3512903f47
Reviewed-on: https://gerrit.libreoffice.org/17579
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noelgrandin@gmail.com>