tdf#93154: Save button often does not save
The problem is that in tiled rendering LO doesn't always consider the
document to be modified. Some operations such as Bold, Italic, etc do
not mark the document as being modified, but we need to be able to save
the changes. This solved the issue by always allowing to save.

Change-Id: Iaf8120abadc768a07b24f42c287c50080b4e4d89
Reviewed-on: https://gerrit.libreoffice.org/17665
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Tested-by: Jan Holesovsky <kendy@collabora.com>