do not parse range fragments as valid ranges
For example, "B2:" or ":B2" were accepted as valid range, with the
missing part set to A1, which resulted in a A1:B2 range.

An omitted sheet specifier always results in a valid relative sheet,
hence that flag is set, so testing flags for non-null before the final
result is available isn't sufficient.

Change-Id: I57d1b7cf71a298e7ac108b0e13aab88208f0f00c