Revert "exclude unnamed Library/Executable from ide integration"
This reverts commit 93cd7b78c29c11ccc87c19c845c6617acb834630.

We need to generate project files for unnamed projects as they seem to
correspond to libraries. If that causes problems for the msvc
integration (which seems strange) then we need to handle these problems
in the MSVC generator and not in the parser.

It also includes:

Revert "Clean trailing whitespace"

This reverts commit 7d92b27aff3b4d1c59aef9866003f4d001c0fc25.