Android: check file modify date to know if save is complete
The API does not allow to set a callback for the save operation, we
work this limitation around by checking the modification date of the
local file periodically. When that date changes, we are sure the
local save operation is complete and we can invoke the document
provider save operation to push the changes to the cloud, if
necessary.

Users may press "save" on a document with no changes, in this case we
have set a 20 seconds limit to stop checking the modification date.

We also add a "save complete" message for the user.

Change-Id: Ib8871fac682a5c03a187a7238e11874984143527