tdf#94237 tiled rendering: Use the entire document as the visual area.
Many places were already adapted for the tiled rendering, and the conditions
checking for the VisArea() were changed so that the checks are ignored when
tiled rendering is active.

Unfortunately there are still some places left, and they are causing that
various areas are invalidated only partially.  Let's sort it out for good, and
report the entire document as the VisArea() when the tiled rendering is
active, and if there are performance problems, let's profile that & act
accordingly.

Change-Id: I53f18915ed0aec898704dd1350a9534cfc3f001b