Don't violate unordered container hash+pred requirements
"Two values k1 and k2 of type Key are considered equivalent if the container's
key equality predicate returns true when passed those values.  If k1 and k2 are
equivalent, the container's hash function shall return the same value for both."
[unord.req]

But for INetContentTypeParameterList, all inserted keys are already lower-case
by construction, and all queries happen to be done with (literal) lower-case
keys, so just document the requirement here that insertion and query needs to be
done with lower-case keys.

Change-Id: Ic46845b7bb0376e995e7736b57fa80c777d1c263