sw table styles: Implement table styles in Writer.
This extends the table auto formats so that SwDoc keeps track of the auto
formats used in the document.  With this in mind, we can update the format of
the table with every operation like adding/removing a line, splitting a table,
etc.

So far we only have the core functionality, and cover inserting a line in
the table; more to come.

Based on work of Alex Ivan <alexnivan@yahoo.com> during GSoC 2013 - thank you!

Change-Id: I7839147e54c2f976988121a523331def9859f4c2