correctly dispose the chart sidebar when leaving chart, tdf#94292
This is a much better approach than my old one that tried to use the
correct XModel with the chart sidebar. Now everytime we leave the chart
edit mdoe we destroy the chart sidebar panels and recreate them when we
go back to chart edit mode.
As a result we no longer need to have worry that much about using the
correct ChartModel and about the lifecycle of the ChartModel.

Change-Id: Ie85ad82968e20a81f059d4bd930a83a70da35fc9