tdf#94924: Add a more systematic OpenCL unit test
Avoid the horrible convention of hard-coding in a C++ unit test code
addresses of data in the spreadsheet document being tested. Instead,
mark the expected (= as calculated by Excel) and calculated (by
LibreOffice) formula results, rectangular blocks of data, so that the
C++ code can easily find it, and then compare. This is much more
flexible. No need to edit hardoded row and column numbers in the C++
code when adding more test data.

The systematic.xls file has documentation on how to maintain it.

Change-Id: I4fb088fe21831dd3b3213d21916460a708aa0842