Updated core Project: help  3f2b535c2dea86ec6c01d83dc37fbb56d9c4e999
Correction to licensing of the patch for ODFF.xhp

After introducing ODFF.xhp using an old version of helpauthoring extension the wrong license
was introduced in the xhp file. Further corrections in the sequence still left mention of AL
where it should not. This patch removes references to AL licence.

Sequence:
2015-09-02      Olivier Hallot  Revert "Remove metadata from embeded xhp files" 80/18280/2
2015-09-02      Olivier Hallot  Remove metadata from embeded xhp files 55/18255/2
2015-07-09      Olivier Hallot  Fix license header in these 2 files 70/16870/2
2015-07-08      Olivier Hallot  Promote ODF Formula standard in Calc function help 10/16110/2

Change-Id: I16236838423cc35e8939d65d77ad14d5291a1b5d
Reviewed-on: https://gerrit.libreoffice.org/19533
Reviewed-by: Olivier Hallot <olivier.hallot@edx.srv.br>
Tested-by: Olivier Hallot <olivier.hallot@edx.srv.br>