Updated core Project: help  3397ffc09895d62067538eafcfef2858f0e42522
Correction to licensing of the patch for func_aggregate.xhp

After introducing func_aggregate.xhp using an old version of helpauthoring
extension the wrong license was introduced in the xhp file.
Further corrections in the sequence still left mention of AL
where it should not. This patch removes references to AL licence.

Sequence:
2015-09-29 tagezi tdf#85228 Added description of AGGREGATE function 21/18821/7

Change-Id: I28036684bf3ca5a942a82c6ec9295e0eb427e36b
Reviewed-on: https://gerrit.libreoffice.org/19535
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>
Tested-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>