Let Gtk handle icon names instead of ResMgr for window and trayicon
The icons in the themes should be deleted in a second step and the defines
should be transformed into an ENUM and stored in one place.
Furthermore we will have to decide whether to live with gtkimagemenuitem
despite its deprecation in Gtk3.10 or whether to pack GtkImages and
GtkLabels in GtkBoxes inside GtkMenuItems and live with the additional
padding, which potentially looks horrible.

Change-Id: I8dec44375e40a4d2b6668792bc4b813efe83ff31
Reviewed-on: https://gerrit.libreoffice.org/19601
Reviewed-by: Björn Michaelsen <bjoern.michaelsen@canonical.com>
Tested-by: Björn Michaelsen <bjoern.michaelsen@canonical.com>