tdf#93837 Add a new PopupLabel property
Some commands in context menus use different labels than in
the menu bar. In the old resource format those labels are
defined in the resource file itself. Note that the menu xml
has the menu:label attribute, but we can't use it because it
lacks localization support, and as such useful only for user
customization.

The order of consideration now is:

1. Label - Used if Context/PopupLabel isn't defined. (It's
   also used as the base for toolbar tooltips.)

2. ContextLabel - Overrides Label when in menu or context menu.
   Useful to drop the context, e.g. "Insert Image" should be only
   "Image" under "Insert" menu. If exists, that's what returned
   when asking for the "Label" property, instead of the full label.

3. PopupLabel - Overrides Label/ContextLabel when in context menu.
   Used to give items in a context menu different labels than in
   the main menu, e.g. "Paste Special" from the main menu turns to
   "More Options" when in context menu. It doesn't affect the Label
   property, and should be asked separately.

Change-Id: I7408fc2bfb8d384b0f1a72a78b8c5d7c50de38da