tdf#95356 Now text decorations are applied correctly
Now the default for using the SVG tiny profile is false.
Now the default for use native text decoration is true.

Fixed the following issues:

1) Text decoration (underline, strike-out) was not exported.

2) Commit 0a4e913 (tdf#56467: improve export of formulas to SVG) had a
side effect: all exported text portions have position attributes (x,y)
even if they are on the same line (y_old == y_new): this caused that
underlined text belonging to 2 different `tspan` element was not
underlined by a continuous line but there was a gap where a `tspan`
ended and the next `tspan` started.

3) In commit (tdf#37650: further improvement of svg export)
StartTextShape method is invoked even when positioned characters are
used but the `implWriteText` method always adds a `text` element by
itself so there is no need to call StartTextShape.
The changes for non-positioned characters have been embedded inside the
`WriteTextPortion` method.

Change-Id: Iee0d6a6816dcbd2d97f2fbdf969f012de678604e
Reviewed-on: https://gerrit.libreoffice.org/19672
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Jan Holesovsky <kendy@collabora.com>