sc lok: fix rounding errors with non-100% zoom
There were two problems here:

1) ScTabView::getRowColumnHeaders() did not expose twip values directly,
but used ScRow/ColBar::GetEntrySize(), which does a twip -> pixel
conversion, and then converted it back to twip. Avoid this unnecessary
roundtrip.

2) ScViewData::ToPixel() trunaces the resulting float to an integer, so
if the result is e.g. 67.7 pixels, then Calc handled that as 67, but
gtktiledviewer rounded that up to 68, resulting in non-matching headers
for the rendered tiles.

Change-Id: Ie6ed1ea923a423d1526eeb235b7b87106fd2f20b