tdf#36946: Organise transitions hierarchically
Introduce the concept of transition groups and sets. (Suggestions for
other terminology welcome.) A collection of transitions that differ
only in the direction the transition is applied, or similar minor
fashion, are called a set, and they show up only once in the long list
of transitions in the UI. Each set also has an icon.

Sets are then collected into groups based on some overall common
feature, like "subtle" or "3D". Groups do not yet show up in the UI,
but are only present in the configuration registry.

I made only a few silly icons as I am not an artist. Not intended to
be a final design in any way for them.

Change-Id: I148cb7f8dc2e3ecd70cae188908dd02053308239
Reviewed-on: https://gerrit.libreoffice.org/19797
Reviewed-by: Tor Lillqvist <tml@collabora.com>
Tested-by: Tor Lillqvist <tml@collabora.com>