oox: check for namespace in MCE is flawed, use a namespace list
In "AlternateContent" nodes we have to check if we support the
namespace which is provided in "Requires" attribute of "Choice".
Currently we tried to resolve the namespace with a call to the
xml filter, however this doesn't work as the filter is already
gone.
In writerfilter we also have to handle a similar situation but
there we just compare it to a list of predefined namespace alias
("wps" and "wpg").
This commit adds a list of supported namespace aliases to
fragmenthandler2 instead of the namespace checking to support
the "p14" namespace alias correctly.

Change-Id: I25c430b97336c9e140bb5641a76a60895734b91f