Use GLyphy for text on Windows (broken)
Work in progress. Produces visible text in some cases, but not nearly
all. The text that is visible is roughly in the right place and of the
right size, but horribly ugly. Compare to glyphy-demo, which does
produce beautiful text, so the problem is not in the GLyphy code but
in the way we use it.

Include sources from GLyphy's "demo" directory (with only slight
modifications to avoid unconditional debug output) and use that "demo"
API from vcl.

The changes to existing vcl code are all in one place, in the
winlayout.cxx file.

Change-Id: I69cce5d66db534c6f4c1ab85d520b6090baf8fe0