sw lok comments: implement painting of the vertical scrollbar
This one is special, as normally its map mode is in pixels, but we need
all sub-widgets to work in twips when tiled rendering.

With this, the scrollbar widget (both the buttons and the
button/background area of the scrollbar itself) is painted at the
correct location when Writer comments have enough content so the
scrollbar is visible.

Change-Id: I4ee9ef8618974b965339078d2262364ec19732ef