vcl: simplify Image internals
Image could be of 2 types - BITMAP or IMAGE, where BITMAP used to
store the content in a Bitmap and IMAGE in a ImplImageData, which
contained a BitmapEx. This was refactored with this commit to
always store the content in a BitmapEx and there are no distinct
image types anymore. This greatly simplfies the code.

Drawing of the image in case of type IMAGE was done in the class
ImplImageBmp which also modified the image according to
DrawImageFlags (for example to create a "disabled" image). This
was moved to ImplImage and the bitmap manipulation code was moved
to BitmapProcessor (done in previous commits).

Change-Id: Iec9f63a7c05618c457d8465f1ec60ed4f16bd579