tdf#96046 SVG: Parse 'em' and 'ex' units correctly
the problem is that real_p parses the 'e' in 'em' and 'ex' as the exponential 
operator instead of as a str.
Change it to parse sequencies like: (11) (1.1) (1,1) (.1) (,1) (1.) (1,)
but now the parser returns a std::string so we need to convert it to double,
and make sure the commas are converted to dots too.

Change-Id: I913be7975e7db1be4773f528fd0d586db67c9461
Reviewed-on: https://gerrit.libreoffice.org/20181
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>