tdf#34957 - don't push multi-page table to next page.
Removing an old hack that is no longer needed.

A table marked as “keep with next” doesn't split and keeps with the
following paragraph already.  It doesn't require this old hack to
function properly anymore.

Preventing tables from splitting had forced
large tables to always start on a new page.  Removing the hack
allows everything to lay out nicely now. There is a fairly large
suite of tests in the bug report that prove the proper layout of
various keep-with-next situations.

Any document depending on the old behaviour can simply  (and properly)
insert a manual page-break before the table in order to force it to start
on a new page.

Change-Id: Ie2036c1a638009aaf95247574998077a1e55b5e5
Reviewed-on: https://gerrit.libreoffice.org/20216
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>