tdf#91083 - .doc: emulate table keep-with-next paragraph
connect table with the following paragraph.  Move forward to a new page
together with the following paragraph if necessary.

Most of the added code allows the table to split if
all of the kept items do not fit on one page - allowing
the emulated table to match the layout of kept tables.

The only difference with .doc occurs when the table itself is larger
than one page.  In that case it ALWAYS starts a new page.  Although
it does not match .odt, it DOES match how MSWord handles that .doc.

Change-Id: Ic6f495c0dc5cd4e9f8029a8cec9e31b4fab4b20f
Reviewed-on: https://gerrit.libreoffice.org/20233
Reviewed-by: Justin Luth <justin_luth@sil.org>
Tested-by: Justin Luth <justin_luth@sil.org>