sc interpreter: Don't throw when we hit something not yet implemented.
Instead create a ScFullMatrix, and perform the operation using that.

This has a performance penalty, as we need to copy the data from
ScVectorRefMatrix to the ScFullMatrix; but this is one-off, and the same
performance penalty that was there before introducing ScVectorRefMatrix.

Additionally now we can optimize the methods that appear to be the
bottlenecks one after another (and whitelist them).

Change-Id: If94a5017b96fb914817cfd7ed1b7ae823db8b5f5