DOCX import: initialize RDF metadata before importing document properties
With this, it is possible to import part of the document as RDF
statements later when parsing document.xml.

Previously SfxBaseModel member functions like
getMetadataGraphsWithType() and addMetadataFile() failed, as they tried
to find the already imported document in UCB, which failed, as the
import was still in progress. To prevent that, do the same as the ODT
import in XMLReader::Read(), part "RDF metadata".

Change-Id: Ia15af9bb0681cbc25fe745742c24af82cd213e5d