Remove image color conversion menu from Impress
In impress there is (or was) a submenu "Mode" for images that
containted color transforms for images to various bitdepths and
palettes. The action is destructable - which means the original
is replaced by a new image. This also has little use today as I
don't think many people want to shrink image (byte) size this way.
The only useful transform is to 8-bit indexed for certain types
of images to save space without a bigger color degradation, but it
makes more sense to add this option to "compress image dialog".

Change-Id: I7450b2ec3507fb5d180cbd1ec879cb5560a2dd22