related to tdf#73691 - AltX: use alt-c for eu, gl, nl, pt-BR
Unicode Notation toggle's alt-X conflicts with the main menu for Basque,
Galician, Dutch, and Brazilian Portugese.  Replaced with Alt-C which matches
the alternate combination already assigned for German.

If MSWord uses a different combination for these languages, someone can
request a matching change.  Dutch is confirmed as using the conflicting Alt-X,
but the other languages are unknown status in MSWord.

Change-Id: Id07fc825aa7df427b4e7d1db9026dc1665a14a0d
Reviewed-on: https://gerrit.libreoffice.org/20370
Reviewed-by: Samuel Mehrbrodt <Samuel.Mehrbrodt@cib.de>
Tested-by: Samuel Mehrbrodt <Samuel.Mehrbrodt@cib.de>