sw: DOCX export: convert ODF embedded objects to OOXML
If the user edits an embedded object it is converted to ODF, so we
really need to be able to store such objects.

Ensure that the proper MediaType is set in [Content_Types].xml,
the package relationship type in document.xml.rels and
the proper ProgID attribute in document.xml.

Change-Id: I3c78c5ab5b4d534213af5e773fe0c6c2c92d9104