geez, how about actually checking the Move() error return? tdf#92779 related
Handle failure condition where we know how to treat it, i.e. when
updating references, assert in all other places that so far silently
ignored it and implicitly assumed the failing Move() truncating at
bounds would be alright. In case we'll encounter an assertion we'll have
to inspect those places and decide what to do about it.

Noticed this error with a reference like B1048575 and inserting two rows
above, it became B1048576 instead of B#REF!

Change-Id: I00757f3ed2e305b591178047933ed60f1533317e