Get rid of ImplDoTiledRendering()
It returned true only on iOS, and we do use tiled rendering
successfully also on the other non-DESKTOP platform, Android, plus
obviously for LibreOffice Online and other LibreOfficeKit clients, so
it is quite likely that the function was not really needed. Besides,
the code doesn't even build for iOS at the moment.

Change-Id: I229bce61eaab9b795a1d5e959cb6587dcedf1ef1