crashtesting: crash on converting ooo93489-1.ods to pdf
with --headless --convert-to pdf ooo93489-1.ods

probable regression from...

commit d4daad185e9583bedbb5a4eef1fd53e1f22e219b
Author: Jan Holesovsky <kendy@collabora.com>
Date:   Mon Nov 30 10:28:43 2015 +0100

    sc interpreter: Move the code that can create a ScFullMatrix.

In that original code I see there was a check for "Data array is shorter than
the row size of the reference. Truncate it to the data" which doesn't exist
anymore. If I reintroduce that check here the crash is avoided.

Change-Id: I7c7a5979d9c14c133b05e89ce3794e6b739ca61c
Reviewed-on: https://gerrit.libreoffice.org/20794
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Eike Rathke <erack@redhat.com>