tdf#93837 Convert sheet row/col header context menus to xml
This commit shows some weakness of the current approach with
PopupLabel, as in many cases it's exactly the same as Label.
Have some ideas on how to improve that, but will look at it
after converting the remaining menus (at least in sfx2 based
modules), so I could have a broader look at what we have.

Change-Id: I78dfdb4a6f2ea3faf84f3a521067c6c15df96852