toolkit: mark AccessibleScrollBar:XAccessibleEventBroadcaster as a known issue
JunitTest_toolkit_unoapi_1 started to fail for me recently in
toolkit.AccessibleScrollBar, though only when ran via 'make check' or
'make toolkit.subsequentcheck', not via 'make
JunitTest_toolkit_unoapi_1' when other junit tests of toolkit are not
running.

Provided that, it's quite hard to debug the exact problem, but it seems
VCLXAccessibleScrollBar doesn't implement XAccessibleEventBroadcaster
itself, but via comphelper::OAccessibleContextHelper, which perhaps
explains why 4 other components have their XAccessibleEventBroadcaster
interface tests disabled already.

In any case, disable the XAccessibleEventBroadcaster test for
AccessibleScrollBar till it's clear if it's the test that is poor or
if there is a real problem here.

Change-Id: I2de618c11e445e4e6111c7ae0d1d8feef11fd9d3