crashtesting: CreateShellID returns the mediums base url...
when it works the way we want it to work, so presumably we can cut out the
middle man and use getDocumentBaseURL directly which does the same thing.

This has the advantage of being in the IEmbeddedHelper api so assignFrom can be
tweaked to pull the src and dest ids arguments to CopyAndGetEmbeddedObject from
the src and dest IEmbeddedHelpers always available there. So we ensure we
have some ids to present to CopyAndGetEmbeddedObject, avoiding the assert
on converting documents such as ooo89433-1.ppt to pdf

So the explicit args can apparently be dropped, bubbling up then means we don't
need a CloneWithShellIDs (introduced with tdf#71076 and tdf#71767) as normal
Clone should now do the right thing.

And various other arguments in various callers of these, which are typically
filled in with an unhelpful empty OUString placeholder, can go

Change-Id: I9c0be0340e72f98b10ee87ada7d2a021fcc39e23