vcl:  Make ImplFindByAttributes public and rename
Renamed PhysicalFontCollection::ImplFindByAttributes to
FindByAttributes.  This function is actually useful for finding a
font within the collection, so I'm moving it from being a protected
function to public because it is actually useful in its own right,
not just for PhysicalFontCollection.

Note that ImplFontAttrs is a mess anyway, so not going to touch that
(yet).

Change-Id: I884ddd5118cbbb0f92b157540f7ffcf1bf35609b
Reviewed-on: https://gerrit.libreoffice.org/20865
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>
Tested-by: Chris Sherlock <chris.sherlock79@gmail.com>