vcl: remove poor-man's RTTI in PhysicalFontFace
The *only* things relying on the in-built magic number are PDFFontCache
and PDFWriterImpl. PDFWriterImpl::emitBuiltinFont has as its first
parameter a const PhysicalFontFace*, but this is only called upon by
PDFWriterImpl::emitFonts, which instantiates an ImplPdfBuiltinFontData
object from a built-in font map... so basically it is always guaranteed
to get a ImplPdfBuiltinFontData object. ImplPdfBuiltinFontData derives
from PhysicalFontFace, so to prevent the need to check the class being
used is a ImplPdfBuiltinFontData, just make emitBuiltinFont require as
it's first parameter ImplPdfBuiltinFontData!

Unfortunately, we can't get around RTTI completely, because
PDFFontCache::FontIdentifier requires a way of distinguishing between
two different PhysicalFontFaces...

Change-Id: I74c8cd294580bbecd12dd2427ac405f97a7e7f7e
Reviewed-on: https://gerrit.libreoffice.org/21030
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>