the idea behind the test can't work right now
The Test currently tries to test the OpenCL and the SW interpreter and
assumes that it will show different results. However if we succeed in
generating an OpenCL interpreter we will never generate a SW interpreter.

There are a number of problems with the approach taken in this test.

1.) Assuming that we are able to execute OpenCL is an invalid
assumption.

2.) Changing the configuration inside of a test needs to be exeption
save as we are otherwise left with a strange config for the remaining
test cases.

3.) Changing the configuration does not change the formula group
interpreter. This needs to be done explicitly.

Change-Id: I679985ee7551b21aaa52038e107324980e7e1bb6