tdf#93565 Use multiple docs for individual PDF MM
There is actually a comment in the mail merge code:

  // convert fields to text if we are exporting to PDF
  // this prevents a second merge while updating the fields
  // in SwXTextDocument::getRendererCount()

This was included to fix i#93714.

But the reference to 2005-05-23 #122919# in the actual code
doesn't help to understand, why the original code is needed.
I would like to get rid of it, but currently don't have time
for tests.

So the optimization to skip ConvertFieldsToText() and use a
single working document for individual document mail merge
breaks for PDF, as the working document won't contain any
fields after the first document is written.
This uses multiple working documents for PDF export again.

Change-Id: I0cf1c64da27863fc5fb4bb6c7812fed90b737d13