vcl: ComboBox::ImplUserDrawHandler was in the void namespace...
I'm amazed that C++ allows you to do this, but I checked the draft
standard and there appears to be no restrictions around using a
C++ keyword as a namespace identifier!

This hasn't been an issue it appears because we don't have any
user-defined draw signals for combobox's currently.

This was picked up by Doxygen, which spat out the following error:

/home/chris/repos/libreoffice/vcl/source/control/combobox.cxx:1322 warning: no uniquely matching class member found for
  void::ComboBox::Impl::ImplUserDrawHandler(UserDrawEvent *pEvent)

Issue introduced in commit d6a83d3f913, I checked it and it was the
only place this occured in the patch.

Change-Id: Ie4272e5c4e3dba13ed82457bba8579971a5345d0