vcl: (Windows only) Changed ImplWinFontData name to WinFontFace
The base class name for WinFontFace is PhysicalFontFace, which gives
the actual font installed on Windows (this is eventually mapped and
realized into a LogicalFontInstance). However, PhysicalFontFace is
not just a data structure, it also holds the function CreateFontInstance
which must be implemented by the subclasses for each platform to allow
fonts to be used in LibreOffice.

Therefore, to keep the naming consistent I am following the lead of the
base class and renaming ImplWinFontData to WinFontFace (and removing
the "Impl" business, as this is not necessary).

Change-Id: I4debb3af3ab5067b7e12e3659f2c71aba91e848c
Reviewed-on: https://gerrit.libreoffice.org/21298
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>