vcl: Renamed ImplPdfBuiltinFontData to PdfBuiltinFontFace
Base class's name is PhysicalFontFace, which is used to create a
logical font instance. Each platform (including the PDF writer it
appears!) implements it's own factory function - so it's more than
just a "data" class.

I am also aiming to keep naming in the class heirachy consistent,
thus the reason for the name change.

Change-Id: Idb8c5b6a3b23a39ca727d355dcf421307cacbb2f
Reviewed-on: https://gerrit.libreoffice.org/21302
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>