vcl: resplit FontAttributes from ImplFontMetricData
The change I made in commit 7b974e056df3 ("vcl: merge
ImplFontMetricData with ImplFontAttributes") was... ill-advised. For
starters, there really needs to be this split as FontSelectPattern
needs it, and PhysicalFontFace only requires the font attributes, not
the metric data.

So the merge was unfortunately, in my mind, a failure and I'm manually
backing it out now.

Change-Id: Iac38f0815f984541e2f55099c965376dd88eeb43
Reviewed-on: https://gerrit.libreoffice.org/21380
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>