tdf#96522 reset CustomShapeEngine at import
Do not reset UNO API xShape at import time as described in tdf#93994, better
reset the involved XCustomShapeEngine which allocates and holds the
EditEngine and VirtualDevice. Resetting the UNO API object leads to problems
when e.g. an animation effect is set for the shape. This happens since the
animation effect remembers the xShape using a weak reference, thus being
no real on-demand constructable object.

Change-Id: I7b421f578ebf003af2745e449cd02ba368b455dd
Reviewed-on: https://gerrit.libreoffice.org/21401
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>