tdf#96385 - opengl: dynamically adjust priority of swap buffers.
Initially we start with a very low priority, so that the lame
bits of code that do eg. focus, and cursor rendering before the
document is visible do not cause a swap, flash. Then after we've
processed a REPAINT priority idle (hopefully our first paint) we
adjust the swap priority to highest.

Essentially a fusion of Tor's approach and mine.

Change-Id: Ib9b78a18fb9359ac8c3dee1cfeb30177c08fe162
Reviewed-on: https://gerrit.libreoffice.org/21405
Reviewed-by: Tor Lillqvist <tml@collabora.com>
Tested-by: Tor Lillqvist <tml@collabora.com>