editeng: handle SdrModel::isTiledSearching()
Given that the edit/outliner views can come and go, avoid the lifecycle
problems with just passing a pointer to the sdr model to editeng, and
then it'll always have the up to date "are we searching" information.

editeng can't depend on svx, so provide an interface class SdrModel can
implement.

Change-Id: I3b98011593b00ac0fab05b6b9c591dd20d94c579