vcl: add symbol and charset accessors and mutators to Font
Rules for the vcl::Font class for setting character set and
the symbol flag:

If the characterset changes to anything other than
RTL_TEXTENCODING_SYMBOL then the symbol flag should be off.

If the characterset changes to RTL_TEXTENCODING_SYMBOL then
the symbol flag should be on.

If the symbol flag is set to false and the characterset is
already RTL_TEXTENCODING_SYMBOL then set the characterset to
RTL_TEXTENCODING_DONTKNOW and set the symbol flag to false.
However, if we are setting the symbol flag from false to
false (i.e. we know the characterset) then we can keep the
characterset as it is.

Unit test written in this commit to test this is working.

Change-Id: Iced44659ab88ff66b711c560cb68bd4681ecb537