tdf#95292 MM: cleanup bEndOfDB handling
bEndOfDB was handled inconsistently in ToNextRecord for record
selections and normal record iteration. The new unified semantics
are defined to set bEndOfDB for the first failing record fetch.

This includes a rename of the function ExistsNextRecord, which
did never check for a next record, but just verified the current
record set to be valid.

This way we can also drop the bAfterSelection selection, which
was just needed for the inconsistent bEndOfDB handling.

Change-Id: Ibbd7ae95e5969b51e32f44b442e5dc778655e8a9