tdf#95376 DOCX import: fix incorrectly indented tab stops
Regression from commit f4badd9a485f32f787d78431ed673e2932973887
(tdf#92454 DOCX import: allow overriding para prop from num style in
para style, 2015-09-22), the problem was yet another priority
mishandling in the maze of various styles and indentation handling.

In the tdf#92454 bugdoc, both a numbering-from-paragraph-style and a
paragraph-style defined indentation, and Word preferred the numbering,
while Writer preferred the paragraph style, that's why the import-time
conversion was added.

However, it turns out there is a 3rd source that's still not direct
indentation formatting: a direct numbering. So the correct priority is:

direct-ind > ind-from-num > ind-from-parastyle > ind-from-num-from-parastyle

Which means in this case the indentation should not be set directly: the
two conflicting value (ind-from-num and ind-from-parastyle) will be
resolved correctly by Writer core.

Given that we always first get the para style info, and only then the
numbering, we just need to undo the conversion added for the other bug
if we see a direct numbering, then both the old bugdoc and this new one
will be handled properly.

Change-Id: I09cc84605d5df6159da985ad069d46b580a53358