tdf#83890 sw html export: fix missing start numbering attribute
In the sequence
1 Title 1
1.1 Subtitle 1.1
Not numbered item
1.2 Subitle 1.2
2 Title 2

The numbering element for the Subitle 1.2 was missing the html start attribute
when it was followed by a numbering element of lower level.
So let's force it, when it comes after a not numbered item.

Adding a test case that validate that the Subtitle now contains the right attribute

Removed a reference to an openoffice bug that brings no value too

Change-Id: I928592a667f59128a4aa079d6b042ffb14244181
Reviewed-on: https://gerrit.libreoffice.org/22191
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>