Autocompletion: complete year (part) to current date
to check or insert it during typing.

For example, typing "201" or "2016-0", Writer will
suggest "2016-01-30" - the current date in ISO 8601
format - in a tooltip, and it's possible to insert it
pressing enter.

There is no autocompletion for single years (e.g. "2016")
to avoid unintentional text insertion at line ending, e.g.
typing "30 January 2016" in a letter header.

Change-Id: I230a449b0e19d5316b8369d780d7c7fda57fea37
Reviewed-on: https://gerrit.libreoffice.org/21928
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Andras Timar <andras.timar@collabora.com>