Remove leftover printf-style %f
But seriously, who builds with OSL_DEBUG_LEVEL > 2?

Either debugging printouts inside such ifdefs are completely pointless
and should be removed, or then the information might actually be
useful, and the ifdefs should be removed, so that the information can
be selected to be printed using the normal SAL_LOG environment
variable mechanism (in a build where SAL_INFO is functional).

Change-Id: Ia5d2c7c7b525ed133fe3f077a005f583b8a856e7