sfx2: fix locking after signing OOXML documents
File -> digital signatures looks like a way to view and modify digital
signatures, internally it's a sign action, that at the end may not modify
signatures after all. For this to work, SfxObjectShell::ImplSign() calls
GetMedium()->CloseAndRelease() to release the document's lock file, invokes the
signing dialog, then DoSaveCompleted() creates the lockfile again.

When signing OOXML documents, the lock file is not re-created, as
DoSaveCompleted() only creates the lockfile in case
IsPackageStorageFormat_Impl() (== own format) is true. Fix this by adding a
mode that creates the lock file, even in case of a foreign format.

With this, closing the digital signatures dialog for OOXML documents no longer
results in a confusing "Document in Use" dialog after closing the signatures
dialog.

Change-Id: Ie9e56b88768825e61765669b27a89082cdc1981f