tdf#97700 - vcl: opengl - add asserts for horrible size mismatch.
We really need to be sure that our texture and its wrapper agree on
the size of the texture, and particularly the buffer it is reading
into to avoid DMA'ing junk over the heap.

Add paranoid assertions, also add a canary at the end of the texture
so we fail hard and fast in this case.

Change-Id: Ibf4869fb5cba562aa117943ce0f2f3df21ca7036