GetActiveAccessibleCell: remove recently obsoleted IF clause
Follow-up to commit 1fda8bf08123b521d9ad77eb542d7d1a3c63f54a

The recently added ELSE component is more inclusive than the IF component,
so there is no longer any need to check if a cell IsTextEditActive().
As long as the cell is active in any way, it will be returned.

Change-Id: I92cf2a231db1e1457f81a2ef77811ed40078e647
Reviewed-on: https://gerrit.libreoffice.org/22262
Reviewed-by: Justin Luth <justin_luth@sil.org>
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>