tdf#96685 - ensure FindFocus a11y context is valid EditableText
signalIMDelete and Retrieve Surrounding Text search for the accessible
context that has the focus.  When a context with the FOCUSED state was
found it was automatically returned, assuming that the reference was valid.

In Draw tables, especially when using the arrow keys to move between
cells, that often was not true.  So, instead of returning a broken
reference, keep searching through the children until a valid
AccessibleEditableText item that is marked as FOCUSED is found.

Change-Id: I71e3e9bfda56d1dfbdbd93945882560a39e40714
Reviewed-on: https://gerrit.libreoffice.org/22263
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>