xmlsecurity: fix reading OOXML signature from temp storage
When reading a signature, we just take the ODF signature stream or OOXML
signature storage.

When writing, we write to an ODF temporary stream / OOXML temporary
storage, read it back, and only the signatures dialog close handler
writes the signature back to the original file.

The underlying OOXML temporary storage's stream is the signature stream,
so only try to load signatures from the stream if it's not OOXML,
throwing a ZIP file at an XML parser is not a great idea.

Change-Id: I6e42d117b9c97676aaeaad158e78e39a2c39a5e8