Updated core Project: help  81aeff99e0dc94adbee9f0568d636350616ca29b
tdf#97818 "Add a link here -->" in the wikihelp

It does not show in the local help, so it is a help-to-wiki
converter issue.

Moved <comment></comment> on its own line and out of the <paragraph>
in an attempt to ease the wiki converter.

But I can't test in the wiki.

Change-Id: I57bd988de80dd325004eb21ea9f4e256055fb0fe
Reviewed-on: https://gerrit.libreoffice.org/22341
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>