don't fill the matrix cell by cell, tdf#67071
We are now at a place where filling one cell takes about a second with
the big document. We could in theory for the special case of this
document add some caching but that would fail in other cases. This
document is already quite large with an external reference to 70k cells
for each formula cell.

I consider it already quite nice and useable again.

Change-Id: I804094838471507e6bb7b0e0e3387e0c7fe53e4b
Reviewed-on: https://gerrit.libreoffice.org/22388
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Markus Mohrhard <markus.mohrhard@googlemail.com>