sw: fix assert in SwAccessibleFrameBase::GetStates()
We get this assert(bIsSelected && "bSelected out of sync");
when double clicking on frames in the navigator.

Reason is that SwCursorShell::UpdateCursor() calls MakeSelVisible()
to scroll to it before calling InvalidateAccessibleCursorPosition().

Scrolling causes loads of CHILD a11y events and the ATK backend
registers event handlers everywhere and calls GetStates(),
which sees that its frame is selected in the shell but its own
bIsSelected flag wasn't set yet.

Change-Id: Id69c9de2591a5d2d10d181c2ab2339daf97fe488