Removing #defines for String_XXX in sbxres.hxx
This is a [partial] patch for BUG #84938:
https://bugs.documentfoundation.org/show_bug.cgi?id=84938

The objective is to substitute all apparitions of #defined constants
for enum class.

The victim here is the series of constants STRING_xxx that were
located in sbres.hxx and were changed for a StringId enum class.
Obviously some other files where the type is used must be changed too.

Change-Id: I31fa684eb5eb3508d2025c9a319d5b245cc177af
Reviewed-on: https://gerrit.libreoffice.org/22655
Reviewed-by: Noel Grandin <noelgrandin@gmail.com>
Tested-by: Noel Grandin <noelgrandin@gmail.com>