tdf#84938 Eliminate B_xxx defines in filefmt.hxx for FileOffset scoped enum.
Second step: Make SbiOpenRecord() accept a FileOffset as second parameter

This is a [partial] patch The objective is to substitute all apparitions of
#defined constants for enum class.
The victim here is the series of constants B_xxx that were located in
filefmt.hxx and were changed for a FileOffset enum class.

Obviously some other files where the type is used must be changed too.

Change-Id: I7406bdc8e780ee89802ab6823f9c91c4c9b765c0
Reviewed-on: https://gerrit.libreoffice.org/22697
Reviewed-by: Noel Grandin <noelgrandin@gmail.com>
Tested-by: Noel Grandin <noelgrandin@gmail.com>