tdf#90319: make image flipping work when aLinearContext(...) code path chosen
Image flipping can be processed following different code paths in bitmap.cxx.
Has been verified that it was broken in Linux Mint (with cinnamon) because
there a buggy code path was chosen which was not fixed
in commit 3119440a80282692640378fde5e37974ab63f096 in master.
This commit solves the problem for this code path

Change-Id: I22257e70761ca5469c0424b5f9925681cfd4e2e3
Reviewed-on: https://gerrit.libreoffice.org/22753
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>