sfx2 classification: fix category name duplication
A category name was stored in m_aName and also as a label value with key
name PROP_BACNAME(). Fix this by always using m_aName, and doing the
conversion in both directions in
SfxClassificationHelper::Impl::pushToObjectShell() and the
SfxClassificationHelper() ctor, as document metadata wants to work with
only key-vaulue pairs.

Change-Id: Iad386840038966e4483e3503c5fcdc046ca9effc