Fix java home detection on gentoo
On gentoo (and some other distros?) /usr/bin/javac points at a script
that chooses the version of java to run. So when configure resolves
the link it doesn't end up where it expects to be. This mod detects
this situation, clears java_home, and lets it fall through to the next
test which completes successfully.

Change-Id: I7400ac06485475203abd584f60aed7eca81c7f90
Reviewed-on: https://gerrit.libreoffice.org/22847
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noelgrandin@gmail.com>