sd lok: Remove LOK_CALLBACK_PARTS_COUNT_CHANGED from the API.
LOK_CALLBACK_PARTS_COUNT_CHANGED did not cover more complex scenarios, like
more pages deleted at the same time, etc.

Instead, we need to trigger the LOK_CALLBACK_DOCUMENT_SIZE_CHANGED, and the
client has to be smart enough to update the previews in a way that does not
consume too much resources - ie. it has to have a link between the part names
(which are unique) and the previews, and request only the missing ones.

Change-Id: I36ff5dc86f360e3c7bd2c55fae8aaa99e054b4eb