-Werror,-Wundef
For MSVC/clang-cl: "__STDC__: Indicates conformance with the ANSI/ISO C99
tandard. Defined as the integer literal constant 1 only if the /Za compiler
option is given and you are not compiling C++ code; otherwise is undefined."
(<https://msdn.microsoft.com/en-us/library/b0084kay.aspx>)

Presumably fine to just replace #if with #ifdef here (as is used in all the
other cases surrounding it), and not care whether it actually expands to true.

Change-Id: Ie58c76d7e72321b06a11a370fd28221e85fcf871