opengl: deferred and optimized (text) texture drawing
Switching between textures is not cheap, so minimizing the amount
of switching performs better. So instead of immediate drawing we
can accumulate texture draw actions and defer drawing as long as
possible. After that switch all accumulated textures and draw
everything needed with one GL draw call.

This is beneficial for text drawing as we cache many glyphs in
per textue.

Change-Id: I1b94b9ac6a5f2c1a3dbbd75f4df76436a5d40f31