genLang update
Last singleton gone, structure is now

handler (local in gLang.cxx) holds
   a variable of l10nMem
   a temporary convert_gen variable (to analyze file)

l10nMem contains hash list of all PO information

convert_xxx is inherited from convert_gen and instanciated
with a static function convert_gen::createInstance

the lex functions (in c) uses a "this" pointer to find back
to the class. This needs to be done better.

l10nMem contains a link to convert_PO, to save files, this
needs to be split.

Change-Id: I3ad31aac27aac739845062f8da61c8c1c3bf9c31