Upgrade libxmlsec to 1.2.16
No instant drop of any of our patches this time, but a considerable
amount of xmlsec1-customkeymanage.patch.1 is now redundant, as part of
the key wrapper code is available in this upstream release for both
mscrypto and nss.

But that can be cleaned up in a separate follow-up commit.

Change-Id: I197eaffe3a52f2f9c02af982872185e017965006
Reviewed-on: https://gerrit.libreoffice.org/23344
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>