genlang .tree ==> .pot identical.
the conversion .tree to .pot works

Added script gRun.sh with local call of genlang until integrated
gList.sh find all files to be translated.
For now, files that should be translated but are not, are ignored
this is due to 100% compatibility with current executables.

Research shows that .tree and .xhp are handled by the same old
executable to the .pot files are not complete.

Change-Id: I1634207b72075e74438ebfbc22a7753cdddd20ed