tdf#75372 rework sheet-copy names for converting global names to sheet-local
This during copying a sheet copies global names that from a copied
formula cell reference the copied sheet and converts the copied global
names to sheet-local names. References to the original sheet are updated
to point to the new sheet.

It works for names used in formula cells copied but needs enhancement to
pick up nested names, i.e. names used in names that do not reference the
sheet but the nested name does.

Change-Id: I1aa16cb28c9f7b3581bec289435492c21e6fcd73