tdf#98989: vcl: fix handling of non-scalable fonts like "Courier"
For a VirtualDevice only scalable fonts are cloned, but for non-scalable
bitmap fonts still an empty PhysicalFontFamily with no PhysicalFontFace
is created, which causes text to disappear (height 0).

Suppress creation of such families like it was done in LO 4.3, so that
the fall-back can handle it and map "Courier" to "Courier New".

(regression from 8d6697587776136f3121733e1c29d4200720dbd9)

Change-Id: I6542a3f7a01bdf46ae2bcf328fa04064f7f86332