tdf#99140 DOCX import: fix table horizontal aligment to be 'from left' ...
... when it was 'manual'.

Regression from commit c1e563f6efd09cd3463f1b92a3022ae288c92087
(fdo#76741 [DOCX] Table Alignment and width type, 2014-04-04), DOCX
import code had to deal with two kinds of horizontal alignment when it
came to floating tables: the alignment of the table itself, and the
alignment of the float parameters. The problem is, in general it's
wanted that the table is aligned according to the floating parameters,
but in Writer the "from left" UI setting is described differently for
tables and fly frames: tables use LEFT_AND_WIDTH for that, while fly
frames use NONE.

Fix the problem by touching the default only in case the floating
parameters have something that's different from NONE.

With this, the width of tables is no longer lost when they are described
to be floating ones in the DOCX markup, but FloatingTableConversion()
decides to ignore that.

Change-Id: Idd41c3e03e6ded8552e9d15b6080e4b45eb18d3d
Reviewed-on: https://gerrit.libreoffice.org/23923
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>