tdf#93060 Start Center: Update this string to dodge problematic term
I’ve never liked the chosen wording of this string; it has been merely
a decorative item, of limited usefulness. So I’ve decided to replace it
with something that describes a somewhat hidden functionality (drag’n’drop)
while avoiding the misuse of the word “sidebar” to refer to the bar
containing the application launcher buttons.

Change-Id: I299e9e1e97a4e7aa176bc7bbc24cf0e9bea5abb5