tdf#98512: sw: remove duplicate undo history add
Follow-up to commit ed81bf39dd3431d28860fed6a2d4e8814126cc61.

In almost all cases SetAttr() and InsertHint() will already add
the history hint themselves, so this code in SwRegHistory::InsertItems()
looks somewhat pointless as it inserts a duplicate.

It turns out it's needed for hints that insert dummy characters, because
firstly for the hints-without-end the NoteInHistory() was not called, and
secondly when InsertText() inserts the dummy character it may actually
delete the hints array if there are no pre-existing hints, and a new
hints array will be created where the SwRegHistory isn't registered.

It's not obvious how the hints array actually can become empty in
SwTextNode::InsertText().

Let's fix that and assert instead that the history hint was inserted.

Change-Id: If8e0ff1fe28a162f144149a536a7566f94219c0d