extract the minidump uploader code into a static lib
The plan for the near future is to still ship the executable but replace
the interal use if possible with using the static library.

At some point when it is not needed for debugging anymore and everything
works correctly we should only build the uploader executable in dev
configurations. The huge disadvantage of the interal solution is that it
is nearly impossible for a user to upload a crash report if LibO starts
working correctly. Also LibO overwrites the file with the information
after the upload whereas the executable does not.

Change-Id: Ib9854946be3a34e580964c18e1a9c0cce16221d1
Reviewed-on: https://gerrit.libreoffice.org/25862
Reviewed-by: Markus Mohrhard <markus.mohrhard@googlemail.com>
Tested-by: Markus Mohrhard <markus.mohrhard@googlemail.com>