Try to fix tinderbox build running in RDP session from Linux on Windows pre-10
It is a complicated scenario, but I think something like this is going
on:

Before Windows 10, only OpenGL 1.1 was supported in RDP sessions. Our
code detects RDP and tries to disable the use of OpenGL in that
case. This code for this is in the neighbourhood of the blacklisting
of OpenGL implementations. But the use of OpenGL in GLTF and chart2
does not go through code paths that would take that into
consideration, so it tries to use OpenGL anyway, and eventually we end
up in InitGLEW(), where the recently added check for OpenGL 3.0 or
newer causes an (intentional) abrupt termination of the process.

Change-Id: Ia57bfc8d2cefe4e6d7a00282973608d17967d3fd