tdf#100458 Don't forcefully change cell type
Just quoting the source comment from the breaking commit
61c39eae570d6d6040b65bfe93127b30e6080cc8.

// If a matrix formula has a matrix reference cell that is intended to
// have // a blank text result, the matrix reference cell is actually
// saved(export) as a float cell with 0 as the value and empty <text:p/>.

Ok

// Import works around this by setting these cells as text cells so that
// the blank text is used for display instead of the number 0.

This seems to be wrong.

Forcing the cell to a util::NumberFormat::TEXT actually kills
the cell value, if the text is a result of the formular.

I'm not sure if forcefully changing the cell format at this point is
correct at any time, if a cell type was already determinated.

I imagine that an undefined cell type should probably be fixed at the
end of the row import, if Calc detected an error case, or even generally,

This patch is just a minimalistic initial fix. If it's correct, it
probably needs an additional code comment.

Change-Id: If7d9d69467d56c7d8fd2444f9e8b3ae906285264
Reviewed-on: https://gerrit.libreoffice.org/26435
Reviewed-by: Eike Rathke <erack@redhat.com>
Reviewed-by: Markus Mohrhard <markus.mohrhard@googlemail.com>
Tested-by: Markus Mohrhard <markus.mohrhard@googlemail.com>