Check that EASTERSUNDAY with negative value causes error
...even though that's not obvious from the implementation of
ScInterpreter::ScEasterSunday (sc/source/core/tool/interpr2.cxx), which calls
SvNumberFormatter::ExpandTwoDigitYear on that value, which also only checks that
the value is < 100, but not whether it is non-negative.  However,
ExpandTwoDigitYear takes its argument as sal_uInt16, so any negative value is
guaranteed to be converted into a large sal_uInt16 value that is not < 100.

Change-Id: I6f10672e510918f6e9010061541829fcd98e6dd0