Resolves: tdf#100574 Crash when selecting and applying border style...
when a first torn off instance was opened just before the second one
and then the second one sends focus back to the first when its popped
down.

The second one listens to losing the focus and disposes itself which leads to
dereferencing deleted stuff.

So add a reference count to the places these tear offs call popdown so its
still gets disposed but not deleted during the popdown and then protect against
members being disposed with a enough checks to get back to safely

Change-Id: Id5f8eb4771df36305e308a2a9a5035018948f121