tdf#48767 Fix orientation issue after export to .xlsx
After save to .xlsx in usePrinterDefaults was added into PageSetup.
Due to bug in MS Excel, when usePrinterDefaults is present,
(no matter what value), the orientation of the Page is alway "portrait".
To resolve that issue, if usePrinterDefaults is set to default value,
then usePrinterDefaults XML field is not added during export to .xlsx

Change-Id: I3ab310babb278913bce3d66f7fae1192ed9727b9
Reviewed-on: https://gerrit.libreoffice.org/26871
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Markus Mohrhard <markus.mohrhard@googlemail.com>