Resolves: tdf#101076 drag-create creates an unfilled shape...
after using the slide properties->background->all slides

This SetAttributes with nothing selected will end up setting the
default object properties. Here I think we are setting the bg color to
the master, and then unsetting the bg color of the slides to none
so the master color shows through.

So the additional SetAttributes call unsets the bg of the default
object style

Change-Id: I5cfd74eef5637b87d2a8778d4b8c8a0cbc0960db
Reviewed-on: https://gerrit.libreoffice.org/27487
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Katarina Behrens <Katarina.Behrens@cib.de>