svx lok: draw text edits in all views
Use case: a document has a shape with text, and two views. If one view
starts to edit the shape text, the model text is hidden (via
TextHierarchyEditPrimitive2D), and the in-progress text was painted
directly only in the view that edits it.  The remaining views presented
the shape as if it had no text. This commit addresses the subset of this
use case when the views are already created and they already show the
same draw page.

Fix the problem by looking for other views showing the same draw page,
create a view of the text edit for them, and paint them in all views
after the own text edit is painted.

There is not much LOK-specific in this code, except that some extra code
would be needed to not paint the blinking cursor of the text edit in
other views, and then it could be enabled in the non-LOK case as well.

Change-Id: Ib3096a3369fa56663ee209794e102090d362de66
Reviewed-on: https://gerrit.libreoffice.org/27535
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>