Resolves: tdf#100046 the dreaded 1945-04-02 Double DST, also tdf#79663
A coward not having changed this earlier, partially taken from a patch that was
submitted for tdf#79663 but broke API, thanks to Isamu Mogi again.

This removes all the clutter around timezone and DST correction that was
necessary for early ICU versions, which apparently gets things right now and
the clutter can still be confused by corner cases.

Change-Id: I9a90f933e8db8c6e0db145520ebf71cc27621abc