Resolves: rhbz#1364406 inherit the actual format index also for date and time
So summing [HH]:MM cells or calculating with dates uses the same format
in the result, not just the default format of a type.

This also fixes the apparently broken state stored by 4.4 (and earlier,
later?) where no type information was stored with the formula cell,
which may be just due to the old behavior of not applying the actual
format but determining it on the fly instead.

Change-Id: I14d0a7d07185bf5c77e0d7f6989a4a1d1a468d27