Intermediate iOS hack
Just to get this one file to compile. More errors come later.  I just
spent a short time on this while waiting for something else.

Actually I have no idea what we should do on iOS nowadays. Do we want
to use cairo? Do we want to use OpenGL? Would it make sense to mimic
what we do on Android as much as possible? (But what do we do on
Android, and is that by choice or accident?) Even if that might mean
not using APIs native to iOS, but slower (not HW accelerated) FLOSS
alternatives that perform the same functionality, broadly speaking?

Change-Id: Id88a895b90f753417eced744141376656bcf72c3