Use English category names instead of localized ones in theme search tdf#87597
Added an OUString array to hold the original English strings
for search categories. The array is initialized in ctor by using
an initializer list of strings.

Also grouped the related strings of category names. So it can be
iterated easily without writing their names one by one.

Buttons in the UI will be displayed with localized strings,
but the search will take place with the original strings
coming from the array created in ctor.

Change-Id: Ieb72644203df1325e97693b14d5b3476020c6ff1
Reviewed-on: https://gerrit.libreoffice.org/28027
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>