Don't re-check valid SwPageDescs when layouting
nCheckPageNum is set to USHRT_MAX, if all SwPageDesc were checked.
Code manipulating SwPageFrm objects checks the pages via
CheckPageDescs(). If this is currently forbidden, CheckPageDescs()
sets nCheckPageNum to the first page to check.

I guess the !pPage check is there to check the pages - at the end,
if someone has missed the CheckPageDescs call. And we have to run
the check, if the current page is unchecked AKA
  pPage->GetPhyPageNum() >= nCheckPageNum
to do correct layouting.

Change-Id: Ib67878115cde04b7161c919a67131a1e1dc67d8d