Resolves: tdf#92906 badly rendered elements dock after hide+show
The show after hide is seen as "InitShow" which find the optimal size for its
contents. That's ok if it was a toplevel dialog, but its not in this
circumstance of a docked docking window and has to live with the given size.

This seems to have basically come unstuck with

commit 05d4077b724f91fca736d3c3fd64f28e304d7172
Author: Caolán McNamara <caolanm@redhat.com>
Date:   Mon Sep 1 07:17:09 2014 +0100

    rearrange matters to get FloatingWindows working loaded from .ui

Change-Id: If47bb392c8235d946bb2fd07448c6f48479a5163