Massage ifdefs etc for iOS to avoid undefineds
I think it's best to not use cairo on iOS, even if we do use it on
Android. We probably want to use native APIs for the functionality
that cairo would provide. Just like we do on OS X.

No idea whether the resulting TiledLibreOffice will still work like it
used to in May last year, when I last tried.

Change-Id: Ie15cad6918d7a66e2aff7faabfcade7f3246b060