Add handling of a +RELATIVETIMER flag in the SAL_LOG environment variable
Outputs a timestamp in decimal seconds (with millisecond accuracy).

Simplified the handling of SAL_LOG if no "level" is specified. Now
just a totally unset (or empty) SAL_LOG causes the default of "+WARN"
to be used. Given how the code works, it would have become too
unwieldy to check for all combinations of TIMESTAMP and RELATIVETIMER
but no WARN or INFO.

Change-Id: I7bb5bb665d4e764e7eee447e93486f6467042e97