tdf#101589 MS export: use dont-split-row for 1-row table
.doc and .docx do not have a setting for "do not split table" so
it was being emulated by "keep paragraph with next" for all but
the last table-row.  That means that a single-row table with
lots of content might be split.

There is a setting to prevent a row from splitting, so use
that instead in this corner case.  It needs to be
separate from the other function in order to be
set in time for docx to write w:cantSplit w:val="true"
before the row contents. For some reason it did not work
to move the whole function here.

Change-Id: Idc11626e0e2cf1706b87a83f2bd4a802348cb633
Reviewed-on: https://gerrit.libreoffice.org/28352
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>