tdf#100002 MM: directly try filename from column
If we generate multiple files via mail merge with the same name,
we append a number to any file, so the "basename" filename won't
actually be generated.

But if we get the filename from a database column, the user
expects to try the original file name before trying a name with
an appended number.

Change-Id: I68411901ed90d8e9a21423de9233a5be4f41eb12