tdf#62847 import image hyperlink in docx
The image hyperlink is a resource id in the document and needs to be translated into real URL.
First I define a new type CT_Hyperlink_URL in the model and associate it with an action handleHyperlinkURL.
In OOXMLFastContextHandlerProperties::handleHyperlinkURL I dispatch it to OOXMLHyperlinkURLHandler to translate resource id to real URL then set the PropertySet with real URL.
Then the correct URL will be captured while resolving GraphicImport, which will be stored in GraphicImport_Impl::sHyperlinkURL as an OUString.
Finally the property will be set in the GraphicImport::applyName if the length of the sHyperlinkURL is not 0.

Also adds a test file image-hyperlink.docx and a test in ooxmlimport.cxx.

Change-Id: I6194b9cc6bcc1bfaa033ab05e94836fe96e33f14
Reviewed-on: https://gerrit.libreoffice.org/28432
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>