change of plans, we are not going to use balrog
Balrog makes a lot of assumptions about the Mozilla infrastructure that
we can't clone. As an example the whole admin interface assumes an
external authenication system, most likely LDAP. Also it would not
support my plans for language pack and SDK handling.

We will copy some of the ideas of Balrog but will need an own simple
webservice for that.

Change-Id: I2e31f3ef7887b19edcd248ddbd02a1965f71b3e3