Toolbar popups need explicit deactivation
... if based on MenuBarManager, because unlike context
menus, we reuse the same menu instead of creating from
scratch each time, so that it suffers from the same
"double activation" problem as in tdf#97665.

Note that for GtkSalMenu the problem is only for the
top level popup menu, as the Deactivate callback is
properly called when sub menus are closed.

The non-native case also had this problem, but only when
closing the menu by selecting one of its items, and was
fixed already in b41deb2ef057c3bd43fa61448bb4d7299138c8d4.

Change-Id: Iee037f7baba0f80c32fb3a45f16511f8585c203e