tdf#94987 Create directories for temp filenames
Per default a temporary file is construted from a path and a leading
pattern for the filename. For mail merge the filename can be read
from a database column. If the path is not existing, a temporary
directory is created.

Normally the temp file function would fail, if the filename contains
a slash and the sub-directory of the filename doesn't exists as a
subdirectory of path.

To implement tdf#94987, this adds an option to the temp file class
to create the parent directories of the filename pattern.

Change-Id: I02bf34294dac85598ee153d8cfcf00bc5d7775af