tdf#76349 SpacingWithoutBorders: enable .doc RT
Microsoft allows spacing in textboxes even if the borders are not
shown, but LO does not.  Added a compatible setting, used an existing
setting to allow the spacing, and changed .doc export not to zero
out the spacing if the border was zero-width.

Using the compatible setting in the export code is almost redundant,
but it does require that the document was LOADED as .doc, and not
"save as" from another format.

This patch simply allows round-tripping - any user attempt to modify the
border settings will enforce normal LO border rules.

Change-Id: I60ac036e1bfac381eea15e33c21495ad3800277a
Reviewed-on: https://gerrit.libreoffice.org/28601
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>