tdf#82840 Add 'Reset' button to toolbar customization
Add 'Reset' button to the toolbar customization dialog,
and remove the 'Restore Default Settings' item from the
'Toolbar' menu.

Note that this 'Reset' button does not only reset
the selected item but also restores the selected
top-level toolbar to default settings.

Since there is no 'Restore Default Settings' option
in the other tabs of the config dialog, the button is
disabled and hidded in the parent class ctor, then
made visible again in the child class ctor.

In the following commits, more widgets will be added
and moved around to achieve the suggested appearance
and functionality in the related bug report.

Change-Id: I028c082daf423761e3511101069a0dc842743e3b
Reviewed-on: https://gerrit.libreoffice.org/28621
Tested-by: Jenkins <ci@libreoffice.org>
Tested-by: Yousuf Philips <philipz85@hotmail.com>
Reviewed-by: Samuel Mehrbrodt <Samuel.Mehrbrodt@cib.de>