tdf#102274 Closing LibreOffice should not kill active UNO connections
When closing the last window, check whether there are active UNO connections.
If that's the case, just close the window, don't terminate the application
so that the connected application keeps working.

This doesn't affect the behavior of "File->Exit LibreOffice". In that case,
the application still gets terminated and existing connections are closed.

Change-Id: If2d22d51c9b566be8abd51969f35c80896ed4767
Reviewed-on: https://gerrit.libreoffice.org/29018
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>
Tested-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>