sw: remove defensive programming bullshit in lcl_AssureFieldMarksSet
In CppunitTest_sw_ooxmlfieldexport testFdo81492 a TextFieldmark is
inserted at a position where already another TextFieldmark starts. The
defensively programmed lcl_AssureFieldMarksSet notices there is aleady a
dummy character at the start position, and does not insert another one,
but then the dummy character for the end position is inserted, moving
the start position index, which puts the start position behind another
bookmark.

So we end up with a field mark that has a end character but not a start
character and an un-sorted m_vAllMarks.

Change-Id: Icd15e83471e18f607eb41b2f7b0c2ce61c94ff9f