Unify drag handles for toolboxes
The custom shape popup toolbars used their own drawing code for
the drag handle. This was dropped in favour of the default toolbox
drag handle code.

This also drops the drag highlighting in favour of a move mouse
cursor as the drag indicator. The drag handle is currently always
drawn on the top, like a window title.

Change-Id: I08cbf715f1e240c2eb6d9a61fad2b705f9bd8014