tdf#64038 ODF import: fix handling of space following <text:s>
The whitespace collapse algorithm given in ODF 1.2 part 1,
"6.1.2 White Space Characters", was not implemented as well as it could be.

Although the problematic case is arguably invalid, because "6.1.3
<text:s>" says:

    This element shall be used to represent the second and all following
“ “ (U+0020, SPACE) characters in a sequence of “ “ (U+0020, SPACE)
characters.

Hence it is probably irrelevant in practice that a space following <text:s>
was ignored, because a conforming document cannot have such content.

Change-Id: Ic30347fff27176c511ea317d46a1011b410e57d5