tdf#101814 sw: MSWord export: don't override first-page with heuristics
Revert commit dcea8ba3b4fc347e7b6d9e0d5ec2fd99bd77ce6f, and do the
opposite.  The added "titlePage" check makes no sense, because in that
case there is already an explicit first-page style in the PageDesc, and
overriding that cannot be right.

The test added in that commit still succeeds, which suggests that the
root cause of that bug was elsewhere and was fixed in the meantime.

Change-Id: I9b21b6dc115be28cb0ee3360615c538d4cfee3cb