tdf#63846 assign macros after VBA project fully loaded.
fixes regression caused by d4743045a0b320449d07a957463a76bb8b13f939.

see the documentation for VbaMacroAttacherBase, which indicates that
the macros are to be attached after all objects are finished loading.

Made attachMacros() public so it could be called separately since so
many things are inter-dependent in the logic flow.

Change-Id: I11f7f4a1a75d6f77a97e12e3359d6ea8a995f518
Reviewed-on: https://gerrit.libreoffice.org/30337
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Markus Mohrhard <markus.mohrhard@googlemail.com>