Enable the new text layout engine by default
The handful of bugs reported so far have been fixed, and I think it had
enough basic testing to show that it is usable and can be switched on
for wider testing.

We can always revert back to the old engine if needed before or even
during 5.3 series.

The environment variable SAL_NO_COMMON_LAYOUT can be set to disable it
at runtime.

Change-Id: If319a0161897ec676d3123595944a0401c67b63f
Reviewed-on: https://gerrit.libreoffice.org/30443
Reviewed-by: Khaled Hosny <khaledhosny@eglug.org>
Tested-by: Khaled Hosny <khaledhosny@eglug.org>