Use a single method of converting between points and 100th mm in starmath
The SmFormat ctor (starmath/source/format.cxx) uses SmPtsTo100th_mm(12) to
initialize aBaseSize, resulting in a height of 422 100th mm.  Getting that value
(and writing it out to a file) through SmModel::_getPropertyValues mapped back
to 12 point, but in turn setting that value (as read from a file) through
SmModel::_setPropertyValues resulted in a height of 423 100th mm.  That, in
turn, caused SmDocShell::GetSize (starmath/source/document.cxx) to compute
diverging sizes, so e.g. clicking into an OLE-embedded formula in a Writer
document caused the Writer document to be marked as modified (as the embedded
formula's size had changed).

Change-Id: Ia90654aeaaf779690997b5c3bb868b193eaf46c6