chartx: (regression) fix crash on label import
The sparse chart import moved from assuming that the number of elements
in the list parsed from ooxml is the same as the real number of
elements. For this, the patch relies on a new member that was not always
initialized. This patch fixes a missing initialization. According to
'grep' this should be the last one.

Change-Id: I31d8a653f227100436360deef4a53c9418de9d93
Reviewed-on: https://gerrit.libreoffice.org/30838
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>