Demonstrate that (extension) .xcu can contain <item> with nested <node>s
This sprang from an IRC conversation with uniscript about allowing .xcd in
extensions, so that an extension with (localization) data for many packages
would not need one .xcu file per package, but rather only a single file.  That
is possible with a single .xcu file using the <item> syntax, as shown here.  (I
had thought that originally only registrymodifications.xcu supported the <item>
syntax, but can't find any traces of that; so looks like an extension with an
.xcu file using <item> syntax should be supported by OOo and derivates "since
ages".)

Change-Id: I1101793e6b14ba9c1e7647edab76927d4d31a81c