tdf#101464 tdf#91260 sw: a11y for partially off-page drawing object
The problem is that the drawing object in the bugdoc is not disposed
when the document view is closed at a scroll position that does not show
the page with the anchor frame of the drawing object, but does show the
top of the following page, because it will only be disposed via its anchor
page.

SwAccessibleChild::GetBox() must take into account that the
model-determined bounding box of the SdrObject can be partially outside
the page where the object is anchored since commit
a4dee94afed9ade6ac50237c8d99a6e49d3bebc1.

Change-Id: I4276ee636d054320977e2a0ed0f69e53bbc778a3