tdf#104434 tdf#104451 Revert "tdf#46738 Fix exporting .xlsx of coloured em"...
This reverts commit 5239cefff56875c7bb45c046977f1724aace0cfb.

That lead to including / looping through all formatted rows, which
easily happens if an entire column is formatted.

Also when printing it would print all formatted rows and printing a
million empty coloured rows certainly is not desired, only one sheet of
paper maybe. Which was the original intention of SC_VISATTR_STOP and as
the name GetLastVisibleAttr() suggests the function is for visible
attribution.