the return from ImplInit *is* checked
bool WinOpenGLContext::init(HDC hDC, HWND hWnd)
{
... return ImplInit();
}

rtl::Reference<OpenGLContext> WinOpenGLSalGraphicsImpl::CreateWinContext()
{
... if (!xContext->init(mrWinParent.mhLocalDC, mrWinParent.mhWnd))
}

etc.

Another problem here is ...

"If we notice that OpenGL is broken the first time being called, it is not too
late to call disableOpenGLAndTerminateForRestart(). The first time this will be
called is from displaying the splash screen"

isn't true when the first start causes opengl to be disabled, then on subsequent
starts you can enter this function from launching an opengl slide transition.

(note opengl is disabled if opengl is < 30, so all the versioning checks the
slide transitions do is probably pointless now. The only actually useful use
of opengl we have seems to be somewhat unloved)

Change-Id: I1e159f826447520f39ea063fd4c0c83220b1385f