tdf#91260 sw: fix nasty corner case of SdrGroupObject in header
or footer.  Some very special handling going on to paint these in
VOCOfDrawVirtObj with a GetSwDrawVirtObj().GetOffset() added;
they are falsely identified as invisible because their
sub-objects have an anchor set but it is always on the first page
so to properly check them we'd have to add the same GetOffset()
thing but checking the top-level object should be sufficient.

(regression from 689cead9e0837dc932e3a4cd765f7d319b529018)

Change-Id: Ida6fcd41948f46410ca02193549e426bc9dae8c5