Fix smoketest under right-to-left locales
The calc macro test tries to select a range of cells by moving selection
right, but this fails when the UI direction is right-to-left as the
first cell will be the right most and moving right will do nothing.

Fix this by forcing en-US locale like we do with JunitTest.

Change-Id: I08d0bc50013b0936af85af75bca82a44b044d48a
Reviewed-on: https://gerrit.libreoffice.org/32243
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>