Fix the table border toolbar popup
Using ImplPopupFloatWin for it isn't a good idea:
It doesn't need its dragging code, nor its painting code
(which draws the ugly black border), and the resizing code
conflicts with the title-based border which results in a
wrong size (and we might have even bigger problems, if we
start to use .ui based layouts for the ToolbarPopup class).
We can of course put checks all over the place, but why do
it, if we can simply use plain FloatingWindow?

Change-Id: Ibc9a5c220309d15a60f6425835e1cc7a1b19c530