Updated core Project: help  5063054f398d3d354ac215cdae56492f5450ad98
Move the 'how does word count work' embed to the end

Currently it is embedded in the main section, which is
embedded on the Writer Tools menu page and looks too terse.
Also the 'how to get' section should not be the last
on the page as it is now.

Change-Id: I4da0acfe8ca0733faaf51a7ee3dedfc35ee1c121
Reviewed-on: https://gerrit.libreoffice.org/32525
Reviewed-by: Gabor Kelemen <kelemeng@ubuntu.com>
Tested-by: Gabor Kelemen <kelemeng@ubuntu.com>