Updated core Project: help  92e858687cb974b9f6074e9643f5377f24eb73a0
tdf#98643 Fix links between UI and the 'Define Names' page

So that they work in the Manage Names dialog too.
Fixed up the text as well a bit to better fit the
current state of the windows.
I can't decide yet if it would be better to create a
new Manage Names page or to change this one to
mention Manage Names more (in title, howtoget, etc.)

Change-Id: I5f39b02c54ccebde723b6b8b61bf7e58cd054028
Reviewed-on: https://gerrit.libreoffice.org/32536
Reviewed-by: Gabor Kelemen <kelemeng@ubuntu.com>
Tested-by: Gabor Kelemen <kelemeng@ubuntu.com>