tdf104061 - looking for test documents
This fixes the specific test document here, but likely there
need to be other qualifiers such as "if the table is the first
node in the section".  Likely this commit will be reverted.

We will use Milos' document tester to hunt for the side-effects
of this patch - which should give us a good set of test documents.

Change-Id: I7d3a2168556c82bc4758ad665da319d55df4ba2b
Reviewed-on: https://gerrit.libreoffice.org/32759
Reviewed-by: Justin Luth <justin_luth@sil.org>
Tested-by: Justin Luth <justin_luth@sil.org>