Always exit with error if codesign fails
That sanity check was added in
615fae2f67028f3c5c51c70c77dbaa9b9f3856d6 but we lost it at some stage.

The codesigning script has a very confusing history, in part caused by
its use for two purposes, when building for the Mac App Store and when
building for a TDF style distribution on a dmg disk image. Those who
work with the former and those who work with the latter haven't
necessarily checked that it doesn't break for the other case. Sorry.

Change-Id: I79011302f60b1f6551328c8b80e00f5d3698504c