Rename mail dispatcher members
Makes the members conform with most LO variable naming,

This especially switches Active <=> Running in the member names.

"Running" is normally referred to as the thread state, while it
exists (also look into the osl::Thread API) and now active is the
state, when messages are processed, controlled by the start() and
stop() functions.

Change-Id: I3b8437d4ff24731348272a84720ff81fdea90db7