tdf#96277 revert ww8 import: fix upper margin of floating table
commit 3755c87d0426a499d4755e82d8fbc77caa240764
mis-positions non-textframed floating tables.

The patch only worked under very ideal circumstances
that cannot be identified during import time. Lots of bug
reports indicate real world problems with it.

This is a pretty safe revert because the code only
processes a document one time, so old documents don't
depend on this to work.

Change-Id: I3b94d27901a20e46c2f5fd4305041d1d28acf33a
Reviewed-on: https://gerrit.libreoffice.org/33322
Reviewed-by: Justin Luth <justin_luth@sil.org>
Tested-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>