bffvalidator: fix a retro-fitted by ms after the fact warning
nVerBuiltInNamesWhenSaved (2 bytes): An unsigned integer that specifies the
version number of the style names as defined by the application that writes the
file. This value SHOULD be 0.

old number is 2: which means...

"Word 97 2"

Specifying a value of 0 is recommended for maximum compatibility, as it will
cause all versions of Word to update the names to whatever set of application
defined style names is current, with little performance penalty.

http://msdn.microsoft.com/en-us/library/0E246123-E907-4AD1-9DFC-558512E2B052

Change-Id: I9a04f62f950c7965d6eb5e6ff23d9c09fe38d99e