tdf#105672 framework managed menu button
This adds a menu button that can use a popup menu controller
to manage its menu. It supports 2 cases:

- Use any controller that is registered in Controller.xcu,
  by specifing its .uno command.

- Manage an arbitrary popup menu with MenuBarManager
  (assuming its items have proper .uno commands in their
  MenuItemData::aCommandStr). It means that a menu that
  was defined in a .ui file, can be used inside that .ui
  file without any additional code.

This commit uses the new control to fix some currently
non-working buttons in Calc's Notebookbar (but there are
more that can be fixed the same way). It's not clear how
long we will continue to use buttons (instead of
toolboxes) for the Notebookbar, but hopefully this control
will be useful in other places too.

Change-Id: Ie00cde7cd7e39948948960ca2eff76e9db837109
Reviewed-on: https://gerrit.libreoffice.org/34103
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>