i#92948 sw: WW8 import: fix table in Apo in header
The document contains a table inside of a frame anchored in the header.

In WW8TabDesc::WW8TabDesc() the update to nOrigPos makes a mess of
things because the new nOrigPos value contains the value of nCpOfs (93)
already, so later on when SwWW8ImplReader::TestApo() calls
WW8PLCFxDesc::Save() / WW8PLCFxDesc::Restore(), the call to SeekPos()
positions it on CP 325 instead of 232 and the subsequent call to
GetSprms() finds nothing on the wrong position; thus TestApo() concludes
that the paragraph has different Apo properties than the previous one
and a new fly frame is created.

(regression from 4bd341e2c8a1f3a4f4677eb879a60a4ffc3369db)

Change-Id: I697c4526c7c3a12119947ff5659c11ea9ee7fff8