Revert "sw lok: Move this test to sw/"
This reverts commit b7640cbb4b34df2cb05c6fbe60346bca96846924.

Not much idea yet but moving it to sw/ for some reason makes callback
pertaining to last 'InsertAnnotation' uno command not firing and on
ubsan tinderbox even the first 'InsertAnnotation' command is a noOp
failing the tests there. Lets keep it in desktop/ for now where it was
earlier.

Conflicts:
	desktop/qa/desktop_lib/test_desktop_lib.cxx

Change-Id: I67f5bdeffa218ae2a7afe0896f7cdde5324a5f09