sw: Fix STL assert on DOCX export of ooo29679-42.odt
The problem was that the StartField_Impl() was called again and again on
the same field (actually 0-length ToXmark with dummy char), or in other
words, EndField_Impl() wasn't called and it wasn't removed from
m_Fields.

So tweak the south-pointing chariot DocxAttributeOutput::EndRun() again
in the hope it will go another km or two before it starts pointing east.

This doesn't actually produce the elements in the ideal order in some
cases, but given that this code has clearly passed the complexity event
horizon that is too much to ask for.

Change-Id: I9b031ce8fae91d3e5af3cb1148390807574fb748