doc/docx: make unit test for tdf#99120 round-tripable
When a page description is chained to a non-chained follow,
it will only be used for a single page. Thus it might be able
to be combined into the follow as a titlePage instead.
However, if this page description has distinct even/odd pages,
don't merge it as a titlepage.

Change-Id: I62237a48c062e7c393a5224285a23597c479c1c9
Reviewed-on: https://gerrit.libreoffice.org/34889
Reviewed-by: Justin Luth <justin_luth@sil.org>
Tested-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>