lcms2: try to get rid of $(DEVENV) /Upgrade
It's not clear to me that this is actually required, particularly why it
would be required for VS 2015 but not VS 2017, and why only for this
external and not the others that use MSBuild.exe.

Reportedly this can fail if you have an expired or not yet registered VS,
while strangely enough everything else compiles fine in that case,
so rather than try to find out how to check for that issue in configure,
avoid the problem by removing the /Upgrade.

Change-Id: I55566e109e57117f65febb91de7580667c984a54
Reviewed-on: https://gerrit.libreoffice.org/34947
Reviewed-by: David Ostrovsky <david@ostrovsky.org>
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Michael Stahl <mstahl@redhat.com>