ios LibreOfficeLight, FileManager
Online consist in reality of an independent FileManager
and a DocumentViewer. Adding a FileManager that "knows"
ios specific fileHandling.

Added UI dialog (storyboard)

FileManager consist of 2 parts, a FileManagement class
connecting to IOS (will later handle differencies between
local storage and iCloud) and a FileManagerViewer giving
the user a UI to select files.

Change-Id: I30b4eafe48a3396c100e7475ba7118f0ee2de73d
Reviewed-on: https://gerrit.libreoffice.org/35451
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: jan iversen <jani@libreoffice.org>