tdf#95992 "Edit" > "Compare Document..." opens dialog that is...
misnamed "Insert"

Issue is that the caption of the dialog is handled by generic code,
which was not designed with comparison in mind. This fix extends the
captions that can be used by such generic codes, without modifying
the underlaying architecture.
Letting callers directly provide caption text might be more versatile,
though.

The fix is extended to:
* Writer Merge functionality
* Calc Compare/Merge functionalities
which use the same dialog (and thus were also displaying "Insert").

Change-Id: I452b37bf7d0024924c87316cd47572c09b373b65
Reviewed-on: https://gerrit.libreoffice.org/35285
Reviewed-by: Michael Stahl <mstahl@redhat.com>
Tested-by: Michael Stahl <mstahl@redhat.com>