uitest: add logging for UI actions
The long term goal for this logging is that it generates content in a
DSL for the UI testing. The generated file can then be interpreted by
the UI testing and replay the interaction with the UI.

For now the plan is to have a readable output of what happens in the UI
layer that allows to quickly transform it to a UI test.

Change-Id: Ic536db766e41d03d048c920f6d551047af6fbb74
Reviewed-on: https://gerrit.libreoffice.org/35447
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Markus Mohrhard <markus.mohrhard@googlemail.com>