Resolves: rhbz#1438876 detect preferred help module for csv dialog
For calc import before any toplevel is created the context is otherwise
unknown. If a help id starts with a specific module name, then default to that
module name as default help category context. Cosmetic, same help is shown in
any case because its in the shared section, but title bar would state "Writer"
when context is expected to be "Calc"

Change-Id: I3323bc55ef23ec162dad4349efb5e84ec3e8054c