tdf#106933 vcl: delete D2DWriteTextOutRenderer before exit()
As it happens this DirectWrite stuff is using some thread pool
internally, and that must be shutdown before exit(), as Win32 will
terminate all other threads at that point, and then the thread pool
wants to talk to threads that don't exist any more.

https://blogs.msdn.microsoft.com/oldnewthing/20120427-00/?p=7763/

So convert this from a global variable that is deleted from DllMain()
to a member of SalData, so it is deleted from DeInitVCL().

Change-Id: I51408a07c78758cf0c193ab66b9214d0c9dbd9e3