tdf#107204: Make sure we always do Unicode Bidi, take two
[ Now that the side-effect will be fixed by
<https://gerrit.libreoffice.org/#/c/37050/>.]

Writer seems to have an “optimization” that only applies bidi algorithm
if the text direction is nor LTR and there is some characters that we
classify as complex script. This is a very simplistic optimization as
evidenced by the bug above. We certainly don't want to skip bidi the
text differently based on the default direction (whatever that means),
and not all RTL scripts are complex ones, even if they are we might be
behind Unicode in our script classification.

Just trust that bidi algorithm does the right thing and don’t do
premature optimization. If this turns out to be a real performance
issue, we will need to find a better optimization.

Change-Id: I276e2ae0ac92fb51d2e37ef2ad0b11efc6b441ef