external/epoxy: Work around missing clang-cl DLL-init CONSTRUCT support
Looks like neither the __GNUC__ specific __attribute__((constructor)) definition
of CONSTRUCT nor the _MSC_VER specific one (if the #ifs were reordered so that
clang-cl would pick the latter) with __declspec(allocate(".CRT$XCU")) are
supported by clang-cl, and both are rather silently ignored.  That means that
library_init is not called, library_initialized remains false, and the first
call to get_dlopen_handle aborts.

But this whole "verify that get_dlopen_handle isn't called too early" business
is somewhat pointless here (and that's the only use of the CONSTRUCT macro, and
DESTRUCT isn't used at all), so just short-circuit it for clang-cl for now.

Change-Id: I5d50df3574d350f9591e807ef0fb6a1b02dc34ec