Remove SlotType sdi property
sdi files have two ways to specify the type of a slot:

(a) Put it at the beginning of the first line
(b) Using the SlotType property

Where (b) have higher priority, and (a) is actually
ignored for "method slots" (i.e. the ones that have
parameter parentheses, even empty, and therefore get
the SfxSlotMode::METHOD flag), which is very
confusing.

The goal here is to make exclusive use of (a) for all
kinds of slots, and eventually remove the SlotType
support code from svidl (in a follow-up commit).

Change-Id: Ib6654d34e2f8aa8ebd0bd3c41aa2a86e30e9a634