tdf#82326 calc 32bits unable to open files with a lot of cond formatting
The code to build and compute formulas is using an fixed size array.
The size of this array has been changed from 512 to 8192 by commit
9c1ca6dca3b553c302a635357e33591605343b99

Since then, LO consumes a LOT of memory while opening scalc
files which are making extensive use of conditional formatting.

This is a silent issue when working with a 64bits versions of LO because
the memory is immediately released at the end of the load phase.
However, with 32bits versions, some files can not be opened anymore as it
can consume the full process memory space (2GB on windows).

The parser is creating ScCondtionEntry objects to store conditional formatting expressions
which are immediately compiled in the constructor. However the compile
method did not return a flattened token array.

Change-Id: Id290bc1f5838c246d4c2322b297b13307defaf88
Reviewed-on: https://gerrit.libreoffice.org/37527
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Eike Rathke <erack@redhat.com>