Clean obsolete code from menu document writer
Remove special handling of New, Wizards, and Addon sub-menus.
Many years ago they used to be injected into the menu structure
at runtime, so we had to not write their contents back to xml.
Now we have a clear separation between the contents of a menu
at UNO API level, and the actual contents of the VCL menubar.

Change-Id: I5e791e8963faca9f26ad31c9b4e6b8763b70d1d9