tdf#79571 sc FilterSelect: accept empty strings
If a drop-down list has a valid entry that is an empty string,
it should be acceptable as a data entry value, instead of just
ignoring it and leaving the old value in the cell.

Yes, the other simpler option is to just hit delete, but anyone using
the UI will get confused, and if not paying attention will have
incorrect data entry since it doesn't work as expected.

In the bug tested here, the selection dropdown was created with
data validation. Unsure of other times that create the grid...

Change-Id: I3b49c34d01079d4c1ec9e140e2c51123408b10a6
Reviewed-on: https://gerrit.libreoffice.org/38124
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Markus Mohrhard <markus.mohrhard@googlemail.com>