writerfilter : remove dead code with m_sFootnoteFontName
The string variable is never initialized or set,
and the only use is if it is not empty.

In 2007, author Oliver Specht 2007-06-12 04:41:16 in
range fe8e4e774087465a10ecd8f366b7f0c272d8319a
introduced m_sFootnoteFontName in a pair of commits.

The call to SetFootnoteFontname() was removed in 2008
with commit 009e3f4841378f13a2764a2c84e147e401b70525

The SetFootnoteFontName() function was removed as unused in
commit 6ea66657ebca5b3ec63f93b496b812888e0d53c9

Since the footnote name is never set, it is always
empty, and thus this code is useless.

Change-Id: I64898ce9fa6673fbef0106054b5aa5a817c145d5
Reviewed-on: https://gerrit.libreoffice.org/38257
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>