tdf#102930 Add custom HIDS for New Address Block page
The New Address Block and Edit Address Block windows
in Mail Merge wizard use the same ui file as the Custom
Salutation windows, yet they differ slightly in appearance
and a bit more in functionality.
There is already a help page for this, so we need a few HIDs
to be able to hook into that.

Change-Id: I5d1e11d62e28c02d74e0e6833e493923a5d699cf
Reviewed-on: https://gerrit.libreoffice.org/38327
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Katarina Behrens <Katarina.Behrens@cib.de>