odk: remove PackageSet_odk_headers
This is quite confusing: the gb_Helper_register_packages_for_install
odk_headers does not actually use the Package odk_headers, but the
PackageSet odk_headers, because the name is the same and the PackageSet
directory comes first in the search path.

This means that the Package odk_headers_generated is installed despite
there being no obvious reason why.

The PackageSet doesn't provide much value here, so just remove it.

Change-Id: I564f3b9fc6acaabe700328bc8c3db70c3b2de0cd