tdf#82173 writerfilter: charStyle XnoteReference->Xnote Characters
In MSWord, there is a single character style (Footnote Reference) that
affects both the in-text anchor number and the in-footnote number.
In Writer, that is split out into "Footnote anchor" for in-text and
"Footnote Characters" for in-footnote.

The biggest difference in LO is that Footnote Characters does NOT
default to superscript, while Footnote anchor does.
In order to handle the default situation, it is best to map
"Footnote Reference" to "Footnote Characters" so that the style
differences between the two will be removed. It seems
that WW8 also does the same thing.

The "regression" will come if someone has done custom formatting that
they prefer to show up in-text rather than in-footnote. Now it will be
reversed, where the custom formatting will show up in-footer and not
in-text. However, it is preferable to handle the default case and
"fix" the missing superscript problem.

(In WW8, the stiXnoteRef's are RES_POOLCHR_FOOTNOTE/ENDNOTE,
as mapped in sw/source/filter/ww8/writerwordglue.cxx,
which turns out to be "Xnote Characters" styles.)

Change-Id: I9152b4a85f6facf3f8192c943e66312988718b0d
Reviewed-on: https://gerrit.libreoffice.org/38306
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>