tdf#108524 sw: try to split rows that contain 1-col sections
We used to not even attempt to split a row that contains sections. Relax
this condition and try to split the row in case the table itself is not
in a section (to avoid recursion) and the section has no columns.

This is needed, but not enough to split the section in the bugdoc.

Change-Id: I6ad0d6eb18611f108ae29e4feea7101ffe552c48
Reviewed-on: https://gerrit.libreoffice.org/38824
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>