don't parse align and scale in VclAlignment
allow it to pass up to superclasses for parsing, from discussion:

<noel_grandin> caolan, is it intended that the VclAlignment x/y
align/scale members are effectively unused?
<caolan> noel_grandin: its based off the GtkAlignment so that's why it
had them in the first place. Never got around to doing anything about
scale so no surprise there and might as well remove those
<noel_grandin> caolan, will do, thanks
<caolan> I would have expected the xalign/yalign to do something, I
think looking closer that we should be using those and probably that
stub impl is blocking the underlying Window::set_property from using the
supported left/right top/bottom values, so you can try remove those too,
but I'd do it in a separate commit from the definitely useless scale
ones
<noel_grandin> caolan, at the moment, the change will remove them, but
not the parsing of the values, so it should not change the functionality
<noel_grandin> caolan, I'll do the removal of the parsing in a separate
commit, which will pass them up to Window::set_property

Change-Id: I5bb043a5c38d494355bf4766af8521838b7fc3f2
Reviewed-on: https://gerrit.libreoffice.org/38957
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>