tdf#107505: Impress - crash with two people co-editing one slide
In tiledrendering mode cursor inside shapes are not handled very well now.
Different LO instances manipulates the same cursor instead of having
one separate cursor for all instances (as in case of simple text editing
in Writer).

Crashes caused by that one cursor is owned (created and destroyed) by one
LO instance but used by an other instance in the meantime. We can avoid
this by not making this cursor available to the other instances. Cursor is
hidden anyway so don't need to have it set to the corresponding Window.

Change-Id: Id78f79e73eda88f3c2c6aa7f534c32f3f8aba174
Reviewed-on: https://gerrit.libreoffice.org/39432
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tamás Zolnai <tamas.zolnai@collabora.com>