Fix instant 0ms scheduler wakeup for headless
This is a regression from commit
  503eba23c9a199583eddee9e169a4fddbecf416f

Due to rounding errors, as the Scheduler uses milliseconds, but
the headless backend uses nanoseconds the StartTimer assumed it
would wake up this ms, but the headless check function would still
wait. This is more of a workaround, so instant wakeup for the
headless backend works again.

Change-Id: I2ba9b4ad2b67ec99eeb4dd098ded6457d3753127