SfxItemSet::Equal: ignore parents if the pools are different
If there are parents, they are necessarily different pointers for
different pools; the only sensible way to compare them is by-value,
which is a bit tricky because the items overwritten in the child
should of course be ignored.

Rather than rewrite the whole function, just ignore the parents for
now, because this comparison is currently used to replace compares
of the serializations of item sets, and SfxItemSet::Store()
completely ignores the m_pParent, so this looks like the best
replacement.

Change-Id: I0a4bf44d4fef49e47a3ec78bc7d977a6fcbd789d
Reviewed-on: https://gerrit.libreoffice.org/41186
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Michael Stahl <mstahl@redhat.com>