tdf#107480 RTF import: fix HTML automatic spacing in styles
\htmautsp changes the meaning of \sbauto and \saauto, but the auto
keywords may appear early in style definitions before \htmautsp arrives.
This wasn't a problem before commit
1be0a3fa9ebb22b607c54b47739d4467acfed259 (n#825305: writerfilter RTF
import: override style properties like Word, 2014-06-17), as there the
right auto keyword interpretation was repeated as direct formatting.

So fix the regression by not deduplicating these keywords, which brings
back the correct layout. A larger rework is necessary in the future to
delay the parsing of the stylesheet table, that would be a cleaner
approach.

Change-Id: I783d2175483bc80ead0a511ad15c4b75e4ab663c
Reviewed-on: https://gerrit.libreoffice.org/41187
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>