writerfilter: default graphics in header to the background.
Instead of always qualifying each bOpaque query with an additional
check to see if it is in the header/footer, base the default
value conditionally on IsInHeaderFooter().

This doesn't cause any logic change, but it does re-enforce the
concept that for compatibility, graphics in header/footers are
always behind the text. Especially helpful to prevent errors
in future uses of bOpaque that otherwise might miss this
important qualification.

Change-Id: I7c0142f220d454c09ff14af65382317273402a86
Reviewed-on: https://gerrit.libreoffice.org/41346
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>