WalkAndMatchElements: really really limit the match, tdf#108292 follow-up
getRemainingCount() could deliver a wrapped around overflow value if mnIndex
was already greater than the end index, which could happen if when/for
non-matching larger block sizes were added, and if then a match was found
behind those blocks a non-requested/unexpected index was returned, which in
turn led to the assert() being hit in ScInterpreter::CalculateLookup(). In
non-debug could result in an invalid block position access.

This happened with the bug case document of tdf#111943 which in master can be
loaded.

Also, the start and end index are not dynamic and don't have to be recalculated
each time, so make them const; column argument values are unused after.

Change-Id: Ic294cade4e8e7828bee394e5ade61d7127be6bbb