emfplus: Corrected ReadWindowMetafile
There are two places which do hand over not a
complete SvStream staring at pos zero, but pass
a seek position indirectly in that stream. Thus
this needs to be used. There is one usage in sc
that copies the data to a MemStream to avoid that,
so this hints that this 'trap/feature' was not
known to everyone using it

Change-Id: I94139b86c8bdd82879124c574bc3014e02d9ab5f
Reviewed-on: https://gerrit.libreoffice.org/41399
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Armin Le Grand <Armin.Le.Grand@cib.de>