tdf#111895 writerfilter: avoid hiding shapes behind background
In MSO, page background is "in hell" zOrder, however
in LO, page background is at the same zOrder as the text,
so avoid hiding shapes behind the page background unless
doing so would hide text.

When importing gradients/pattern/hatch backgrounds are supported,
HasPageBackground will need to be expanded to include those as well.

Change-Id: I4496fd3f4c644ac9ac1854090658017078ce3e88
Reviewed-on: https://gerrit.libreoffice.org/41486
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>