tdf#112074 doc export: for RTL tables, LEFT_AND_WIDTH is RIGHT
In MSO tables, the only side that can have an indent is the
logical left orientation. In right-to-left tables, the indent
is measured from the physical right margin. So, we need to
save the table's orientation as RIGHT aligned.

In LO tables, only the physical left orientation can specify
the indent. Importing already calculates the corrected index size.

On export, the indent size will need to be re-calculated to account
for the reversed the margins, but that will come in a separate patch.

Change-Id: Ie2457791f913af2360af96f5038063d305e366de
Reviewed-on: https://gerrit.libreoffice.org/41665
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>