Adapt external/beanshell to Java 9
For one, drop a spurious target="1.5" so that ant instead uses the passed-in
JAVA_TARGET_VER (which needs to be 1.6 for Java 9).

For another, the deprecated java.awt.Component.getPeer method finally got
removed, so bsh.util.AWTConsole would no longer compile.  But it looks like that
class is not used by us anyway, so just remove the problematic code.
(workdir/UnpackedTarball/beanshell/docs/manual/bshmanual.html: "Web Browser
Access:  After starting the server you can connect your web browser to the port
you specified.   BeanShell will respond by sending an HTML page offering you a
choice of the Swing based JConsole or the older AWTConsole.  You may choose
whichever is appropriate for your web browser.  [...]  The bsh.util.AWTConsole
is a legacy implementation of the GUI Console using AWT instead of Swing.  This
console does work, but it is not as slick or pretty as the JConsole.  The
primary reason it is still here is to support remote access from generic web
browsers using only Java 1.1.")

Change-Id: I9f1112caebae9a55e4eec7f9baa0114b2254ccda