Resolves: tdf#112258 correctly pick items from AutoFilter selection
Combining only one child with one parent to lookup is not sufficient if we have
year,month,day chains. The entire chain is needed.

Regression from

    commit 2a39dc74724d3648ff76aa900edfebe0dd19b296
    Date:   Thu Jun 1 18:27:13 2017 +0200

        tdf#108259 Fix nested checkbox handling in autofilter popup

and

    commit 511fb8e80d298d42f5c45e7410bf64f2a25b441e
    Date:   Wed May 31 10:59:42 2017 +0200

        tdf#108259 Enable autofilter with many different values

which combined landed in 5.4 as

    commit 0163957ef808cffa332c2ddd3267409c5ae1494a
    Date:   Wed May 31 10:59:42 2017 +0200

        tdf#108259 Enable autofilter with many different values

This on the other hand makes the change of these doubtful, as for each leaf
item the concatenation and the lookup has to be done. Something to be
investigated.

Change-Id: I07fd6367bb8da2adab94a43c45fe88391179f496