tdf#97648 ooxmlimport: horizontal line's horiOrient
Although LO won't visually do anything with this property
(because o:hr is improperly imported AS_CHAR), at least
import and assign the property so that it is there for
the future if o:hr is changed from AS_CHAR to AT_*.

The unit test just adds a test for the existing status - to
highlight what still needs to be fixed.

o:hr probably ought to be AT_CHAR, with no wrapping. However,
simply doing that wraps text the wrong way (always underneath)
which seems to be a LO limitation of not putting preceeding
chars above an AT_CHAR fly. So that's why I'm not implementing
a change to AT_CHAR.

Change-Id: I4aec99d3976c6d461715dba304cdf9ad98ab12b6
Reviewed-on: https://gerrit.libreoffice.org/42440
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>