TSCP: add advanced classification dialog
This adds a advanced classification dialog, which enables the user
to manually construct a header/footer message from classification
properties in cases where the user needs more control. All the
text is inserted as fields into the end document and can be changed
by changing the document properties. The simple classification
still functions as it did before, which is what most users will
want to use in this case.

The BAF policy has been extended with new xml extension elements:
- loext:Marking
- loext:IntellectualPropertyPart
- loext:IntellectualPropertyPartNumber

They are used to fill the values in the advanced classification
dialog.

Change-Id: Ie2d638d69b8a9b0799cff9c2b785eb789f8af1d5
Reviewed-on: https://gerrit.libreoffice.org/42474
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Tomaž Vajngerl <quikee@gmail.com>