sw: fix ~SwIndexReg assert in DelFullPara() importing tdf108272-1.docx
The code in DelFullPara() returns early if it can't move the rPam
out of the to-be-deleted paragraph, but it only checks that the point
of the rPam is outside; it then rather pointlessly assigns the 0
index in the node where the mark is, which leaves the mark on the
to-be-deleted paragraph.

In this modern day and age, after much technological progress,
DeleteMark() actually does unregister its SwIndex, so just do that.

Change-Id: Ia1643cf0b976857809ee4130efdec37ce831fdef