sw: fix infinite grammar checking idle loop
The grammar checker always wants to be started in DoIdleJobs(),
even if all paragraphs are already marked as checked.

This is because there is currently no call anywhere of
SwRootFrame::SetNeedGrammarCheck(false) to reset the flag
and prevent DoIdleJobs from trying to start the grammar checker.

This call was already there before but was removed without
any justification in commit 9160fe814a46e93da6907e169ce9d58e46fa37f2.

This has become an infinite loop in several Junit tests with commit
53da556c600fa82ba84bc7fdce6a594b43f2b097.

Change-Id: Ibe7ad93442070aac0577725d044281912307d9e8