tdf#112211 RTF import: fix unwanted direct formatting for other indents
Commit 56a695fddb915bcba13b088b5b2b4e0841d4acbc (tdf#112211 RTF import:
fix unwanted direct formatting for left indents, 2017-09-26) fixed left
indents, and given that it was a regression fix, left the other indent
types untouched.

As it has been pointed out in the bug comment, the original bugdoc
actually needs the other indent types removed as well, so let's do that.

Change-Id: Ia4ea7e2214b7df27536f46b046f90bd703c107be
Reviewed-on: https://gerrit.libreoffice.org/43303
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>