Fix CppunitTest_services for constructor-based implementations...
(i.e., using constructor="..." in the .component file) that implement multiple
services, but are not implemented as single-instance objects (i.e., always
return different objects from that constructor function).  In that case, the
"factory" would be the object itself, and would thus be different for different
instantiations of that implementation.  erAck ran into such a case with new code
he's writing.

Change-Id: Ib134a4c6d1400b43a7df3cf6a115f8206fb80c93