tdf#113520 sw split sections in tables: avoid empty page after content del
The problem was that since split sections are allowed in nested tables,
deleting enough content so that the full content of a page is removed
results in an empty page, having only a 0 height table.  I.e. the split
section gets removed properly, the inner table (and its row and column)
as well, but the outer table is still there, though it is marked to have
0 height.

This results in a situation that the previous page doesn't have free
space (the in-table section tries to take as much space as it can), but
on the other hand we try (and fail) to move the 0 height table on the
current page to the previous one, as it doesn't have free space. At the
end the "empty" page still has an invisible table frame, so it is not
removed.

Fix the problem by allowing the move of a 0 height follow table frame
from the current page to the previous one, even it has no empty space.

Change-Id: I2a5fac88b8b7dc2b91d041b58a4ad1b328f56a6b
Reviewed-on: https://gerrit.libreoffice.org/44059
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>