TSCP: read/write recently used in advanced classification dialog
This implements the reading and writing of the recently used
content in the advanced classification dialog. The recently used
content is stored to a xml file in the user profile folder under
"classification" folder - filename "recentlyUsed.xml". The content
is just a reflection of the resulting ClassificationResult vector
in a simple xml form.

Change-Id: If148ae123a6c826874b86dc1f374cdaab6e8fea9
Reviewed-on: https://gerrit.libreoffice.org/44150
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>