icu: don't skip building the data just because tools were disabled
building for e.g. android disables the tests and requests data in a
static library, but icu completely skips building the data directory in
case --disable-tools was specified:
icu/source/Makefile.in
    @TOOLS_TRUE@DATASUBDIR = data
will become
    #DATASUBDIR = data
and then
    SUBDIRS =  stubdata common i18n $(LAYOUTEX) $(ICUIO) $(TOOLS) $(DATASUBDIR) $(EXTRA) $(SAMPLE) $(TEST)
will not have the data dir and make will ignore it.  Add it back by
specifying it when invoking make and all is fine.

Change-Id: I0af693f22938ebabdc189a97f1cfc3f8b1c042ee
Reviewed-on: https://gerrit.libreoffice.org/45107
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Tested-by: Jenkins <ci@libreoffice.org>