etags: pass the "-e" flag directly to ctag binary
'man ctags' says that emacs mode will be enabled if the ctags binary is
renamed as etags or '-e' flag is provided to ctags binary.

Before this patch, the script assumes that host system has an 'etags'
binary renamed from 'ctags' program. This is not always the case in all
hosts. Eg: In Fedora, 'etags' binary is provided by emacs-common package
which doesn't understand the flags given later in the script.

It is safe to just explicitly enable the emacs mode via '-e' flag to the
ctags binary.

Change-Id: Ic7ded56cff32683fc5e9d3fcc7405e79da4c23b7
Reviewed-on: https://gerrit.libreoffice.org/45358
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: pranavk <pranavk@collabora.co.uk>