Remove arch-pinning for gpgme on Windows again
With e0d94237878d4949573fc9b6d97eea3146cada6d the dependency on gcc
is gone, so we should be able to build on all cygwin flavours again
for all Windows arch targets.

Well almost - we still have windres on cygwin32 only able to produce
pe-i386 targets, so let's exclude the case cyg32 and WIN64 for the
moment.

Change-Id: Ibe7f967585995586985177d4c36cc33d9850d9de
Reviewed-on: https://gerrit.libreoffice.org/45422
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>