tdf#41650 DOCX export: split para on section break
Given two pages with entirely different style (resp. not plausible for a
single section) and a paragraph which starts on the first page and has a
follow on the second.

Those page styles are represented in doc/docx structure with sections.
Doc/Docx does not support having a paragraph in more than one sections.

Workaround is to split the paragraph into more paragraphs and put the
section break in between.

Change-Id: I4014c45f97e18132470d0d0647bce84831b2460a
Reviewed-on: https://gerrit.libreoffice.org/45666
Reviewed-by: Tamás Bunth <btomi96@gmail.com>
Tested-by: Tamás Bunth <btomi96@gmail.com>