Get rid of the css::util::NumberFormat::TEXT abuse as "special" "key"
It is a bit, not a key, which demanded special treatment at all
places, and a key with the same value could not be handled at all.

Abusing the css::util::NumberFormat::TEXT flag as number format
key to signal the "special" meaning, having a number format key
equal to that value would had resulted in undesired behaviour.
The bit value is 256 and a key 256 means 156 user defined number
formats in the default locale, rarely in the wild but can happen.

Change-Id: Idfd5b07d524c222df3491d201095ef65ad4a46c9
Reviewed-on: https://gerrit.libreoffice.org/45705
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Eike Rathke <erack@redhat.com>