Resolves: tdf#114231 do not mask ErrCode with GetRest()
Regression from

    commit 7ca950ec744b7af1d15724ec2abc296573a641e4
    Date:   Wed Aug 23 19:25:02 2017 +0200

	no need to use ERRCODE_RES_MASK here

	the relevant usage sites already call GetRest() before comparing

which exactly is the reason that it didn't work anymore.

Old StringArray ItemList resources stored only 16-bit values,
hence ERRCODE_RES_MASK was used to mask the ErrCode values in the
resource for which code had to mask ERRCODE_RES_MASK as well to
compare values. Now the full ErrCode is stored, so code must not
use GetRest() on a value to compare against, or use GetRest() on
both values (which theoretically could lead to ambiguities, but
probably doesn't in resources that are restricted to one module).

Change-Id: I835e47424bb008bc680dc4f8c502c9558397db36