Related: tdf#114251 disable ocName named expressions for vectorization
Recursive checking of their token arrays would be needed.

This solves the slowness in the non-threading case of tdf#114251
because no arrays are generated for the named ranges, which is
rather a lucky workaround than a solution.

The wrong handling of range references in general (array where
scalar is expected, no implicit intersection) still persists.

Change-Id: I57385bd95efa584108d93d724921285c97a91a45
Reviewed-on: https://gerrit.libreoffice.org/46623
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Eike Rathke <erack@redhat.com>