menu: pass paint size to ImplPaint and account for buttons
When a menu bar is painted it needs to take into account that
the buttons (close) is positioned at the right side of the paint
area. To do this we need to pass what the output size is when
painting (ImplPaint) on the common menu code, instead of assuming
the whole area can be used.

Change-Id: I2e9d6c686929fe1cd7e28368a8055c1e2df13c49
Reviewed-on: https://gerrit.libreoffice.org/46710
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>