tdf#112140 fix some NPEs re touch/zoom while doucment is not ready
touching the screen or trying to zoom were trying to get/adjust zoom
factor despite not ready yet, similarily swiping from the left to show
the parts view would crash as well in case the document is not ready
yet.
Most likely to see this with a debug build (where loading is slower
to begin with), or with long documents like the license text.

Change-Id: Iea5e8c07a9b0973ce50f90114eb015e8e896e390
Reviewed-on: https://gerrit.libreoffice.org/46848
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>