tdf#114519 Update cell content on sidebar number format change
In Calc, when the cursor is flashing in a cell (input mode) and the
sidebar Number Format panel 'Select a category of contents' drop down
selection is changed the cell contents does not update on selection
change.

This patch ends cell input mode on 'Select a category of contents' drop
down change allowing cell to be updated to the selected number format.

Change-Id: Ia1f46499ada2cfbcc82bd5babb6cf56ae0d6259c
Reviewed-on: https://gerrit.libreoffice.org/46883
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Eike Rathke <erack@redhat.com>