Log just one line in cppuhelper::ServiceManager::preloadImplementations()
That *is* what we want, I guess, so better I do it proactively before
I am told to do it. Sorry for the back and forth, but my main point in
touching this code was anyway the filtering out of the component (the
KDE4Backend thing) that causes a crash for me, at least, and which it
presumably is fairly pointless to preload anyway.

Also, filter out libraries that for some reason have an empty name.

Change-Id: I2cffd51e67c23977e0779e7f0d6faaf59d8819d9