writerfilter: FindStyleSheetBy... remove StyleName option
It is confusing having three undocumented Find functions.
Reduce by one - either search on the imported WW8 (ISTD) name
or on the internal (converted) name.

The one remaining use of ByStyleName was testing against an
already set property, so by definition it was set using
convertedStyleName.  (Plus, that was my code anyway...)

Change-Id: If5a352bedb7d297698f449fbb629a1350e27c51c
Reviewed-on: https://gerrit.libreoffice.org/48037
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>