mailconfig: default to secure port 587
A very readable explanation of ports 25 / 465 / 587 was
http://blog.mailgun.com/25-465-587-what-port-should-i-use/

In short, 465 was a very temporary and never official port
to use for encrypted sending. Port 587 is the official
port for encrypted (or even non-encrypted) mail submissions.

For a very long time now, every public mail server (and likely
even private internal servers) supports STARTTLS, so default
to using secure connections.

Change-Id: Ic8a6886221fa2e7b733b2535a84a584fe64d3165
Reviewed-on: https://gerrit.libreoffice.org/48075
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Heiko Tietze <tietze.heiko@gmail.com>
Tested-by: Heiko Tietze <tietze.heiko@gmail.com>
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>
Reviewed-by: Justin Luth <justin_luth@sil.org>