tdf#115057 Never use ImplPopupFloatWin for anything except ToolBox
ImplPopupFloatWin draws the window title and frame directly on the client
window, which is incompatible with .ui layouts, as they try to occupy the
whole available space. Instead just use a plain FloatingWindow whenever
possible, and change the border window code to mimic the behavior of
ImplPopupFloatWin, wrt removing part of the frame of toolbar popups w/o a
title.

Change-Id: If72ea89e31f913470faad7fcb8083f9b9787c9a8
Reviewed-on: https://gerrit.libreoffice.org/48094
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>
Tested-by: Maxim Monastirsky <momonasmon@gmail.com>