Introduce dedicated SwXMLTableContext::MAX_WIDTH, replacing USHRT_MAX
For one, this should make it more obvious what the magic constant USHRT_MAX
meant in the context of SwXMLTableContext::m_nWidth (and shows that it should
arguably have value SAL_MAX_UINT16, not USHRT_MAX).

For another, at least some Android builds are stuck with a broken C library that
defines USHRT_MAX to be of type unsigned int instead of signed int, which caused
various -Wsign-compare that are removed as a side effect.

Change-Id: If2676954f4e7159b0c0d3656b8bc0186f0771e10
Reviewed-on: https://gerrit.libreoffice.org/48661
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Michael Stahl <mstahl@redhat.com>