tdf#114168 If minor axis unit is automatic, then set it to 5
Based on OOXML implementation in MS Excel,
if Minor axis Unit is set to automatic,
then during chart import, LibreOffice should set
Interval Count to 5, to mimic behaviour of MS Excel.
Becaues default Interval Count for LibreOffice is 2,
we need to override it to 5.
With that solution, the Minor axis unit is preserved
also after saving to .ods file format.

During .xlsx export, if Interval Count is set to 5,
then treat is as automatic axis unit.

Change-Id: Iab9209fb3950ef73e79229329606363b528d35fe
Reviewed-on: https://gerrit.libreoffice.org/49327
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Bartosz Kosiorek <gang65@poczta.onet.pl>