tdf#115554:  Create tar archives using 'fakeroot'
'tar' must be run as root or faked to be run as root,
to make sure that file ownerships and permissions inside
the created tar archives are as desired (root:root).

Have fakeroot take care about creating an appropriate
environment rather than using the custom "libgetuid",
to no longer have to care about tar internals by ourselves.

This fixes the problem that file ownerships are incorrect
when tar version >= 1.24 is used for

* tar archives holding all the generated deb/rpm packages
  (created in 'download.pm')
* tar archives created by using the
  '--with-pacakage-format=archive' autogen option
  (created in 'simplepackage.pm')

Change-Id: Id20ccce4d002ff95c75292eda8080ca299eee3a5
Reviewed-on: https://gerrit.libreoffice.org/49682
Reviewed-by: Michael Stahl <mstahl@redhat.com>
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>