tdf#115938: Properly detect shadow size control's changed state
This change makes sure that only if the shadow size edit box' text
is changed, it is converted to the size in twips. It sets wrapper's
default value on each SetControlValue, to return proper initial
size when the size text wasn't modified.

Also both ExampleSet and OutSet in SfxTabDialog's Reset handler are
processed to set relevant items back to initial value; otherwise,
as the restored item (marked as unchanged) will be absent in tab's
result set, it would not modify the old items in the two, and they
would keep previous modified values.

Change-Id: Ie4362811968c6e6cbe0f9229bd3c9b29462f5704
Reviewed-on: https://gerrit.libreoffice.org/50196
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>