StyleSheetUser use in dynamic_cast should better be SAL_DLLPUBLIC_RTTI
...so that the dynamic_cast is guaranteed to work on macOS (where RTTI
equivalence is determined by address, not by strcmp).  The relevant dynamic_cast
is in SdStyleSheet::IsUsed in sd/source/core/stlsheet.cxx, and at least during a
`make check` it appears to only be fed SdStyleSheet instances derived from
StyleSheetUser (and SdStyleSheet is also defiined in Library_sd, so the
dynamic_cast happens to work fine also on macOS in that case).  But better be
conservative and mark StyleSheetUser as SAL_DLLPUBLIC_RTTI after all.

Change-Id: I760f58357b7fe3b4826ea496934e03da3499419e
Reviewed-on: https://gerrit.libreoffice.org/50246
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>
Tested-by: Stephan Bergmann <sbergman@redhat.com>