tdf#114076: Expand ClipRange to next PixelBound
Do this by going to basegfx::B2DRange, adding a
single pixel size and using floor/ceil to go to
full integer (as needed for pixels). Also need
to go back to basegfx::B2DPolyPolygon for the
creation of the needed MaskPrimitive2D.
The general problem is that Writer is scrolling
using blitting the unchanged parts, this forces
this part of the scroll to pixel coordinate steps,
while the ViewTransformation for paint nowadays has
a sub-pixel precision. This results in an offset
up to one pixel in radius. To solve this for now,
we need to expand to the next outer pixel bound.
Hopefully in the future we will someday be able to
stay on the full available precision, but this
will need a change in the repaint/scroll paradigm.

Change-Id: Idb4a596bc298d1b80aab6f36ad4204afe94d54f7
Reviewed-on: https://gerrit.libreoffice.org/50369
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Armin Le Grand <Armin.Le.Grand@cib.de>