swarm: don't assert the result in testUnconstrained
It can happen the algorithm doesn't converge the values towards
the solution in the timeframe and number of generations it has
available (slow execution, high cpu load, bad luck) so for now
don't assert the expected value in the test, but still run it
so we know the algorithm does end eventually and doesn't cause
any seg. faults.

Change-Id: Ic889ad1618bf057ac7b281c15bce735f8ec5e13c
Reviewed-on: https://gerrit.libreoffice.org/50500
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>