remove some GraphicObject URL funct. from GraphicImportHelper
GraphicImportHelper is an implementation of XGraphicObjectResolver
which is mostly used to "resolve" a GraphicObject URL to the
internal storage URL (and save the graphic to the storage in the
process). Most of the GraphicObject URL functionality was removed
so the required overrides from XGraphicObjectResolver interface
now return an empty result or do nothing.

Change-Id: Iee6bb71b15411d03ef82ab4d20d234ff0e834425
Reviewed-on: https://gerrit.libreoffice.org/51331
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>