OperationSmiley: Adapted TestDocument
The OperationSmiley fix is about removing the old hack
that for CustomShapes in top-left and bottom-right points
were added. In the ScTiledRenderingTest::testGraphicInvalidate
UnitTest using this document exactly that top-left wrong polygon
was hit which is now no longer there. The used shape was an
ellipse, so this makes a difference - there *is* no geometry
in the top-left corner of this shape. To fix that, I exchanged
the ellipe CustomShape with a rectangle CustomShape. That way,
the old hit behaviour will be replicated.

Change-Id: I4b96ad79ba91166f097207226577d34087625589