tdf#108122 HTML export: fix lost images on copy
As in the copy part of copy&paste. The problem was that the document has
an empty base URL during copy, and this was images were not exported at
all.

An alternative fix would be to use embedded images, but sadly Word's
HTML import doesn't understand that markup, so use tempfiles instead.

Change-Id: Iab8c555ac244d943c4958f24f8ac61cba4ec3aba
Reviewed-on: https://gerrit.libreoffice.org/52003
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>