Allow URL-Check for non-selected TableText
The part to detect a hit on a TableObjectText
in Draw/Impress for a non-selected Table was
setting the HitType to SdrHitKind::TextEditObj.
This prevents the latter happeing check for URL
(see below). Not directly setting to TextEditObj
is okay, this is determined and set later. I
checked various possible changes to keep the
behaviour as for 'normal' SdrObjects, so please
be careful when doing changes in the HitTest code

Change-Id: I326a73b88fc56fd902b5052cc3a2ddecf3c3ba61
Reviewed-on: https://gerrit.libreoffice.org/52016
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Armin Le Grand <Armin.Le.Grand@cib.de>