tdf#115574 sot: fix Excel -> Writer paste
Reported problem is that nothing happens for paste. Direct cause is that
BITMAP is selected as the format, and Excel advertises BITMAP, but when
we try to import that, it fails. There are 3 interesting commits in the
recent history for this topic:

- commit c47db038f98aaf7aec3cbe57c4e5683591afa23e (fdo#52547 SOT: Prefer
  embedding image data to embedding linked image., 2014-02-07) was a
  bugfix due to newer firefox
- commit 538c13f3d1756f2d105115f64ab1bc0b7426eebc (fdo#78801 fdo#52547
  Paste preference is image, then html, then text., 2014-05-28) was a
  regression fix from the previous fix
- commit a96a7ce51aa98fb9ee97ea3803e2b7e648611008 (fdo#81835 Don't
  prefer GDI Metafiles to RTF/HTML, 2014-08-05), was a regression fix
  from the previous fixes

Going back to the original state shows that the Excel -> Writer use-case
used to be RTF.

Restore the old Excel -> Writer (RTF) behavior by:

- going back to the original state, ignoring the enum class conversions
- re-fix fdo#52547: prefer bitmap over html, but leave everything else
  unchanged
- fdo#78801 needs no fix in this case
- fdo#81835 needs no fix in this case
- tdf#115574 selects RTF -> table shows up

After all these complications, the actual fix is surprisingly simple.

Change-Id: I2d728afa7d1dd7888fa43525366c197d806eea6c
Reviewed-on: https://gerrit.libreoffice.org/52120
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>