tdf#95843: Wait for fire_glxtest_process also in --headless mode
Discussed with mmeeks on IRC that fire_glxtest_process is probably called as
early as possible so that its reuslt is ready by the time it is needed in the
non-headless case.  So best fix for headless is probably to just wait for the
sub-process at an opportune point, instead of redesigning the whole mess so that
fire_glxtest_process would only be called once its result is actually needed.

Change-Id: I4ea9c9d54b83c9695a3b72317e68fed0c410da0e
Reviewed-on: https://gerrit.libreoffice.org/53154
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>