fix memory management in BitmapConvolutionMatrixFilter
was deleting a stack allocated buffer.

Also
(*) drop the BitmapConvolutionMatrixFilter default constructor, the
current code does not permit mpMatrix to be nullptr
(*) declare the mpMatrix field as a reference to a fixed length array,
to be more precise
(*) pass the array in the constructor so that call sites will be
properly type- and length-checked.

Change-Id: I650d56cdfac0dae4ea77df7c0c03e19d658c00c8
Reviewed-on: https://gerrit.libreoffice.org/53312
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>