tdf#108894 Drop Office version numbers in filter names
These are prone to be out of date. Besides, to be extra precise we would
otherwise need to mention Office 365 as well there, and that’s overkill.
What every Office version since 2007 has had in common is that their
formats are called “OOXML”, so just mention that and be done with it.

Change-Id: I0393990e6fd67d039666b3954cc9fc4951a7fe08