Add a distro-config file for LibreOffice Vanilla on the Mac App Store
In addition to the options in this file, the following are present in
our autogen.input for the 6.0 branch:

--disable-ccache
--enable-macosx-code-signing=<redacted>
--enable-macosx-package-signing=<redacted>
--with-external-tar=/Users/tml/lo/src
--with-lang=bg br ca ca-valencia cy cs da de el en-US en-GB es et eu fi fr ga gd gl hr hu id is it ja km lt lv nb nl nn pl pt\
 pt-BR ro ru sk sl sv ta tr uk zh-CN zh-TW
--with-macosx-bundle-identifier=com.collabora.libreoffice-free
--with-product-name=LibreOffice Vanilla
--with-package-version=<depends on version>
--with-referenced-git=/Users/tml/lo/core
--without-parallelism

The language list is a guesstimate of what localisations are good
enough. Suggestions of languages that should be dropped (or added)
welcome.

(We don't want a language with immediately visible holes (in top level
menus or common dialogs) in the translation coverage. That is an
insult to a user who would want to use that language and worse than
not providing a localisation in that language at all. In my opinion.)

Change-Id: I0dd1c2848cd964f6bf9e17d95d1735b81b235b1c