writerfilter NFC: save ConvertedParaStyleName, not WW8Name
and rename m_sCurrentParaStyleId -> m_sCurrentParaStyleName

This is prep work for tdf#117504, where the value
saved in m_sCurrentParaStyleId is not accurate. Since the
actual value saved inside the paragraph properties is the
ConvertStyleName(), lets use the LO-version of the
style name, so that FindStyleSheetByConvertedStyleName()
can be used for either the paragraph's PropertyValue or
m_pImpl->m_sCurrentParaStyleName.

Change-Id: I8aed80094417ea91f2515e666dc05ecbb021a128
Reviewed-on: https://gerrit.libreoffice.org/54084
Reviewed-by: Justin Luth <justin_luth@sil.org>
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Miklos Vajna <vmiklos@collabora.co.uk>