tdf#117427 missing API for determining and loading linked graphic
This adds 2 properties to GraphicDescriptor:
"Linked" - to indicate if the graphic is a linked one
"OriginURL" - the URL of the location from where the graphic was
or will be loaded (and is linked to)

Additionally GraphicProvider was extended with a "LoadAsLink"
(media)property, which can be passed using queryGraphic method.
If "LoadAsLink" is set to true, then the Graphic will be created
as a linked one.

In addition the tests for this API was added to VCL (home of this
API implementors).

Change-Id: Ibdabe5e7780d1b372d1e66c3e19b18779fee2c9a
Reviewed-on: https://gerrit.libreoffice.org/54090
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>